% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{addPos}
\alias{addPos}
\title{add pos information}
\usage{
addPos(object, ...)
}
\arguments{
\item{object}{either a partition, a partitionCluster or a keyness object}

\item{...}{further arguments}
}
\value{
the original, enhanced object
}
\description{
Add information on part-of-speech tags of tokens to tokens. The method is
available for objects of the classes 'partition', 'partitionCluster' and
'keyness' respectively.
\code{method?addPos("partitionCluster")}, \code{method?addPos("keyness")}).
}

