% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustEncoding-method.R
\name{adjustEncoding}
\alias{adjustEncoding}
\title{adjust encoding}
\usage{
adjustEncoding(.Object, ...)
}
\arguments{
\item{.Object}{the object}

\item{corpusEncoding}{the encoding of the corpus}
}
\description{
make sure that encoding is equal to partition encoding
}

