% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.cqp}
\alias{as.cqp}
\title{Convert a string to a CQP query}
\usage{
as.cqp(queries, normalise_case = FALSE, collapse = FALSE)
}
\arguments{
\item{queries}{a character vector}

\item{normalise_case}{logical}

\item{collapse}{whether collapse the queries into one}
}
\value{
a character vector
}
\description{
Takes a simple string as an imput and converts it to a valid CQP query
}

