% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.cqpQuery}
\alias{as.cqpQuery}
\title{Convert a string to a CQP query}
\usage{
as.cqpQuery(queries, collapse = FALSE)
}
\arguments{
\item{queries}{a character vector}

\item{collapse}{whether collapse the queries into one}
}
\value{
a character vector
}
\description{
Takes a simple string as an imput and converts it to a valid CQP query
}

