% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.mallet_method.R, R/store_method.R
\docType{methods}
\name{as.mallet}
\alias{as.mallet}
\alias{as.mallet,partitionBundle-method}
\alias{store,jobjRef-method}
\title{as.mallet}
\usage{
as.mallet(.Object, ...)

\S4method{as.mallet}{partitionBundle}(.Object, pAttribute = "word",
  stoplist = stopwords("de"), mc = TRUE, verbose = TRUE)

\S4method{store}{jobjRef}(object, filename = NULL)
}
\arguments{
\item{.Object}{partition- or partitionBundle-object}

\item{...}{further parameters}

\item{pAttribute}{the pAttribute to use, typically "word" or "lemma"}

\item{stoplist}{a list of stopwords to be dropped from matrix, defaults to stopwords("de") from tm package}

\item{mc}{whether to use multicore}

\item{verbose}{logical, whether to be verbose}

\item{object}{the rJava-object}

\item{filename}{where to store the Java-object}
}
\description{
Save polmineR-object as mallet object
}
\examples{
use("polmineR.sampleCorpus")
bundle <- partitionBundle("PLPRBTTXT", def=list(text_year="2009"), var=list(text_protocol_no=NULL), pAttribute=NULL)
instanceList <- as.mallet(bundle)
store(instanceList, filename=NULL) # output to a tempfile
}

