% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as.mallet-method.R
\docType{methods}
\name{as.mallet}
\alias{as.mallet}
\alias{as.mallet,partitionCluster-method}
\title{as.mallet}
\usage{
as.mallet(.Object, ...)

\S4method{as.mallet}{partitionCluster}(.Object, pAttribute = "word",
  stoplist = stopwords("de"), mc = TRUE, verbose = TRUE)
}
\arguments{
\item{.Object}{partition- or partitionCluster-object}

\item{...}{further parameters}
}
\description{
Save polmineR-object as mallet object
}
\examples{
\dontrun{
clu <- partitionCluster("ARENEN", def=list(text_newspaper="Süddeutsche Zeitung"), var=list(text_date=NULL), tf=NULL)
foo <- as.mallet(clu)
}
}

