% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.markdown_method.R, R/html_method.R
\docType{methods}
\name{as.markdown}
\alias{as.markdown}
\alias{as.markdown,plprPartition-method}
\alias{as.markdown,pressPartition-method}
\alias{html,character-method}
\title{turn partition into markdown}
\usage{
as.markdown(object, ...)

\S4method{as.markdown}{plprPartition}(object, meta)

\S4method{as.markdown}{pressPartition}(object, meta = c("text_newspaper",
  "text_date"))

\S4method{html}{character}(object)
}
\arguments{
\item{object}{object to be converted}

\item{...}{further arguments}

\item{meta}{metainformation i.e. s-attributes) to be displayed}
}
\description{
for further processing to html
}

