% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.sparseMatrix.R
\docType{methods}
\name{as.sparseMatrix}
\alias{as.sparseMatrix}
\alias{as.sparseMatrix,simple_triplet_matrix-method}
\alias{as.sparseMatrix,TermDocumentMatrix-method}
\alias{as.sparseMatrix,bundle-method}
\title{Type conversion - get sparseMatrix.}
\usage{
as.sparseMatrix(x, ...)

\S4method{as.sparseMatrix}{simple_triplet_matrix}(x, ...)

\S4method{as.sparseMatrix}{TermDocumentMatrix}(x, ...)

\S4method{as.sparseMatrix}{bundle}(x, col)
}
\arguments{
\item{x}{object to convert}

\item{...}{further parameters}

\item{col}{column name to get values from (if x is a bundle)}
}
\description{
Turn objects into the sparseMatrix as defined in the Matrix package.
}
