% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.speeches_method.R
\docType{methods}
\name{as.speeches}
\alias{as.speeches}
\alias{as.speeches,partition-method}
\title{Split partition into speeches}
\usage{
\S4method{as.speeches}{partition}(.Object, sAttributeDates, sAttributeNames,
  gap = 500, mc = FALSE, verbose = FALSE)
}
\arguments{
\item{.Object}{a partition .Object}

\item{sAttributeDates}{the s-attribute that provides the dates of sessions}

\item{sAttributeNames}{the s-attribute that provides the names of speakers}

\item{gap}{number of tokens between strucs to identify speeches}

\item{mc}{whether to use multicore, defaults to FALSE}

\item{verbose}{logical, defaults to TRUE}
}
\value{
a partitionBundle object
}
\description{
A method designed for corpora from the
PolMine corpora of plenary protocols. A partition is split into speeches.
}

