% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.speeches_method.R
\name{as.speeches}
\alias{as.speeches}
\title{Split Partition Into Speeches}
\usage{
as.speeches(.Object, s_attribute_date = grep("date", sAttributes(.Object),
  value = TRUE), s_attribute_name = grep("name", sAttributes(.Object), value =
  TRUE), gap = 500, mc = FALSE, verbose = TRUE, progress = TRUE)
}
\arguments{
\item{.Object}{a partition .Object}

\item{s_attribute_date}{the s-attribute that provides the dates of sessions}

\item{s_attribute_name}{the s-attribute that provides the names of speakers}

\item{gap}{number of tokens between strucs to identify speeches}

\item{mc}{whether to use multicore, defaults to FALSE}

\item{verbose}{logical, defaults to TRUE}

\item{progress}{logical}
}
\value{
a partitionBundle object
}
\description{
A partition is split into speeches.
}
\examples{
use("polmineR")
bt <- partition("GERMAPARLMINI", date = ".*", regex = TRUE)
speeches <- as.speeches(bt, s_attribute_date = "date", s_attribute_name = "speaker")
  
# step-by-step, not the fastest way
speeches <- enrich(speeches, pAttribute = "word")
tdm <- as.TermDocumentMatrix(speeches, col = "count")
  
# fast option (counts performed when assembling the sparse matrix)
# tdm <- as.TermDocumentMatrix(speeches, pAttribute = "word")
# termsToDropList <- noise(tdm)
# whatToDrop <- c("stopwords", "specialChars", "numbers", "minNchar")
# termsToDrop <- unlist(lapply(whatToDrop, function(x) termsToDropList[[x]]))
# tdm <- trim(tdm, termsToDrop = termsToDrop)
}
