% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.VCorpus.R
\docType{methods}
\name{as.VCorpus,partitionBundle-method}
\alias{as.VCorpus,partitionBundle-method}
\title{Coerce partitionBundle to VCorpus.}
\usage{
\S4method{as.VCorpus}{partitionBundle}(x)
}
\arguments{
\item{x}{a partitionBundle object}
}
\description{
Coerce partitionBundle to VCorpus.
}
\examples{
use("polmineR")
P <- partition("GERMAPARLMINI", date = "2009-11-10")
PB <- partitionBundle(P, s_attribute = "speaker")
VC <- as.VCorpus(PB)
}
