% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,context-method}
\alias{as.data.frame,keyness-method}
\title{Get statistics table from an object}
\usage{
\S4method{as.data.frame}{keyness}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{object with a statistics table}

\item{row.names}{defaults to NULL}

\item{optional}{see documentation of as.data.frame}

\item{...}{any further arguments}
}
\description{
Get statistics table from an object
}

