% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blapply_method.R
\docType{methods}
\name{blapply}
\alias{blapply}
\alias{blapply,bundle-method}
\alias{blapply,list-method}
\title{apply a function over a list or bundle with and without verbose parallelization}
\usage{
blapply(x, ...)

\S4method{blapply}{list}(x, f, mc = TRUE, progress = TRUE,
  verbose = FALSE, ...)

\S4method{blapply}{bundle}(x, f, mc = FALSE, progress = TRUE,
  verbose = FALSE, ...)
}
\arguments{
\item{...}{further parameters}

\item{f}{a function that can be applied to each object contained in the bundle}
}
\description{
apply a function over a list or bundle with and without verbose parallelization
}

