% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{browse,html-method}
\alias{browse}
\alias{browse,html-method}
\alias{browse,kwic-method}
\title{display in browser}
\usage{
\S4method{browse}{html}(object)

browse(object, ...)

\S4method{browse}{kwic}(object)
}
\arguments{
\item{object}{what is to be displayed}

\item{...}{further parameters}
}
\description{
bla
}

