% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bughunt-method.R
\docType{methods}
\name{bughunt}
\alias{bughunt}
\title{find bugs in a corpus}
\usage{
\S4method{bughunt}{character}(.Object, nested = c(p = 10), ...)
}
\arguments{
\item{.Object}{a character vector providing the name of a CWB corpus}

\item{nested}{name of an element that is nested in an unwanted fashion}
}
\description{
This is utility to trace back the cause of errors that cwb-encode
throws out when XML elements are nested.
}

