% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bundle_class.R, R/partitionBundle_methods.R
\docType{class}
\name{bundle-class}
\alias{+,bundle,bundle-method}
\alias{+,partitionBundle,partition-method}
\alias{[[,bundle,ANY-method}
\alias{bapply}
\alias{bapply,bundle-method}
\alias{bundle-class}
\alias{length,bundle-method}
\alias{names,bundle-method}
\alias{unique,bundle-method}
\title{bundle class}
\usage{
\S4method{[[}{bundle,ANY}(x, i)

\S4method{length}{bundle}(x)

\S4method{names}{bundle}(x)

\S4method{unique}{bundle}(x)

bapply(x, ...)

\S4method{bapply}{bundle}(x, f, ...)

\S4method{+}{bundle,bundle}(e1, e2)

\S4method{+}{partitionBundle,partition}(e1, e2)
}
\arguments{
\item{x}{a bundle object}

\item{i}{index to address an object}

\item{...}{further parameters}

\item{f}{a function that can be applied to each object contained in the bundle}

\item{e1}{object 1}

\item{e2}{object 2}
}
\description{
A class to bundle several objects (partition, context, keyness, cooccurrences objects)
in one S4 object.
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"}}

\item{\code{pAttribute}}{Object of class \code{"character"}}
}}
\author{
Andreas Blaette
}

