% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize character vector.}
\usage{
capitalize(x)
}
\arguments{
\item{x}{A \code{character} vector.}
}
\description{
Make the first character of the elements of a \code{character} vector have upper
case and the rest lower case.
}
\details{
The \code{capitalize()} function may be useful when applying lowercased
dictionaries of stoplists, a sentiment dictionary etc. on a CWB corpus that
maintains capitalization (tokens are not lowercased).

This function is inspired by a method Python offers for string objects.
}
\examples{
capitalize(c("oil", "corpus", "data"))
}
