% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterCount_method.R
\docType{methods}
\name{characterCount}
\alias{characterCount}
\alias{characterCount,partition-method}
\alias{characterCount,partitionBundle-method}
\title{count characters}
\usage{
characterCount(.Object, ...)

\S4method{characterCount}{partition}(.Object, pAttribute,
  regexCharsToKeep = "[a-zA-Z]", toLower = TRUE, decreasing = TRUE)

\S4method{characterCount}{partitionBundle}(.Object, decreasing = TRUE,
  mc = FALSE, ...)
}
\arguments{
\item{.Object}{object}

\item{...}{parameters that will be passed}

\item{pAttribute}{the p-attribute}

\item{regexCharsToKeep}{if not NULL, a regex}

\item{toLower}{whether to lower tokens}

\item{decreasing}{logical, passed into order call}

\item{mc}{logical}
}
\description{
count characters
}

