% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\docType{methods}
\name{chisquare}
\alias{chisquare}
\alias{chisquare,features-method}
\alias{chisquare,context-method}
\alias{chisquare,cooccurrences-method}
\title{Perform chisquare-text.}
\usage{
chisquare(.Object, ...)

\S4method{chisquare}{features}(.Object)

\S4method{chisquare}{context}(.Object)

\S4method{chisquare}{cooccurrences}(.Object)
}
\arguments{
\item{.Object}{object}

\item{...}{further parameters}
}
\value{
a table
}
\description{
Perform Chisquare-Test based on a table with counts
}
\details{
This function deliberately uses a self-made chi-square test for performance
reason
}
\author{
Andreas Blaette
}
\keyword{textstatistics}
