% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coerce-methods.R
\docType{methods}
\name{as.TermDocumentMatrix}
\alias{as.DocumentTermMatrix,partitionCluster-method}
\alias{as.TermDocumentMatrix}
\alias{as.TermDocumentMatrix,character-method}
\alias{as.TermDocumentMatrix,partitionCluster-method}
\title{coerce object to TermDocumentMatrix/DocumentTermMatrix}
\usage{
as.TermDocumentMatrix(x, ...)

\S4method{as.TermDocumentMatrix}{partitionCluster}(x, pAttribute = NULL,
  weight = NULL, rmBlank = TRUE, verbose = TRUE, ...)

\S4method{as.DocumentTermMatrix}{partitionCluster}(x, pAttribute = NULL,
  weight = NULL, rmBlank = TRUE, ...)

\S4method{as.TermDocumentMatrix}{character}(x, pAttribute, sAttribute,
  from = NULL, to = NULL, strucs = NULL, rmBlank = TRUE,
  verbose = TRUE, robust = FALSE, mc = FALSE)
}
\arguments{
\item{x}{some object}

\item{...}{to make the check happy}

\item{pAttribute}{the p-attribute}

\item{weight}{whether to introduce a weight ("tfidf" and "rel" are implemented)}

\item{rmBlank}{bla}

\item{verbose}{bla}

\item{sAttribute}{the s-attribute}

\item{from}{bla}

\item{to}{bla}

\item{strucs}{bla}

\item{robust}{bla}

\item{mc}{logical}
}
\value{
a TermDocumentMatrix
}
\description{
A set of methods to transform objects into the TermDocumentMatrix-objects, or
DocumentTermMatrix-objects imported from the \code{tm} package.
}
\examples{
\dontrun{
foo <- as.TermDocumentMatrix(
  x="ZEIT", pAttribute="word", sAttribute="text_id",
  from="1946_01_auf-einen-von-bomben-zerschlagenen-engel.html",
  to="1951_08_gespraeche-zum-interzonenhandel.html", robust="LESUNG", mc=TRUE, verbose=TRUE, rmBlank=TRUE
)
}
}
\author{
Andreas Blaette

me
}

