% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collocationsBundle-class.R, R/collocationsBundle-methods.R
\docType{class}
\name{collocationsBundle-class}
\alias{as.TermDocumentMatrix,collocationsBundle-method}
\alias{collocationsBundle}
\alias{collocationsBundle-class}
\title{S4 collocationsBundle class}
\usage{
\S4method{as.TermDocumentMatrix}{collocationsBundle}(x, col, directed = TRUE,
  rel = FALSE, mc = TRUE)
}
\arguments{
\item{x}{a collocationsBundle object}

\item{col}{the column to extract}

\item{directed}{to be explained}

\item{rel}{to be explained, too}

\item{mc}{logical, whether to use multicore}
}
\description{
class to organize information of multiple context analyses
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"} a list of context objects}
}}

