% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collocationsCluster-class.R, R/collocationsCluster-methods.R
\docType{class}
\name{collocationsCluster-class}
\alias{as.TermDocumentMatrix,collocationsCluster-method}
\alias{collocationsCluster}
\alias{collocationsCluster-class}
\title{S4 collocationsCluster class}
\usage{
\S4method{as.TermDocumentMatrix}{collocationsCluster}(x, col, directed = TRUE,
  rel = FALSE, mc = TRUE)
}
\arguments{
\item{x}{a collocationsCluster object}

\item{col}{the column to extract}

\item{directed}{to be explained}

\item{rel}{to be explained, too}

\item{mc}{logical, whether to use multicore}
}
\description{
class to organize information of multiple context analyses
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"} a list of context objects}
}}

