% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collocationsReshaped-class.R
\docType{class}
\name{collocationsReshaped-class}
\alias{collocationsReshaped-class}
\title{S4 context class}
\description{
class to organize information of context analysis
}
\section{Slots}{

\describe{
\item{\code{call}}{Object of class \code{"character"} the call that generated the object}

\item{\code{partition}}{Object of class \code{"character"} the partition the analysis is based on}

\item{\code{partitionSize}}{Object of class \code{"numeric"} the size of the partition}

\item{\code{leftContext}}{Object of class \code{"numeric"} number of tokens to the right}

\item{\code{rightContext}}{Object of class \code{"numeric"} number of tokens to the left}

\item{\code{pAttribute}}{Object of class \code{"character"} p-attribute of the query}

\item{\code{corpus}}{Object of class \code{"character"} the CWB corpus used}

\item{\code{stat}}{Object of class \code{"data.frame"} statistics of the analysis}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{posFilter}}{Object of class \code{"character"} part-of-speech tags filtered}

\item{\code{method}}{Object of class \code{"character"} statistical test(s) used}

\item{\code{cutoff}}{Object of class \code{"list"} cutoff levels that have been applied}

\item{\code{svg}}{Object of class \code{"character"} - valid XML with svg representation}
}}

