% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_class.R, R/enrich_method.R
\docType{class}
\name{comp-class}
\alias{comp-class}
\alias{compCooccurrences-class}
\alias{compNgrams-class}
\alias{enrich,comp-method}
\alias{show,comp-method}
\alias{summary,comp-method}
\title{S4 class for comparing corpora}
\usage{
\S4method{summary}{comp}(object)

\S4method{show}{comp}(object)

\S4method{enrich}{comp}(object, addPos = NULL, verbose = TRUE)
}
\arguments{
\item{object}{a comp object}

\item{addPos}{logical, whether to add POS information to comp object}

\item{verbose}{whether R should be talkative}
}
\description{
object resulting from compare-method

Wrapper for adding a statistic on pos attributes to comp object
}
\section{Slots}{

\describe{
\item{\code{corpus}}{Object of class \code{"character"}}

\item{\code{pAttribute}}{Object of class \code{"character"}}

\item{\code{encoding}}{Object of class \code{"character"}}

\item{\code{corpus}}{Object of class \code{"character"}}

\item{\code{stat}}{Object of class \code{"data.frame"}}

\item{\code{sizeCoi}}{Object of class \code{"numeric"}}

\item{\code{sizeRef}}{Object of class \code{"numeric"}}

\item{\code{included}}{Object of class \code{"logical"} whether corpus of interest is included in reference corpus}

\item{\code{method}}{Object of class \code{"character"} statisticalTest used}

\item{\code{call}}{Object of class \code{"character"} the call that generated the object
@section Methods:
 \describe{
  \item{summary}{\code{signature(object = "comp")}: Display essential information }
  \item{addPos}{\code{signature(object = "comp")}: add POS attribute to statistics table }
  \item{as.data.frame}{\code{signature(object = "comp")}: get the statistics table}
  }}
}}
\author{
Andreas Blaette
}

