% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_method.R
\docType{methods}
\name{compare}
\alias{compare}
\alias{compare,cooccurrences-method}
\alias{compare,missing-method}
\alias{compare,ngrams-method}
\alias{compare,partition-method}
\alias{compare,partitionBundle-method}
\title{compare features of two partitions}
\usage{
compare(x, ...)

\S4method{compare}{partition}(x, y, included = FALSE, method = "chisquare",
  verbose = FALSE)

\S4method{compare}{partitionBundle}(x, y, included = FALSE,
  method = "chisquare", verbose = TRUE, mc = TRUE, progress = FALSE)

\S4method{compare}{cooccurrences}(x, y, included = FALSE, method = "ll",
  mc = TRUE, verbose = TRUE)

\S4method{compare}{missing}()

\S4method{compare}{ngrams}(x, y, included = FALSE, method = "chisquare",
  verbose = TRUE, ...)
}
\arguments{
\item{x}{a partition or partitionBundle object}

\item{...}{further parameters}

\item{y}{a partition object, it is assumed that the coi is a subcorpus of
ref}

\item{included}{TRUE if coi is part of ref, defaults to FALSE}

\item{method}{the statistical test to apply (chisquare or log likelihood)}

\item{verbose}{defaults to TRUE}

\item{mc}{logical, whether to use multicore}

\item{progress}{logical}
}
\value{
The function returns a data frame with the following structure:
- absolute frequencies in the first row
- ...
}
\description{
compare features of two partitions
}
\author{
Andreas Blaette
}
\references{
Manning / Schuetze ...
}

