% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_class.R
\docType{class}
\name{compBundle-class}
\alias{compBundle-class}
\alias{summary,compBundle-method}
\title{compBundle-class}
\usage{
\S4method{summary}{compBundle}(object)
}
\arguments{
\item{x}{object}

\item{col}{the column to extract}

\item{rmBlank}{remove blank rows}

\item{verbose}{logical, whether to be verbose}
}
\description{
The class inherits from the bundle class.
}
\section{Slots}{

\describe{
\item{\code{objects}}{an object of class \code{list}}
}}

