% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatenate.R
\docType{methods}
\name{concatenate}
\alias{concatenate}
\alias{concatenate,data.table-method}
\alias{concatenate,matrix-method}
\title{Concatenate tokens to phrases}
\usage{
concatenate(.Object, ...)

\S4method{concatenate}{data.table}(.Object, regions, col, corpus)

\S4method{concatenate}{matrix}(.Object, corpus, p_attribute)
}
\arguments{
\item{.Object}{Either a \code{matrix} or a \code{data.table}.}

\item{...}{Further arguments that may be defined for the method.}

\item{regions}{A two-column \code{matrix} with corpus positions defining 
regions.}

\item{col}{If \code{.Object} is a \code{data.table}, the column to concatenate.}

\item{corpus}{A length-one \code{character} vector, the coropus ID of the corpus
from which regions / the \code{data.table} representing a decoded corpus is derived.}

\item{p_attribute}{The positional attribute for which tokens shall be decoded.}
}
\description{
Concatenate sequences of tokens defined by regions to phrases.
}
\examples{
lexical_units_cqp <- c(
  '"Deutsche.*" "Bundestag.*"',
  '"sozial.*" "Gerechtigkeit"',
  '"Ausschuss" "für" "Arbeit" "und" "Soziales"',
  '"soziale.*" "Marktwirtschaft"',
  '"freiheitliche.*" "Grundordnung"'
)
phrases_regions <- cpos("GERMAPARLMINI", query = lexical_units_cqp, cqp = TRUE)
phrases <- concatenate(phrases_regions, corpus = "GERMAPARLMINI", p_attribute = "word")

dt <- decode("GERMAPARLMINI", p_attribute = "word", s_attribute = character(), to = "data.table")
y <- concatenate(dt, regions = phrases_regions, col = "word", corpus = "GERMAPARLMINI")
}
