% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R
\docType{class}
\name{context_bundle-class}
\alias{context_bundle-class}
\alias{show,context_bundle-method}
\alias{summary,context_bundle-method}
\alias{[,context_bundle-method}
\alias{[,context_bundle,ANY,ANY,ANY-method}
\alias{[[,context_bundle-method}
\title{S4 context_bundle class}
\description{
class to organize information of multiple context analyses
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"} a list of context objects}
}}

\section{Methods}{

  \describe{
    \item{show}{output of core information}
    \item{summary}{core statistical information}
    \item{[}{specific cooccurrences}
    \item{[[}{specific cooccurrences}
   }
}

