% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextBundle_class.R
\docType{class}
\name{contextBundle-class}
\alias{contextBundle-class}
\alias{show,contextBundle-method}
\alias{summary,contextBundle-method}
\alias{[,contextBundle-method}
\alias{[,contextBundle,ANY,ANY,ANY-method}
\alias{[[,contextBundle-method}
\alias{as.TermContextBundle,contextBundle-method}
\title{S4 contextBundle class}
\description{
class to organize information of multiple context analyses
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"} a list of context objects}
}}

\section{Methods}{

  \describe{
    \item{show}{output of core information}
    \item{summary}{core statistical information}
    \item{[}{specific cooccurrences}
    \item{[[}{specific cooccurrences}
   }
}

