% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrence_method.R
\docType{methods}
\name{cooccurrence}
\alias{cooccurrence}
\alias{cooccurrence,partition-method}
\title{find cooccurrences}
\usage{
cooccurrence(object, ...)

\S4method{cooccurrence}{partition}(object, x, y, dist = 5,
  as.character = TRUE)
}
\arguments{
\item{object}{the partition}

\item{...}{further parameters}

\item{x}{cqp query one}

\item{y}{cqp query two}

\item{dist}{distance between queries}

\item{as.character}{logical, defaults to TRUE}
}
\description{
a story to be told
}

