% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences_class.R, R/view_method.R
\docType{class}
\name{cooccurrences-class}
\alias{cooccurrences-class}
\alias{cooccurrencesReshaped-class}
\alias{summary,cooccurrences-method}
\alias{show,cooccurrences-method}
\alias{cooccurrencesBundle-class}
\alias{cooccurrencesBundle}
\alias{as.data.frame,cooccurrencesBundle-method}
\alias{view,cooccurrencesReshaped-method}
\title{Cooccurrences class.}
\usage{
\S4method{summary}{cooccurrences}(object)

\S4method{show}{cooccurrences}(object)

\S4method{as.data.frame}{cooccurrencesBundle}(x)

\S4method{view}{cooccurrencesReshaped}(.Object)
}
\arguments{
\item{object}{object to work with}

\item{x}{object to work with}

\item{.Object}{object to work with}
}
\description{
S4 class to organize information of context analysis
}
\section{Slots}{

\describe{
\item{\code{call}}{Object of class \code{"character"} the call that generated the object}

\item{\code{partition}}{Object of class \code{"character"} the partition the analysis is based on}

\item{\code{partitionSize}}{Object of class \code{"numeric"} the size of the partition}

\item{\code{left}}{Object of class \code{"numeric"} number of tokens to the right}

\item{\code{right}}{Object of class \code{"numeric"} number of tokens to the left}

\item{\code{pAttribute}}{Object of class \code{"character"} p-attribute of the query}

\item{\code{corpus}}{Object of class \code{"character"} the CWB corpus used}

\item{\code{stat}}{Object of class \code{"data.frame"} statistics of the analysis}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{pos}}{Object of class \code{"character"} part-of-speech tags filtered}

\item{\code{method}}{Object of class \code{"character"} statistical test(s) used}

\item{\code{cutoff}}{Object of class \code{"list"} cutoff levels that have been applied}

\item{\code{svg}}{Object of class \code{"character"} - valid XML with svg representation}
}}

