% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cooccurrences_method.R
\docType{methods}
\name{cooccurrences}
\alias{cooccurrences}
\alias{cooccurrences,partition-method}
\alias{cooccurrences,partitionBundle-method}
\title{get all cooccurrences in a partition}
\usage{
cooccurrences(.Object, ...)

\S4method{cooccurrences}{partition}(.Object, pAttribute = c("word", "pos"),
  window = 5, keep = list(pos = c("NN", "ADJA")), method = "ll",
  mc = FALSE, progress = TRUE, verbose = TRUE, ...)

\S4method{cooccurrences}{partitionBundle}(.Object, pAttribute = "word",
  window = 5, method = "ll", keep = list(pos = c("ADJA", "NN")),
  mc = FALSE)
}
\arguments{
\item{.Object}{a partition object}

\item{pAttribute}{p-attribute to define tokens (can be length > 1)}

\item{window}{no of tokens to the left and to the right of nodes}

\item{keep}{list with tokens to keep}

\item{method}{statistical test to use (defaults to "ll")}

\item{mc}{whether to use multicore}
}
\value{
a cooccurrences-class object
}
\description{
get all cooccurrences in a partition
}
\examples{
\dontrun{
bt17merkel <- partition("PLPRTXT", list(text_lp="17", text_type="speech", text_speaker="Angela Merkel"))
bt17merkelColl <- cooccurrences(bt17merkel)
}
}
\author{
Andreas Blaette
}

