% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences.R
\name{cooccurrences_reshaped}
\alias{cooccurrences_reshaped}
\alias{merge,cooccurrences_reshaped-method}
\title{Methods for manipulating \code{cooccurrences_reshaped}-class-objects}
\arguments{
\item{x}{cooccurrences for a corpus of interest}

\item{y}{cooccurrences for a reference corpus}
}
\description{
Methods for manipulating \code{cooccurrences_reshaped}-class-objects
}
