% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences_class.R
\docType{class}
\name{cooccurrencesBundle-class}
\alias{cooccurrencesBundle}
\alias{cooccurrencesBundle-class}
\title{S4 cooccurrencesBundle class}
\arguments{
\item{x}{a cooccurrencesBundle object}

\item{col}{the column to extract}

\item{directed}{to be explained}

\item{rel}{to be explained, too}

\item{mc}{logical, whether to use multicore}
}
\description{
class to organize information of multiple context analyses
}
\section{Slots}{

\describe{
\item{\code{objects}}{for cooccurrenceBundle-objects - Object of class \code{"list"} - list of cooccurrence objects}
}}

