% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/corpus.R
\docType{class}
\name{corpus-class}
\alias{corpus-class}
\alias{zoom}
\alias{corpus}
\alias{get_corpus}
\alias{name,corpus-method}
\alias{corpus,character-method}
\alias{get_corpus,corpus-method}
\alias{corpus,missing-method}
\alias{show,corpus-method}
\alias{$,corpus-method}
\title{S4 class to manage CWB corpora.}
\usage{
\S4method{name}{corpus}(x)

\S4method{corpus}{character}(.Object, server = NULL)

\S4method{get_corpus}{corpus}(x)

\S4method{corpus}{missing}()

\S4method{show}{corpus}(object)

\S4method{$}{corpus}(x, name)
}
\arguments{
\item{x}{An object of class \code{corpus}, or inheriting from it.}

\item{.Object}{An object inheriting from the \code{textstat} or \code{bundle}
superclasses.}

\item{server}{The name of an Open CPU server (can be an IP address) that
hosts corpora.}

\item{object}{An object.}

\item{name}{A (single) s-attribute.}

\item{...}{Further arguments.}
}
\description{
Corpora indexed using the Corpus Workbench (CWB) offer an
  efficient data structure for large, linguistically annotated corpora. The
  \code{corpus}-class serves to manage basic information on a CWB corpus.
  Corresponding to the name of the class, use the \code{corpus}-method to
  instantiate objects of class \code{corpus}. Calling the
  \code{corpus}-method without an argument will return a \code{data.frame}
  with basic information on the corpora that are available. This
  documentation object also includes the methods to extract basic information
  on \code{corpus} objects.
}
\details{
A \code{corpus} object can have a name, which can be retrieved using
the \code{name}-method.

Applying the `$`-method on a corpus will return the values for the
  s-attribute stated with argument \code{name}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{name}: Get the name of a \code{corpus} object.

\item \code{corpus}: If \code{.Object} is an object inheriting from the \code{textstat},
or the \code{bundle} class, the corpus used to generate the object is
returned.

\item \code{get_corpus}: Get the corpus ID from the \code{corpus} object.

\item \code{corpus}: Calling \code{corpus()} will return a \code{data.frame}
listing the corpora described in the active registry directory, and some
basic information on the corpora.

\item \code{show}: Show basic information on the \code{corpus} object.

\item \code{$}: Quick access to the structural attributes of a corpus.
}}

\section{Slots}{

\describe{
\item{\code{corpus}}{A length-one \code{character} vector, a CWB corpus.}

\item{\code{data_dir}}{The directory where the files for the indexed corpus are.}

\item{\code{type}}{The type of the corpus (e.g. "plpr" for a corpus of plenary protocols).}

\item{\code{name}}{An additional name for the object that may be more telling than the corpus ID.}

\item{\code{encoding}}{The encoding of the corpus, given as a length-one \code{character} vector.}
}}

\examples{
g <- corpus("GERMAPARLMINI")
g$date
corpus("GERMAPARLMINI")$date

sc <- subset("GERMAPARLMINI", date == "2009-10-27")
sc$date
}
\seealso{
Other classes to manage corpora: \code{\link{regions}},
  \code{\link{subcorpus}}
}
\concept{classes to manage corpora}
