% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/corpus.R
\docType{class}
\name{corpus-class}
\alias{corpus-class}
\alias{zoom}
\alias{show,corpus-method}
\title{S4 class to wrap information on CWB corpora.}
\usage{
\S4method{show}{corpus}(object)
}
\arguments{
\item{object}{An object.}

\item{...}{Further arguments.}
}
\description{
S4 class to wrap information on CWB corpora.
}
\section{Slots}{

\describe{
\item{\code{corpus}}{A length-one \code{character} vector, a CWB corpus.}

\item{\code{data_dir}}{The directory where the files for the indexed corpus are.}

\item{\code{type}}{The type of the corpus (e.g. "plpr" for a corpus of plenary protocols).}

\item{\code{encoding}}{The encoding of the corpus, given as a length-one \code{character} vector.}
}}

\seealso{
Other classes to manage corpora: \code{\link{regions}},
  \code{\link{subcorpus-class}}
}
\concept{classes to manage corpora}
