% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/corpus.R
\docType{methods}
\name{corpus-methods}
\alias{corpus-methods}
\alias{name,corpus-method}
\alias{get_corpus,corpus-method}
\alias{show,corpus-method}
\alias{$,corpus-method}
\title{Corpus class methods}
\usage{
\S4method{name}{corpus}(x)

\S4method{get_corpus}{corpus}(x)

\S4method{show}{corpus}(object)

\S4method{$}{corpus}(x, name)
}
\arguments{
\item{x}{An object of class \code{corpus}, or inheriting from it.}

\item{object}{An object of class \code{corpus}, or inheriting from it.}

\item{name}{A (single) s-attribute.}
}
\description{
A set of generic methods is available to extract basic information from
objects of the \code{corpus} class.
}
\details{
A \code{corpus} object can have a name, which can be retrieved using
the \code{name}-method.

Use \code{get_corpus}-method to get the corpus ID from the slot
  \code{corpus} of the \code{corpus} object.

The \code{show}-method will show basic information on the
  \code{corpus} object.

Applying the `$`-method on a corpus will return the values for the
  s-attribute stated with argument \code{name}.
}
\examples{
use("polmineR")
g <- corpus("GERMAPARLMINI")
g$date
corpus("GERMAPARLMINI")$date

use("polmineR")
sc <- subset("GERMAPARLMINI", date == "2009-10-27")
sc$date
}
