% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_method.R
\docType{methods}
\name{count}
\alias{count}
\alias{count,character-method}
\alias{count,partition-method}
\alias{count,partitionBundle-method}
\alias{count-method}
\title{get counts}
\usage{
count(.Object, ...)

\S4method{count}{partition}(.Object, query, mc = F, verbose = T)

\S4method{count}{partitionBundle}(.Object, query, mc = F, verbose = T)

\S4method{count}{character}(.Object, query, pAttribute = NULL,
  verbose = TRUE)
}
\arguments{
\item{.Object}{either a partition or a partitionBundle object}

\item{...}{further parameters}

\item{query}{a character vector (one or multiple terms to be looked up)}

\item{mc}{logical, whether to use multicore (defaults to FALSE)}

\item{verbose}{logical, whether to be verbose}
}
\description{
Method to obtain term frequencies for one or multiple CQP queries. If
CQP syntax is not needed, use subst as partition-method.
}
\examples{
# generate a partition for testing 
test <- partition("PLPRBTTXT")
count(test, "Wir") # get frequencies for one token
count(test, c("Wir", "lassen", "uns")) # get frequencies for multiple tokens
count(test, c("Zuwander.*", "Integration.*"), method="grep") # get frequencies using "grep"-method
count("PLPRTXT", c("machen", "Integration"), "word")
}
\seealso{
count
}

