% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_class.R
\docType{class}
\name{count_class}
\alias{count_class}
\alias{count-class}
\title{Count class.}
\description{
S4 class to organize counts. The classes \code{polmineR} and
\code{ngrams} inherit from the class.
}
\section{Slots}{

\describe{
\item{\code{stat}}{Object of class \code{data.table}}

\item{\code{corpus}}{Object of class \code{character} the CWB corpus the partition is based on}

\item{\code{encoding}}{Object of class \code{character} encoding of the corpus}

\item{\code{name}}{Object of class \code{character}, a name for the object}

\item{\code{size}}{Object of class \code{integer}, the size of the partition or
corpus the count is based upon}
}}

\seealso{
The \code{count}-class inherits from the \code{\link{textstat-class}}
}
\author{
Andreas Blaette
}
