% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{cp}
\alias{cp}
\title{Copy S4 textstat object.}
\usage{
cp(x)
}
\arguments{
\item{x}{An object (\code{textstat} or class inheriting from \code{textstat}).}
}
\description{
It is not possible to add columns to the \code{data.table} in the \code{stat} slot of a
\code{textclass} object, when the object has been saved and loaded using
\code{save()}/\code{load()}. This scenario applies for instance, when the objects of an
interactive R session are saved, and loaded when starting the next
interactive R session. The \code{cp()} function will create a copy of the object,
including an explicit copy of the \code{data.table} in the \code{stat} slot. Inplace
modifications of the new object are possible. The function can also be used
to avoid unwanted side effects of modifying an object.
}
\examples{
k <- kwic("REUTERS", query = "oil")
rdata_file <- tempfile(fileext = ".RData")
save(k, file = rdata_file)
rm(k)

load(rdata_file)
k <- cp(k) # now it is possible to columns by reference
enrich(k, s_attribute = "id")
"id" \%in\% colnames(k)
}
\seealso{
The \code{\link{restore}} function addesses a related scenario.
}
