\name{crosstabulationSizes}
\alias{crosstabulationSizes}
\title{Size of sub-partitions}
\usage{
crosstabulationSizes(Partition, rows, cols)
}
\arguments{
  \item{Partition}{a partition object}

  \item{rows}{what to find in rows}

  \item{cols}{what to find in cols}
}
\description{
Size of sub-partitions
}

