% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed.R
\name{browse}
\alias{browse}
\title{Defunct methods and functions.}
\usage{
browse(...)
}
\arguments{
\item{...}{Any arguments that may be passed into the defunct function/method.}
}
\description{
Defunct methods and functions.
}
