% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dispersion-class.R, R/dispersion-methods.R, R/tf-method.R
\docType{class}
\name{dispersion-class}
\alias{dispersion-class}
\alias{rel}
\alias{rel,dispersion-method}
\alias{show,dispersion-method}
\alias{t,dispersion-method}
\alias{tf,dispersion-method}
\title{dispersion class}
\usage{
\S4method{rel}{dispersion}(x)

\S4method{t}{dispersion}(x)

\S4method{show}{dispersion}(object)

\S4method{tf}{dispersion}(object)
}
\arguments{
\item{x}{a dispersion object}

\item{object}{a dispersion object}
}
\description{
class to organize results from dispersion analysis
}
\section{Slots}{

\describe{
\item{\code{call}}{call that generated this object}

\item{\code{abs}}{Object of class \code{"character"} the call that generated the object}

\item{\code{rel}}{Object of class \code{"character"} the partition the analysis is based on}

\item{\code{dim}}{dimensions of the object}

\item{\code{query}}{the query / the queries that were used}

\item{\code{sizes}}{Object of class \code{"numeric"} the size of the partition}
}}

