% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_method.R
\docType{methods}
\name{dispersion}
\alias{dispersion}
\title{Dispersion of a query or multiple queries}
\usage{
\S4method{dispersion}{partition}(.Object, query, dim, pAttribute = NULL,
  freq = TRUE, mc = FALSE, progress = TRUE, verbose = FALSE)
}
\arguments{
\item{.Object}{a partition object}

\item{query}{a character vector containing one or multiple queries}

\item{dim}{a character vector of length 1 or 2 providing the sAttributes}

\item{pAttribute}{the p-attribute that will be looked up, typically 'word'
or 'lemma'}

\item{freq}{logical, whether to calculate normalized frequencies}

\item{mc}{logical, whether to use multicore}

\item{verbose}{logical, whether to be verbose}
}
\value{
depends on the input, as this is a wrapper function
}
\description{
The function returns the frequencies of a query or a multiple queries
in sub-partitions defined by one or two dimensions. This is a wrapper function, so the output will depend
on the number of queries and dimensions provided.
}
\examples{
test <- partition("PLPRBTTXT", def=list(text_year="2009"), pAttribute=NULL)
dispersion(test, query="Integration", pAttribute="word", dim=c("text_date"))
foo <- dispersion(test, "Integration", c("text_date", "text_party"))
dispersion(test, '"Integration.*"', c("text_year")) # note the brackets when using regex!
}
\author{
Andreas Blaette
}
\seealso{
\code{crosstab-class}

count
}

