% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_method.R
\docType{methods}
\name{dotplot}
\alias{dotplot}
\alias{dotplot,textstat-method}
\alias{dotplot,features-method}
\alias{dotplot,featuresNgrams-method}
\alias{dotplot,partition-method}
\title{dotplot}
\usage{
dotplot(.Object, ...)

\S4method{dotplot}{textstat}(.Object, col, n = 20L, ...)

\S4method{dotplot}{features}(.Object, col = NULL, n = 20L, ...)

\S4method{dotplot}{featuresNgrams}(.Object, col = NULL, n = 20L, ...)

\S4method{dotplot}{partition}(.Object, col = "count", n = 20L, ...)
}
\arguments{
\item{.Object}{object}

\item{...}{further arguments that will be passed into the \code{dotchart} function}

\item{col}{column}

\item{n}{number}
}
\description{
dotplot
}
