% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates_method.R
\docType{methods}
\name{duplicates}
\alias{duplicates}
\alias{duplicates,partitionBundle-method}
\title{get duplicates}
\usage{
duplicates(.Object, ...)

\S4method{duplicates}{partitionBundle}(.Object, chars = "[a-zA-Z]",
  pAttribute = "word", sAttribute = "text_date", sample = 100, n = 2,
  threshold = 0.9, mc = FALSE, verbose = TRUE)
}
\arguments{
\item{.Object}{a \code{"partitionBundle"} object}

\item{...}{further parameters (verbose, progress, mc) to be passed into functions}

\item{chars}{a regex providing the characters to keep}

\item{sAttribute}{the s-attribute providing the date}

\item{sample}{will be passed as param size into sample-method to achieve a subset of partitionBundle and make char count faster}

\item{n}{number of days before and after a document was published}

\item{threshold}{numeric (0 < x < 1), the minimum similarity to qualify two documents as duplicates}

\item{mc}{whether to use multicore}

\item{verbose}{logical, whether to be verbose}
}
\description{
The method implements a procedure described by Fritz Kliche, Andre Blessing,
Urlich Heid and Jonathan Sonntag in the paper "The eIdentity Text
ExplorationWorkbench" presented at LREC 2014
(see \url{http://www.lrec-conf.org/proceedings/lrec2014/pdf/332_Paper.pdf}).
}
\details{
The method calls the (internal) \code{"getDuplicates"}-method that will make choices as follows:
(a) If two similar articles have been published on the same day, the shorter article will
be considered the duplicate; (b) if two similar articles were published on different days,
the article that appeared later will be considered the duplicate.
}
\examples{
\dontrun{
foo <- partitionBundle(
  "KEYWORDS",
  def=list(text_newspaper="guardian"),
  var=list(text_id=sAttributes("KEYWORDS", "text_id")[1:500]),
  pAttribute=NULL
 )
doubled <- duplicates(foo)
}
}

