% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/element_class.R
\docType{methods}
\name{element}
\alias{+,element,element-method}
\alias{as.bundle}
\alias{element}
\title{virtual superclass for elementary classes}
\usage{
as.bundle(object, ...)

\S4method{+}{element,element}(e1, e2)
}
\description{
virtual superclass for elementary classes
}

