% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_method.R
\docType{methods}
\name{encoding}
\alias{encoding}
\alias{encoding,partition-method}
\alias{encoding,partitionBundle-method}
\alias{encoding<-}
\title{get/set encoding slot of an object}
\usage{
encoding(object)

encoding(object) <- value

\S4method{encoding}{partition}(object)

\S4method{encoding}{partitionBundle}(object)
}
\arguments{
\item{object}{the object with an 'encoding'-slot}

\item{value}{value to be assigned}
}
\description{
Method for partition, partitionBundle objects.
}

