% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustEncoding_method.R
\name{as.utf8}
\alias{as.corpusEnc}
\alias{as.nativeEnc}
\alias{as.utf8}
\title{Conversion between corpus and native encoding.}
\usage{
as.utf8(x, from = "latin1")

as.nativeEnc(x, corpusEnc)

as.corpusEnc(x, corpusEnc)
}
\arguments{
\item{x}{the object (a character vector)}

\item{corpusEnc}{the encoding of the corpus (e.g. "latin1", "UTF-8")}
}
\description{
Utility functions to convert encoding between the native encoding and the 
encoding of the corpus.
}

