% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{enrich,keyness-method}
\alias{enrich,keyness-method}
\title{Enrich keyness object}
\usage{
\S4method{enrich}{keyness}(object, addPos = NULL, verbose = TRUE)
}
\arguments{
\item{object}{a keyness object}

\item{addPos}{logical, whether to add POS information to keyness object}

\item{verbose}{whether R should be talkative}
}
\description{
Wrapper for adding a statistic on pos attributes to keyness object
}

