% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{enrich,partitionCluster-method}
\alias{enrich,partitionCluster-method}
\title{enrich partition object (S4 method)}
\usage{
\S4method{enrich}{partitionCluster}(object, size = TRUE, tf = c(),
  meta = NULL, addPos = NULL, mc = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{a partition object}

\item{size}{logical, defaults to FALSE}

\item{tf}{character vector providing the p-attributes for which frequency
tables shall be generated}

\item{meta}{character vector providing s-attributes for which metainformation shall be supplied}

\item{addPos}{character vector providing p-attributes}

\item{mc}{logical whether to use multicore parallelization}

\item{verbose}{logical, defaults to TRUE}
}
\description{
Fill slots of a partition object: Add object size, metainformation, pos.
}

