% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/features.R, R/view.R
\docType{class}
\name{features-class}
\alias{features-class}
\alias{featuresCooccurrences-class}
\alias{featuresNgrams-class}
\alias{features_bundle-class}
\alias{summary,features-method}
\alias{show,features-method}
\alias{summary,features_bundle-method}
\alias{view,features-method}
\title{Feature selection by comparison (S4 class).}
\usage{
\S4method{summary}{features}(object)

\S4method{show}{features}(object)

\S4method{summary}{features_bundle}(object)

\S4method{view}{features}(.Object)
}
\arguments{
\item{object}{an object}

\item{.Object}{an object}
}
\description{
object resulting from features-method
}
\section{Slots}{

\describe{
\item{\code{corpus}}{Object of class \code{"character"}}

\item{\code{p_attribute}}{Object of class \code{"character"}}

\item{\code{encoding}}{Object of class \code{"character"}}

\item{\code{corpus}}{Object of class \code{"character"}}

\item{\code{stat}}{Object of class \code{"data.frame"}}

\item{\code{size_coi}}{Object of class \code{"numeric"}}

\item{\code{size_ref}}{Object of class \code{"numeric"}}

\item{\code{included}}{Object of class \code{"logical"} whether corpus of interest is included in reference corpus}

\item{\code{method}}{Object of class \code{"character"} statisticalTest used}

\item{\code{call}}{Object of class \code{"character"} the call that generated the object}

\item{\code{objects}}{an object of class \code{list}}
}}

\author{
Andreas Blaette
}
