% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features_method.R
\docType{methods}
\name{features,partition-method}
\alias{features,partition-method}
\alias{features}
\alias{features,count-method}
\alias{features,partitionBundle-method}
\alias{features,ngrams-method}
\title{Get features by comparison.}
\usage{
\S4method{features}{partition}(x, y, included = FALSE, method = "chisquare",
  verbose = FALSE)

\S4method{features}{count}(x, y, by = NULL, included = FALSE,
  method = "chisquare", verbose = TRUE)

\S4method{features}{partitionBundle}(x, y, included = FALSE,
  method = "chisquare", verbose = TRUE, mc = getOption("polmineR.mc"),
  progress = FALSE)

\S4method{features}{ngrams}(x, y, included = FALSE, method = "chisquare",
  verbose = TRUE, ...)
}
\arguments{
\item{x}{a partition or partitionBundle object}

\item{y}{a partition object, it is assumed that the coi is a subcorpus of
ref}

\item{included}{TRUE if coi is part of ref, defaults to FALSE}

\item{method}{the statistical test to apply (chisquare or log likelihood)}

\item{verbose}{logical, defaults to TRUE}

\item{by}{the columns used for merging, if NULL (default), the pAttribute of
x will be used}

\item{mc}{logical, whether to use multicore}

\item{progress}{logical}

\item{...}{further parameters}
}
\description{
The features of two objects, usually a partition defining a corpus of 
interest, and a partition defining a reference corpus are compared. 
The most important purpose is term extraction.
}
\examples{
\dontrun{
  use("polmineR.sampleCorpus")
  kauder <- partition("PLPRBTTXT", text_name="Volker Kauder", pAttribute="word")
  all <- partition("PLPRBTTXT", text_date = ".*", regex = TRUE, pAttribute = "word")
  terms_kauder <- features(x = kauder, y = all, included = TRUE)
  top100 <- subset(terms_kauder, rank_chisquare <= 100)
}
\dontrun{
  use("polmineR.sampleCorpus")
  kauder <- partition("PLPRBTTXT", text_name = "Volker Kauder", pAttribute = "word")
  x <- as(kauder, "count")
  all <- partition("PLPRBTTXT", text_date = ".*", regex = TRUE, pAttribute = "word")
  y <- as(all, "count")
  z <- features(x, y, included = TRUE)
  top100 <- subset(terms_kauder, rank_chisquare <= 100)
}
\dontrun{
  use("polmineR.sampleCorpus")
  byName <- partitionBundle("PLPRBTTXT", sAttribute="text_name")
  byName <- enrich(byName, pAttribute="word")
  all <- partition("PLPRBTTXT", text_date=".*", regex=TRUE, pAttribute="word")
  result <- features(byName, all, included=TRUE, progress=TRUE)
  dtm <- as.DocumentTermMatrix(result, col="chisquare")
}
}
\references{
Manning / Schuetze ...
}
\author{
Andreas Blaette
}
