% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R, R/cooccurrences.R
\docType{methods}
\name{features}
\alias{features}
\alias{features,partition-method}
\alias{features,count-method}
\alias{features,partition_bundle-method}
\alias{features,count_bundle-method}
\alias{features,ngrams-method}
\alias{features,Cooccurrences-method}
\title{Get features by comparison.}
\usage{
features(x, y, ...)

\S4method{features}{partition}(x, y, included = FALSE, method = "chisquare", verbose = FALSE)

\S4method{features}{count}(
  x,
  y,
  by = NULL,
  included = FALSE,
  method = "chisquare",
  verbose = TRUE
)

\S4method{features}{partition_bundle}(
  x,
  y,
  included = FALSE,
  method = "chisquare",
  verbose = TRUE,
  mc = getOption("polmineR.mc"),
  progress = FALSE
)

\S4method{features}{count_bundle}(
  x,
  y,
  included = FALSE,
  method = "chisquare",
  verbose = !progress,
  mc = getOption("polmineR.mc"),
  progress = FALSE
)

\S4method{features}{ngrams}(x, y, included = FALSE, method = "chisquare", verbose = TRUE, ...)

\S4method{features}{Cooccurrences}(x, y, included = FALSE, method = "ll", verbose = TRUE)
}
\arguments{
\item{x}{A \code{partition} or \code{partition_bundle} object.}

\item{y}{A \code{partition} object, it is assumed that the coi is a subcorpus of
ref}

\item{...}{further parameters}

\item{included}{TRUE if coi is part of ref, defaults to FALSE}

\item{method}{the statistical test to apply (chisquare or log likelihood)}

\item{verbose}{A \code{logical} value, defaults to TRUE}

\item{by}{the columns used for merging, if NULL (default), the p-attribute of
x will be used}

\item{mc}{logical, whether to use multicore}

\item{progress}{logical}
}
\description{
The features of two objects, usually a \code{partition} defining a corpus of
interest (coi), and a \code{partition} defining a reference corpus (ref) are compared.
The most important purpose is term extraction.
}
\examples{
use("polmineR")

kauder <- partition(
  "GERMAPARLMINI",
  speaker = "Volker Kauder", interjection = "speech",
  p_attribute = "word"
  )
all <- partition("GERMAPARLMINI", interjection = "speech", p_attribute = "word")

terms_kauder <- features(x = kauder, y = all, included = TRUE)
top100 <- subset(terms_kauder, rank_chisquare <= 100)
head(top100)

# a different way is to compare count objects
kauder_count <- as(kauder, "count")
all_count <- as(all, "count")
terms_kauder <- features(kauder_count, all_count, included = TRUE)
top100 <- subset(terms_kauder, rank_chisquare <= 100)
head(top100)

# get matrix with features (dontrun to keep time for examples short)
\dontrun{
use("RcppCWB")
docs <- partition_bundle("REUTERS", s_attribute = "id") \%>\%
  enrich( p_attribute = "word")
all <- corpus("REUTERS") \%>\%
  count(p_attribute = "word")
docs_terms <- features(docs[1:5], all, included = TRUE, progress = FALSE)
dtm <- as.DocumentTermMatrix(docs_terms, col = "chisquare", verbose = FALSE)
}
# Get features of objects in a count_bundle
ref <- corpus("GERMAPARLMINI") \%>\% count(p_attribute = "word")
cois <- corpus("GERMAPARLMINI") \%>\%
  subset(speaker \%in\% c("Angela Dorothea Merkel", "Hubertus Heil")) \%>\%
  split(s_attribute = "speaker") \%>\%
  count(p_attribute = "word")
y <- features(cois, ref, included = TRUE, method = "chisquare", progress = TRUE)
}
\references{
Baker, Paul (2006): \emph{Using Corpora in Discourse Analysis}. London: continuum, p. 121-149 (ch. 6).

Manning, Christopher D.; Schuetze, Hinrich (1999): \emph{Foundations of Statistical Natural Language
Processing}. MIT Press: Cambridge, Mass., pp. 151-189 (ch. 5).
}
\author{
Andreas Blaette
}
