% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getObjects}
\alias{getObjects}
\title{Get objects of a certain class.}
\usage{
getObjects(class, ns = ".GlobalEnv")
}
\arguments{
\item{class}{character, class to be looked for}

\item{ns}{character string, namespace to be searched}
}
\value{
a list with the partitions found in the namespace
}
\description{
Get objects of a certain class.
}

