% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getSlot}
\alias{getSlot}
\title{Get slot from object.}
\usage{
getSlot(x, name)
}
\arguments{
\item{x}{object to get slot from}

\item{name}{name of the slot}
}
\description{
Auxiliary function to unify access to slots of S4 or R6 object.
}

