% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTermFrequencies_method.R
\docType{methods}
\name{getTermFrequencies,partition-method}
\alias{getTermFrequencies,partition-method}
\title{get term frequencies}
\usage{
\S4method{getTermFrequencies}{partition}(.Object, pAttribute = c("word",
  "pos"), id2str = TRUE, mc = TRUE)
}
\arguments{
\item{.Object}{a partition object}

\item{pAttribute}{length 1 or 2 character vector, first element is supposed to be "word" or "lemma", second (optionally) "pos"}

\item{id2str}{logical, whether to add rownames}

\item{mc}{logical, whether to use multicore}
}
\description{
Get term frequencies based on corpus positions. Parallelization
is possible. (Performance gains decrease with the use
of additional cores, as all cores need to access the corpus on disk.)
}

