% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getTerms-method.R
\docType{methods}
\name{getTerms}
\alias{getTerms}
\alias{getTerms,character-method}
\title{get terms available in a corpus or partition}
\usage{
getTerms(.Object, ...)

\S4method{getTerms}{character}(.Object, pAttribute, robust = FALSE,
  verbose = TRUE, mc = FALSE)
}
\arguments{
\item{.Object}{the object}

\item{...}{further parameters}

\item{pAttribute}{the pAttribute}

\item{robust}{logical, whether to be robust}

\item{verbose}{logical, whether to be verbose}

\item{mc}{logical, whether to use multicore}
}
\description{
get terms available in a corpus or partition
}

