% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTokenStream_method.R
\docType{methods}
\name{getTokenStream}
\alias{getTokenStream}
\alias{getTokenStream,matrix-method}
\alias{getTokenStream,partition-method}
\title{get token stream}
\usage{
getTokenStream(.Object, ...)

\S4method{getTokenStream}{matrix}(.Object, corpus, pAttribute, encoding,
  collapse = NULL, cposAsNames = FALSE)

\S4method{getTokenStream}{partition}(.Object, pAttribute, collapse = NULL,
  cposAsNames = FALSE)
}
\arguments{
\item{.Object}{an object of classe \code{matrix} or \code{partition}}

\item{...}{further arguments}

\item{corpus}{the CWB corpus}

\item{pAttribute}{the pAttribute to decode}

\item{encoding}{encoding to use}

\item{collapse}{character string length 1}

\item{cposAsNames}{logical, whether to return cpos as names of the tokens}
}
\description{
get token stream
}

