% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\docType{methods}
\name{highlight}
\alias{highlight}
\alias{highlight,character-method}
\alias{highlight,html-method}
\alias{highlight,kwic-method}
\title{Highlight tokens in output.}
\usage{
highlight(.Object, ...)

\S4method{highlight}{character}(.Object, highlight = list(), ...)

\S4method{highlight}{html}(.Object, highlight = list(), ...)

\S4method{highlight}{kwic}(.Object, highlight = list(), regex = FALSE,
  perl = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{.Object}{A \code{html}, \code{character}, a \code{kwic} object.}

\item{...}{Terms to be highlighted can be passed in as named character
vectors of terms (or regular expressions); the name then needs to be a
valid color name.}

\item{highlight}{A \code{list} of character or integer vectors, the names
need to provide the colors, the values of the vector the term to be matched
or a corpus position.}

\item{regex}{Logical, whether character vectors are interpreted as regular
expressions.}

\item{perl}{Logical, whether to use perl-style regular expressions for
highlighting when regex is \code{TRUE}.}

\item{verbose}{Logical, whether to output messages.}
}
\description{
Highlight tokens in fulltext based on exact match, a regular expression or
corpus position in \code{kwic} output or \code{html} document.
}
\examples{
use("polmineR")
P <- partition("REUTERS", places = "argentina")
H <- html(P)
Y <- highlight(H, list(lightgreen = "higher"))
if (interactive()) htmltools::html_print(Y)

# highlight matches for a CQP query
H2 <- highlight(
  H,
  highlight = list(yellow = cpos(hits(P, query = '"prod.*"', cqp = TRUE)))
)

# the method can be used in pipe
if (require("magrittr")){
  P \%>\% html() \%>\% highlight(list(lightgreen = "1986")) -> H
  P \%>\% html() \%>\% highlight(list(lightgreen = c("1986", "higher"))) -> H
  P \%>\% html() \%>\% highlight(list(lightgreen = 4020:4023)) -> H
}

# use highlight for kwic output
K <- kwic("REUTERS", query = "barrel")
K2 <- highlight(K, highlight = list(yellow = c("oil", "price")))
if (interactive()) K2
}
