% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/hits.R
\docType{class}
\name{hits_class}
\alias{hits_class}
\alias{hits-class}
\alias{sample,hits-method}
\title{Hits class.}
\usage{
\S4method{sample}{hits}(x, size)
}
\arguments{
\item{x}{A \code{hits} object.}

\item{size}{A non-negative integer giving the number of items to choose.}
}
\description{
Hits class.
}
\section{Slots}{

\describe{
\item{\code{stat}}{A \code{"data.table"} with at least three columns: Columns
\code{cpos_left} and \code{cpos_right} indicate left and right corpus
positions of the regions that yielded hits for a query or queries. Column
\code{query} specifies the (CQP) query that resulted in a hit.}

\item{\code{corpus}}{A length-one \code{"character"} vector, ID of the corpus with
hits for query or queries.}

\item{\code{query}}{Object of class \code{"character"}, query or queries for}

\item{\code{p_attribute}}{The p-attribute that has been queried, a length-one
\code{character} vector.}

\item{\code{encoding}}{Length-one \code{character} vector, the encoding of the
corpus.}

\item{\code{name}}{Length-one \code{characte} vector, name of the object.}
}}

