% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_method.R
\docType{methods}
\name{html}
\alias{html}
\alias{html,character-method}
\alias{html,kwic-method}
\alias{html,partition-method}
\alias{html,partitionBundle-method}
\alias{show,html-method}
\title{restore fulltext as html}
\usage{
html(object, ...)

\S4method{html}{character}(object)

\S4method{html}{partition}(object, meta = getOption("polmineR.meta"),
  highlight = list(), cqp = FALSE, tooltips = NULL, cpos = FALSE,
  verbose = FALSE, ...)

\S4method{html}{partitionBundle}(object, filename = c(), type = "debate")

\S4method{html}{kwic}(object, i, type)
}
\arguments{
\item{object}{the object}

\item{...}{further parameters}

\item{meta}{metadata for output}

\item{highlight}{list with regex to be highlighted}

\item{cqp}{logical}

\item{tooltips}{tooltips for highlighted text}

\item{cpos}{logical}

\item{verbose}{logical}

\item{filename}{the filename}

\item{type}{the partition type}

\item{i}{to be checked}
}
\description{
restore fulltext as html
}

