% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.corpus_function.R
\name{install.corpus}
\alias{install.corpus}
\title{Install packaged corpus from repository.}
\usage{
install.corpus(pkgs, repo = "http://polmine.sowi.uni-due.de/packages", ...)
}
\arguments{
\item{pkgs}{names of data packages with corpora}

\item{repo}{URL of the repository}

\item{...}{further parameters that will be passed into \code{install.packages}}
}
\description{
Convenience function for making the installation of indexed CWB corpora
wrapped into R data packages as easy as possible.
}
\examples{
\dontrun{
}
}

