% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.cwb_function.R
\name{install.cwb}
\alias{install.cwb}
\title{Install Corpus Workbench}
\usage{
install.cwb()
}
\description{
Install precompiled version of the Corpus Workbench in the directory
extdata/cwb/CWB. This is a precondition to offer full CQP functionality
on Windows systems.
}
