% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_nested}
\alias{is_nested}
\title{Check whether s-attributes of corpus are nested}
\usage{
is_nested(x)
}
\arguments{
\item{x}{A \code{character} vector with corpus ID or a \code{corpus} object.}
}
\value{
A \code{logical} value: \code{FALSE} is data structure is flat and \code{TRUE} if
data structure is nested.
}
\description{
Simple test whether the attribute size of all s-attributes of a corpus is
identical (flat import XML) or not (nested import XML).
}
\examples{
use("polmineR")
use("RcppCWB")
}
