\docType{methods}
\name{mail-concordances-method}
\alias{mail,concordances-method}
\alias{mail-concordances-method}
\title{mail concordances}
\usage{
\S4method{mail}{concordances}(object, to = NULL, nrow = NULL,
  fileFormat = c("csv", "xlsx"))
}
\arguments{
  \item{object}{the concordance object}

  \item{to}{the receiver of the mail message}

  \item{nrow}{the number of rows}

  \item{fileFormat}{csv or xlsx, or both}
}
\description{
still experimental
}

