\docType{methods}
\name{mail-kwic-method}
\alias{mail,kwic-method}
\alias{mail-kwic-method}
\title{mail kwic/concordances}
\usage{
\S4method{mail}{kwic}(object, to = NULL, nrow = NULL,
  fileFormat = c("csv", "xlsx"))
}
\arguments{
  \item{object}{the concordance object}

  \item{to}{the receiver of the mail message}

  \item{nrow}{the number of rows of the table (if NULL, the
  whole table will be sent)}

  \item{fileFormat}{csv or xlsx, or both}
}
\description{
still experimental
}

