% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/keynessBundle-class.R, R/keynessBundle-methods.R
\docType{class}
\name{keynessBundle-class}
\alias{as.TermDocumentMatrix,keynessBundle-method}
\alias{dim,keynessBundle-method}
\alias{keynessBundle-class}
\title{keynessBundle-class}
\usage{
\S4method{as.TermDocumentMatrix}{keynessBundle}(x, col = "chiSquare",
  rmBlank = TRUE, verbose = TRUE)

\S4method{dim}{keynessBundle}(x)
}
\arguments{
\item{x}{object}

\item{col}{the column to extract}

\item{rmBlank}{remove blank rows}

\item{verbose}{logical, whether to be verbose}
}
\description{
The class inherits from the bundle class.
}
\section{Slots}{

\describe{
\item{\code{objects}}{an object of class \code{list}}
}}

