% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/keynessCluster-class.R, R/keynessCluster-methods.R
\docType{class}
\name{keynessCluster-class}
\alias{as.TermDocumentMatrix,keynessCluster-method}
\alias{dim,keynessCluster-method}
\alias{keynessCluster-class}
\title{keynessCluster-class}
\usage{
\S4method{as.TermDocumentMatrix}{keynessCluster}(x, col = "chiSquare",
  rmBlank = TRUE, verbose = TRUE)

\S4method{dim}{keynessCluster}(x)
}
\arguments{
\item{x}{object}

\item{col}{the column to extract}

\item{rmBlank}{remove blank rows}

\item{verbose}{logical, whether to be verbose}
}
\description{
The class inherits from the bundle class.
}
\section{Slots}{

\describe{
\item{\code{objects}}{an object of class \code{list}}
}}

