% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kwic-method.R
\docType{methods}
\name{kwic,partition-method}
\alias{kwic}
\alias{kwic,context-method}
\alias{kwic,partition-method}
\alias{show,kwic-method}
\title{KWIC output}
\usage{
\S4method{kwic}{partition}(object, query, leftContext = 0, rightContext = 0,
  meta = NULL, pAttribute = "word", collocate = c(), verbose = TRUE)
}
\arguments{
\item{object}{a partition object}

\item{query}{what to look up}

\item{leftContext}{to the left}

\item{rightContext}{to the right}

\item{meta}{metainformation to display}

\item{pAttribute}{typically 'word' or 'lemma'}

\item{collocate}{only show kwic if a certain word is present}

\item{verbose}{whether to be talkative}
}
\description{
Prepare and show 'keyword in context' (KWIC). The same result can be achieved by
applying the kwich method on either a partition or a context object.
}
\examples{
bt <- partition("PLPRTXT", def=list(text_date=".*"), method="grep")
foo <- kwic(bt, "Integration")
foo <- kwic(bt, "Integration", leftContext=20, rightContext=20, meta=c("text_date", "text_name", "text_party"))
}

