% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/label.R
\docType{class}
\name{labels-class}
\alias{labels-class}
\alias{initialize,labels-method}
\alias{labels,kwic-method}
\alias{labels<-}
\alias{labels<-,kwic,list-method}
\alias{labels<-,textstat,list-method}
\alias{edit,kwic-method}
\title{Labels class and method}
\usage{
\S4method{initialize}{labels}(.Object, n)

\S4method{labels}{kwic}(object, i, j, value)

labels(x) <- value

\S4method{labels}{kwic,list}(x) <- value

\S4method{labels}{textstat,list}(x) <- value

\S4method{edit}{kwic}(name, viewer = shiny::paneViewer(minHeight = 550),
  ...)
}
\arguments{
\item{.Object}{A \code{labels} object.}

\item{n}{The integer index of a label to retrieve or modify.}

\item{object}{A \code{labels} object.}

\item{i}{The row number (single \code{integer} value) of the
\code{data.table} of the \code{labels} object where a new value shall be
assigned.}

\item{j}{The column number (single \code{integer} value) of the
\code{data.table} of the \code{labels} object where a new value shall be
assigned.}

\item{value}{A value to assign.}

\item{x}{An object to be labelled.}

\item{name}{An S4 object with a slot \code{labels} (e.g. an object of class
\code{kwic}).}

\item{viewer}{The viewer to use, see \code{\link[shiny]{viewer}}.}

\item{...}{Passed into \code{rhandsontable::rhandsontable}, can be used for
settings such as \code{height} etc.}
}
\value{
The modified input object is returned invisibly.
}
\description{
The \code{labels}-class in combination with the \code{label()} generic method
offers the tools to implement a workflow to categorize/annotate
results by assigning labels.

The \code{labels} class and the \code{label}-method supports generating and
working with labelled data. They rely on data in a slot \code{labels} of the
S4 object to be modified. Operations are deliberately in-place.

The \code{edit}-method is designed to be used in a RStudio session. Calling
it generates a shiny gadget (see
\url{https://shiny.rstudio.com/articles/gadgets.html}) shown in the viewer
pane of RStudio.

The method returns the modified input object. Note however that changes of
labels are deliberately in-place operations. Accordingly, the input object is
changed even if you do not close the gadget "properly" by hitting the "Done"
button and catch the modified object, which may be forgotten easily and would
be painful after the work that may have been invested.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: When initializing a \code{labels} class object, an
empty \code{data.table} is assigned.
}}

\section{Slots}{

\describe{
\item{\code{labels}}{A \code{data.table} that keeps the assigned labels.}
}}

\examples{
use("polmineR")

# upon initializing a kwic object, there is a minimal labels object
# in the labels slot of the kwic object, which we can get using the
# labels-method
o <- kwic("REUTERS", query = "oil")
labels(o) # see the result (a data.table)

# assign new columns as follows, using the reference semantics of the
# data.table you get by calling the labels-method on an object 
labels(o)[, "class" := factor(levels = c("a", "b", "c"))]
labels(o) <- list(name = "class", what = factor(x = "a", levels = c("a", "b", "c")))
labels(o) <- list(name = "description", what = "")
labels(o) # see the result

# the label-method can be used to assign values; note that is an in-place
# operation using the reference semantics of the data.table
labels(o, i = 77, j = 1, value = FALSE)
labels(o, i = 78, j = 1, value = FALSE)
labels(o)

\dontrun{
edit(o)
labels(o) # to see changes made

# maybe we want additional metadata
enrich(o, s_attributes = "places")
edit(o)
labels(o)

# to get some extra context
o <- enrich(o, extra = 5L, table = TRUE)
edit(o)

# lineview may be better when you use a lot of extra context
options(polmineR.lineview = TRUE)
o <- enrich(o, extra = 20L)
edit(o)

#' edit(o, js = "DataTables")
labels(o) # to see changes made
}
}
