% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mail-method.R
\docType{methods}
\name{mail-crosstab-method}
\alias{mail,crosstab-method}
\alias{mail-crosstab-method}
\title{mail crosstab}
\usage{
\S4method{mail}{crosstab}(object, to = NULL, nrow = NULL,
  fileFormat = c("csv", "xlsx"))
}
\arguments{
\item{object}{the crosstab object}

\item{to}{the receiver of the mail message}

\item{nrow}{the number of rows of the table (if NULL, the whole table will be sent)}

\item{fileFormat}{csv or xlsx, or both}
}
\description{
For exporting.
}

