\docType{methods}
\name{mail}
\alias{mail}
\alias{mail,keyness-method}
\title{mail a result}
\usage{
\S4method{mail}{keyness}(object, to = NULL, nrow = NULL,
  fileFormat = c("csv", "xlsx"))
}
\arguments{
  \item{object}{an object with some statistics}

  \item{to}{the receiver of the mail message}

  \item{nrow}{the number of rows}

  \item{fileFormat}{either csv or xlsx, or both}
}
\description{
still experimental
}

