% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/means_method.R
\docType{methods}
\name{means}
\alias{means}
\alias{means,DocumentTermMatrix-method}
\title{calculate means}
\usage{
means(.Object, ...)

\S4method{means}{DocumentTermMatrix}(.Object, dim = 1)
}
\arguments{
\item{.Object}{object to work on}

\item{...}{further parameters}

\item{dim}{numeric, 1 or 2 whether to work on rows or columns}
}
\description{
calculate means
}

