% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_method.R
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,character-method}
\alias{meta,partition-method}
\title{metainformation}
\usage{
meta(.Object, ...)

\S4method{meta}{character}(.Object, sAttributes, struc)

\S4method{meta}{partition}(.Object, sAttributes)
}
\arguments{
\item{.Object}{a \code{"partition"} object}

\item{...}{further parameters}

\item{sAttributes}{a character vector}

\item{struc}{integer vector, strucs to get metadata for}
}
\value{
a \code{"data.frame"}
}
\description{
This function is usually part of the procedure to set up a partition. Based
on the corpus positions that have initially been set up, a table with all
metadata is added to the object, and a list with the attributes and the
values that occur.
}
\details{
The values of the metadata are there to determine which subcorpora can be
generated. The table with all metadata can be added as it is a basis for the
values. It can be omitted for performance and memora reasons.
}
\author{
Andreas Blaette
}

