% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,character-method}
\alias{meta,partition-method}
\alias{meta-method}
\title{get values of metadata}
\usage{
\S4method{meta}{partition}(object, sAttribute)
}
\arguments{
\item{object}{a partition object or a character vector (length 1) supplying the name of a corpus}

\item{sAttribute}{the s-attribute to be looked up}
}
\value{
a character vector
}
\description{
Returns the values that an sAttribute takes within a partition.
}

