% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nchars_method.R
\docType{methods}
\name{nchars}
\alias{nchars}
\alias{nchars,partition-method}
\alias{nchars,partitionBundle-method}
\title{Count the number of characters}
\usage{
nchars(x, ...)

\S4method{nchars}{partition}(x, pAttribute, regexCharsToKeep = "[a-zA-Z]",
  toLower = TRUE, decreasing = TRUE)

\S4method{nchars}{partitionBundle}(x, decreasing = TRUE, mc = FALSE, ...)
}
\arguments{
\item{...}{parameters that will be passed}

\item{pAttribute}{the p-attribute}

\item{regexCharsToKeep}{if not NULL, a regex}

\item{toLower}{whether to lower tokens}

\item{decreasing}{logical, passed into order call}

\item{mc}{logical}

\item{.Object}{object}

\item{.Object}{object}

\item{pAttribute}{the p-attribute}

\item{regexCharsToKeep}{if NULL, counts for all charactrs will be returned, else a regex indicating which characters to include in the counting}

\item{toLower}{whether to lower tokens}

\item{mc}{logical}

\item{...}{parameters that will be passed}

\item{decreasing}{logical, passed into order call}
}
\description{
Count the number of characters

count characters
}

