% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ngrams_method.R
\name{ngrams}
\alias{ngrams}
\title{get ngrams}
\usage{
ngrams(.Object, ...)
}
\arguments{
\item{.Object}{object}

\item{n}{number of token}

\item{pAttribute}{p-attribute}
}
\description{
get ngrams
}

