% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_method.R
\docType{methods}
\name{noise,DocumentTermMatrix-method}
\alias{noise,DocumentTermMatrix-method}
\title{detect noise}
\usage{
\S4method{noise}{DocumentTermMatrix}(.Object, minSum = 2,
  minTfIdfMean = 0.005, sparse = 0.995, stopwordsLanguage = "german",
  minNchar = 2, verbose = TRUE)
}
\arguments{
\item{.Object}{an .Object of class \code{"DocumentTermMatrix"}}

\item{minSum}{minimum colsum (for DocumentTermMatrix) to qualify a term as non-noise}

\item{minTfIdfMean}{minimum mean value for tf-idf to qualify a term as non-noise}

\item{sparse}{will be passed into \code{"removeSparseTerms"} from \code{"tm"}-package}

\item{stopwordsLanguage}{e.g. "german", to get stopwords defined in the tm package}

\item{minNchar}{min char length ti qualify a term as non-noise}
}
\value{
a list
}
\description{
detect noise
}

