% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textstat_methods.R
\docType{methods}
\name{nrow,textstat-method}
\alias{nrow,textstat-method}
\title{exportMethod nrow}
\usage{
\S4method{nrow}{textstat}(x)
}
\arguments{
\item{x}{textstat object}
}
\description{
exportMethod nrow
}

