% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencpu.R
\name{ocpu_exec}
\alias{ocpu_exec}
\alias{opencpu}
\title{Execute code on OpenCPU server}
\usage{
ocpu_exec(fn, server, user = NULL, password = NULL, do.call = FALSE,
  ...)
}
\arguments{
\item{fn}{Name of the function/method to execute on remote server (length-one
\code{character} vector).}

\item{server}{The IP/URL of the remote OpenCPU server.}

\item{user}{A username required for authentication, if necessary.}

\item{password}{A password required for authentication, if necessary.}

\item{do.call}{Logical, if \code{TRUE}, the function \code{fn} is passed into
a call of \code{do.call}, which offers some flexibility.}

\item{...}{Further arguments passed into the method/function call.}
}
\description{
\code{ocpu_exec} will execute a function/method \code{fn} on an OpenCPU server
(specified by argument \code{server}), using three dots (\code{...}) to pass
arguments. It is the worker of methods defined for \code{remote_corpus},
\code{remote_subcorpus} and \code{remote_partition} objects.
}
\examples{
\dontrun{
# Get polmineR version installed on remote server
ocpu_exec(
  fn = "packageVersion",
  server = Sys.getenv("OPENCPU_SERVER"),
  do.call = TRUE,
  pkg = "polmineR"
)

ocpu_exec(
  fn = "subset",
  server = Sys.getenv("OPENCPU_SERVER"),
  do.call = TRUE,
  x = iris,
  subset = substitute(Sepal.Width > 4.0)
)
}
}
