% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/papply_function.R
\name{papply}
\alias{papply}
\title{mclapply with progress}
\usage{
papply(index, f, verbose = FALSE, mc = TRUE, param = list())
}
\arguments{
\item{index}{index}

\item{f}{function}

\item{verbose}{logical}

\item{mc}{logical}

\item{param}{list}
}
\description{
mclapply with progress
}

