% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pAttributes_method.R, R/partition_class.R, R/partition_methods.R, R/session_methods.R
\docType{methods}
\name{pAttributes,partition-method}
\alias{[,partition,ANY,ANY,ANY-method}
\alias{[,partition-method}
\alias{[[,partition,ANY-method}
\alias{as.data.frame,partition-method}
\alias{as.partitionBundle}
\alias{as.partitionBundle,partition-method}
\alias{corpus,partition-method}
\alias{dissect}
\alias{dissect,partition-method}
\alias{export}
\alias{export,partition-method}
\alias{hist,partition-method}
\alias{length,partition-method}
\alias{name,partition-method}
\alias{name<-,partition,character-method}
\alias{pAttribute,partition-method}
\alias{pAttributes,partition-method}
\alias{partition-class}
\alias{show,partition-method}
\alias{size}
\alias{size,partition-method}
\alias{split}
\title{partition class definition}
\usage{
\S4method{pAttributes}{partition}(.Object)

\S4method{size}{partition}(x)

\S4method{[}{partition,ANY,ANY,ANY}(x, i)

\S4method{[[}{partition,ANY}(x, i)

\S4method{pAttribute}{partition}(object, from, to, mc = TRUE,
  verbose = TRUE)

\S4method{name}{partition}(x)

\S4method{name}{partition,character}(x) <- value

\S4method{dissect}{partition}(object, dim, verbose = FALSE)

\S4method{length}{partition}(x)

\S4method{as.data.frame}{partition}(x)

\S4method{hist}{partition}(x, ...)

\S4method{corpus}{partition}(object)
}
\arguments{
\item{.Object}{a partition object}

\item{object}{a partition object}

\item{verbose}{logical, whether to be verbose}

\item{dim}{don't know}
}
\description{
partition class definition
}
\section{Slots}{

\describe{
\item{\code{name}}{Object of class \code{"character"} a name that may be useful}

\item{\code{corpus}}{Object of class \code{"character"} the CWB corpus the partition is based on}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{sAttributes}}{Object of class \code{"list"} s-attributes specifying the partition}

\item{\code{explanation}}{Object of class \code{"character"} an explanation of the partition}

\item{\code{cpos}}{Object of class \code{"matrix"} corpus positions}

\item{\code{pos}}{Object of class \code{"list"} with tables "abs", "rel" and "max"}

\item{\code{size}}{Object of class \code{"numeric"} total size of the partition}

\item{\code{metadata}}{Object of class \code{"list"} metadata information}

\item{\code{strucs}}{Object of class \code{"numeric"} the strucs defining the partition}

\item{\code{tf}}{Object of class \code{"matrix"}: term frequencies}

\item{\code{pAttribute}}{Object of class \code{"character"} indicating the pAttribute of the
count in slot tf}

\item{\code{xml}}{Object of class \code{"character"} whether the xml is flat or nested}

\item{\code{sAttributeStrucs}}{Object of class \code{"character"} the base node}

\item{\code{call}}{Object of class \code{"character"} the call that generated the partition}
}}
\section{Methods}{

  \describe{
   \item{show}{\code{signature(object = "partition")}: Display essential information }
   \item{addPos}{\code{signature(object="partition")}: add list with most frequent pos for a token }
   \item{trim}{\code{signature(object="partition")}: trim a partition object }
   \item{tf}{\code{signature(object="partition")}: get term frequencies }
   \item{as.partitionBundle}{\code{signature(object="partition")}: transform a partition object into a partitionBundle (to add further objects) }
   \item{[}{get frequency of a query}
   \item{html}{transform partition to html}
   }
}
\author{
Andreas Blaette
}

