% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/partition_method.R
\docType{methods}
\name{partition}
\alias{partition}
\alias{partition,character-method}
\alias{partition,list-method}
\alias{partition,missing-method}
\alias{partition,partition-method}
\alias{partition,session-method}
\alias{partitionGadget}
\title{Initialize a partition}
\usage{
partition(object, ...)

\S4method{partition}{character}(object, def = NULL, name = c(""),
  encoding = NULL, pAttribute = NULL, meta = NULL, regex = FALSE,
  xml = "flat", id2str = TRUE, type = NULL, mc = FALSE,
  verbose = TRUE)

\S4method{partition}{list}(object, name = c(""), encoding = NULL,
  pAttribute = NULL, meta = NULL, regex = FALSE, xml = "flat",
  id2str = TRUE, type = NULL, mc = FALSE, verbose = TRUE)

\S4method{partition}{session}(object)

\S4method{partition}{partition}(object, def, name = c(""), regex = FALSE,
  pAttribute = NULL, id2str = TRUE, type = NULL, verbose = TRUE,
  mc = FALSE, ...)

\S4method{partition}{missing}()

partitionGadget()
}
\arguments{
\item{object}{character-vector - the CWB-corpus to be used}

\item{...}{parameters passed into the partition-method}

\item{def}{list consisting of a set of character vectors (see
details)}

\item{name}{name of the new partition, defaults to "noName"}

\item{encoding}{encoding of the corpus (typically "LATIN1 or "(UTF-8)), if NULL, the encoding provided in the registry file of the corpus (charset="...") will be used b}

\item{pAttribute}{the pAttribute(s) for which term frequencies shall be retrieved}

\item{meta}{a character vector}

\item{regex}{logical (defaults to FALSE), if TRUE, the s-attributes provided will be handeled as regular expressions; the length of the character vectors with s-attributes then needs to be 1}

\item{xml}{either 'flat' (default) or 'nested'}

\item{id2str}{whether to turn token ids to strings (set FALSE to minimize object.size / memory consumption)}

\item{type}{character vector (length 1) specifying the type of corpus / partition (e.g. "plpr")}

\item{mc}{whether to use multicore (for counting terms)}

\item{verbose}{logical, defaults to TRUE}
}
\value{
An object of the S4 class 'partition'
}
\description{
Set up an object of the partition class. Frequency lists are computeted if a pAttribute is provided.
}
\details{
The function sets up a partition based on a list of s-attributes with respective values.
The s-attributes defining the partition are a list, e.g. list(text_type="speech", text_year="2013").
The values of the list may contain regular expressions. To use regular expression syntex, set the
parameter regex to \code{"TRUE"}. Regular expressions are passed into grep, i.e. the regex syntax
used in R needs to be used (double backlashes etc.).
}
\examples{
use(polmineR.sampleCorpus)
spd <- partition("PLPRBTTXT", def=list(text_party="SPD", text_type="speech"))
kauder <- partition("PLPRBTTXT", def=list(text_name="Volker Kauder"), pAttribute=c("word"))
merkel <- partition("PLPRBTTXT", list(text_name=".*Merkel"), pAttribute="word", regex=TRUE)
sAttributes(merkel, "text_date")
sAttributes(merkel, "text_name")
merkel <- partition("PLPRBTTXT", list(text_name="Angela Dorothea Merkel", text_date="2009-11-10", text_type="speech"), pAttribute="word")
merkel <- subset(merkel, !word \%in\% punctuation)
merkel <- subset(merkel, !word \%in\% tm::stopwords("de"))
}
\author{
Andreas Blaette
}

