% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_bundle.R
\docType{methods}
\name{partition_bundle}
\alias{partition_bundle}
\alias{partition_bundle,partition-method}
\alias{partition_bundle,character-method}
\alias{partition_bundle,context-method}
\title{Generate a bundle of partitions}
\usage{
partition_bundle(.Object, ...)

\S4method{partition_bundle}{partition}(.Object, s_attribute, values = NULL,
  prefix = "", mc = getOption("polmineR.mc"), verbose = TRUE,
  progress = FALSE, ...)

\S4method{partition_bundle}{character}(.Object, s_attribute, values = NULL,
  prefix = "", mc = getOption("polmineR.mc"), verbose = TRUE,
  progress = FALSE, xml = "flat", ...)

\S4method{partition_bundle}{context}(.Object, mc = getOption("polmineR.mc"),
  verbose = FALSE, progress = TRUE)
}
\arguments{
\item{.Object}{character string, a partition, or a list}

\item{...}{parameters to be passed into partition-method (see respective documentation)}

\item{s_attribute}{the s-attribute to vary}

\item{values}{values the s-attribute provided shall assume}

\item{prefix}{a character vector that will be attached as a prefix to partition names}

\item{mc}{logical, whether to use multicore parallelization}

\item{verbose}{logical, whether to provide progress information}

\item{progress}{logical, whether to show progress bar}

\item{xml}{logical}
}
\value{
S4 class \code{partition_bundle}, with list of partition objects in slot 'objects'
}
\description{
A \code{partition_bundle} object is a S4 class object. 
partition_bundle,character-method will create a bundle of partitions,
but not yet enriched.
}
\examples{
use("polmineR")
bt2009 <- partition("GERMAPARLMINI", date = "2009-.*", regex = TRUE)
pBundle <- partition_bundle(bt2009, s_attribute = "date", progress = TRUE, p_attribute = "word")
dtm <- as.DocumentTermMatrix(pBundle, col = "count")
summary(pBundle)
btBundle <- partition_bundle("GERMAPARLMINI", s_attribute = "date")
}
\seealso{
\code{\link{partition}} and \code{\link{bundle}}
}
\author{
Andreas Blaette
}
