% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/p_attributes.R, R/partition.R,
%   R/aggregate.R, R/as.markdown.R, R/enrich.R, R/view.R
\docType{class}
\name{partition_class}
\alias{partition_class}
\alias{partition-class}
\alias{show,partition-method}
\alias{[,partition,ANY,ANY,ANY-method}
\alias{[,partition-method}
\alias{as.partition_bundle}
\alias{as.partition_bundle,partition-method}
\alias{export}
\alias{export,partition-method}
\alias{split}
\alias{plprPartition-class}
\alias{pressPartition-class}
\alias{p_attributes,partition-method}
\alias{name,partition-method}
\alias{name<-,partition,character-method}
\alias{length,partition-method}
\alias{hist,partition-method}
\alias{length,partition-method}
\alias{is.partition}
\alias{partition_add_cpos}
\alias{aggregate,partition-method}
\alias{as.markdown,partition-method}
\alias{enrich,partition-method}
\alias{view,partition-method}
\alias{view,cooccurrences-method}
\title{Partition class and methods.}
\usage{
\S4method{p_attributes}{partition}(.Object, p_attribute = NULL, ...)

\S4method{name}{partition}(x)

\S4method{name}{partition,character}(x) <- value

\S4method{length}{partition}(x)

\S4method{hist}{partition}(x, ...)

\S4method{length}{partition}(x)

is.partition(x)

partition_add_cpos(.Object, xml = "flat", regex = FALSE)

\S4method{aggregate}{partition}(x)

\S4method{as.markdown}{partition}(.Object, meta = getOption("polmineR.meta"),
  template = getTemplate(.Object), cpos = TRUE, cutoff = NULL,
  verbose = FALSE, ...)

\S4method{enrich}{partition}(.Object, size = FALSE, p_attribute = NULL,
  decode = TRUE, verbose = TRUE, mc = FALSE, ...)

\S4method{view}{partition}(.Object)

\S4method{view}{cooccurrences}(.Object)
}
\arguments{
\item{.Object}{a partition object}

\item{p_attribute}{a p-attribute (for enriching)}

\item{...}{further parameters}

\item{x}{a partition object}

\item{value}{value}

\item{xml}{either "flat" or "nested"}

\item{regex}{logical}

\item{meta}{...}

\item{template}{template to use}

\item{cpos}{...}

\item{cutoff}{maximum number of tokens to decode}

\item{verbose}{logical}

\item{size}{logical}

\item{decode}{logical}

\item{mc}{logical or, if numeric, providing the number of cores}

\item{partition}{partition object with list of s-attributes}
}
\description{
S4 partition class and methods for instances of class partition.

\code{partition_add_cpos} will generate the matrix with corpus positions (regions)
based on the s-attributes as defined in the respective slot of the partition
object. The result is found in the slot \code{cpos} of the \code{partition}
object that is returned.

The method \code{aggregate} will deflate the matrix in the slot \code{cpos},
i.e. it checks for each new row in the matrix whether it increments the end
of the previous region (by 1), and ensure that the cpos matrix defines
disjoined regions.
}
\section{Slots}{

\describe{
\item{\code{name}}{Object of class \code{"character"} a name that may be useful}

\item{\code{corpus}}{Object of class \code{"character"} the CWB corpus the partition is based on}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{s_attributes}}{Object of class \code{"list"} s-attributes specifying the partition}

\item{\code{explanation}}{Object of class \code{"character"} an explanation of the partition}

\item{\code{cpos}}{Object of class \code{"matrix"} corpus positions}

\item{\code{annotations}}{Object of class \code{"list"}}

\item{\code{pos}}{Object of class \code{"list"} with tables "abs", "rel" and "max"}

\item{\code{size}}{Object of class \code{"numeric"} total size of the partition}

\item{\code{metadata}}{Object of class \code{"data.frame"} metadata information}

\item{\code{strucs}}{Object of class \code{"numeric"} the strucs defining the partition}

\item{\code{p_attribute}}{Object of class \code{"character"} indicating the p_attribute of the
count in slot tf}

\item{\code{xml}}{Object of class \code{"character"} whether the xml is flat or nested}

\item{\code{s_attribute_strucs}}{Object of class \code{"character"} the base node}

\item{\code{call}}{Object of class \code{"character"} the call that generated the partition}
}}

\examples{
P <- new(
  "partition",
  cpos = matrix(data = c(1:10, 20:29), ncol = 2, byrow = TRUE),
  stat = data.table::data.table()
)
P2 <- aggregate(P)
P2@cpos
}
\seealso{
The \code{partition}-class inherits from the \code{\link{textstat-class}}, see
respective documentation to learn more.
}
\author{
Andreas Blaette
}
