% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_stream.R
\name{partition_to_string}
\alias{partition_to_string}
\title{Decode as String.}
\description{
Decode as String.
}
\examples{
use("polmineR")
p <- partition("GERMAPARLMINI", date = "2009-11-10", speaker = "Angela Dorothea Merkel")
s <- as(p, "String")
}
