% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle_class.R, R/partitionBundle_class.R, R/html_method.R, R/partitionBundle_methods.R, R/sAttributes_method.R
\docType{methods}
\name{names<-,bundle,character-method}
\alias{+,partitionBundle,ANY-method}
\alias{+,partitionBundle,partition-method}
\alias{+,partitionBundle,partitionBundle-method}
\alias{+,partitionBundle-method}
\alias{[,partitionBundle,ANY,ANY,ANY-method}
\alias{[,partitionBundle-method}
\alias{[[,partitionBundle-method}
\alias{as.DocumentTermMatrix}
\alias{as.matrix,partitionBundle-method}
\alias{as.partitionBundle,list-method}
\alias{as.sparseMatrix}
\alias{as.sparseMatrix,TermDocumentMatrix-method}
\alias{as.sparseMatrix,partitionBundle-Method}
\alias{as.sparseMatrix,partitionBundle-method}
\alias{barplot,partitionBundle-method}
\alias{html,partitionBundle-method}
\alias{merge,partitionBundle-method}
\alias{names,partitionBundle-method}
\alias{names<-,bundle,character-method}
\alias{partitionBundle-class}
\alias{sAttributes,partitionBundle-method}
\alias{show,partitionBundle-method}
\alias{summary,partitionBundle-method}
\title{partitionBundle class}
\usage{
\S4method{names}{bundle,character}(x) <- value

\S4method{html}{partitionBundle}(object, filename = c(), type = "debate")

\S4method{show}{partitionBundle}(object)

\S4method{summary}{partitionBundle}(object)

\S4method{[}{partitionBundle,ANY,ANY,ANY}(x, i)

\S4method{barplot}{partitionBundle}(height, ...)

\S4method{sAttributes}{partitionBundle}(object, sAttribute)
}
\arguments{
\item{x}{a partitionBundle object}

\item{value}{the value}

\item{object}{a partitionBundle object}

\item{filename}{filename}

\item{type}{type}

\item{i}{integer index}

\item{sAttribute}{the s-attribute to use}

\item{pAttribute}{the p-attribute}

\item{pBundle}{a partitionBundle object}
}
\description{
A bundle of partition objects.

barplot of a partitionBundle
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"} the partitions making up the bundle}

\item{\code{corpus}}{Object of class \code{"character"} the CWB corpus the partition is based on}

\item{\code{sAttributesFixed}}{Object of class \code{"list"} fixed sAttributes}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{explanation}}{Object of class \code{"character"} an explanation of the partition}

\item{\code{xml}}{Object of class \code{"character"} whether the xml is flat or nested}

\item{\code{call}}{Object of class \code{"character"} the call that generated the partitionBundle}
}}
\author{
Andreas Blaette
}

