% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/partitionCluster-class.R, R/partitionCluster-methods.R, R/sAttributes-method.R
\docType{class}
\name{partitionCluster-class}
\alias{+,partitionCluster,ANY-method}
\alias{+,partitionCluster,partition-method}
\alias{+,partitionCluster,partitionCluster-method}
\alias{+,partitionCluster-method}
\alias{[,partitionCluster,ANY,ANY,ANY-method}
\alias{[,partitionCluster-method}
\alias{[[,partitionCluster,ANY-method}
\alias{[[,partitionCluster-method}
\alias{as.DocumentTermMatrix}
\alias{as.matrix,partitionCluster-method}
\alias{as.partitionCluster,list-method}
\alias{as.sparseMatrix}
\alias{as.sparseMatrix,TermDocumentMatrix-method}
\alias{as.sparseMatrix,partitionCluster-Method}
\alias{as.sparseMatrix,partitionCluster-method}
\alias{barplot,partitionCluster-method}
\alias{length,partitionCluster-method}
\alias{merge,partitionCluster-method}
\alias{names,partitionCluster-method}
\alias{partitionCluster-class}
\alias{sAttributes,partitionCluster-method}
\alias{show,partitionCluster-method}
\alias{summary,partitionCluster-method}
\title{partitionCluster class}
\usage{
\S4method{[[}{partitionCluster,ANY}(x, i)

\S4method{length}{partitionCluster}(x)

\S4method{sAttributes}{partitionCluster}(object, sAttribute)
}
\arguments{
\item{x}{a partitionCluster object}

\item{i}{integer index}

\item{object}{a partitionCluster object}

\item{sAttribute}{the s-attribute to use}
}
\description{
A cluster of partition objects.
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"} the partitions making up the cluster}

\item{\code{corpus}}{Object of class \code{"character"} the CWB corpus the partition is based on}

\item{\code{sAttributesFixed}}{Object of class \code{"list"} fixed sAttributes}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{explanation}}{Object of class \code{"character"} an explanation of the partition}

\item{\code{xml}}{Object of class \code{"character"} whether the xml is flat or nested}

\item{\code{call}}{Object of class \code{"character"} the call that generated the partitionCluster}
}}
\section{Methods}{
 \describe{ \item{show}{\code{signature(object =
  "partitionCluster")}: Display essential information }
  \item{addPos}{\code{signature(object="partitionCluster")}: add list with
  most frequent pos for a token }
  \item{tf}{\code{signature(object="partitionCluster")}: get term frequencies
  } \item{trim}{\code{signature(object="partitionCluster")}: trim a
  partitionCluster object } \item{[}{get frequency of a query} \item{[[}{get
  a partition within the cluster} \item{+}{\code{signature(object =
  "partitionCluster")}: combine two partitionClusters into a new one }
  \item{html}{bla}
  \item{as.sparseMatrix}{bla}
  }
}
\author{
Andreas Blaette
}

