% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/textstat_class.R
\docType{methods}
\name{pAttribute}
\alias{pAttribute}
\alias{pAttribute,textstat-method}
\title{get pAttribute}
\usage{
pAttribute(object)

\S4method{pAttribute}{textstat}(object)
}
\arguments{
\item{object}{a textstat object}

\item{pAttribute}{the pAttribute to get}
}
\description{
get pAttribute
}
