% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pAttributes-method.R
\docType{methods}
\name{pAttributes}
\alias{pAttributes}
\alias{pAttributes,character-method}
\title{get availables pAttributes}
\usage{
pAttributes(.Object)

\S4method{pAttributes}{character}(.Object)
}
\arguments{
\item{.Object}{a character string or partition object}
}
\description{
get availables pAttributes
}

