% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\docType{methods}
\name{pmi}
\alias{pmi}
\alias{pmi,context-method}
\title{Calculate Pointwise Mutual Information.}
\usage{
pmi(.Object)

\S4method{pmi}{context}(.Object)
}
\arguments{
\item{.Object}{An object.}
}
\description{
Calculate Pointwise Mutual Information.
}
\seealso{
See \code{\link{ll}}, \code{\link{chisquare}} and \code{\link{t_test}}.
}
