% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/browse_method.R, R/generics.R, R/polmineR_package.R
\name{browse<-}
\alias{browse<-}
\alias{name}
\alias{name<-}
\alias{polmineR-generics}
\title{generic methods defined in the polmineR-package}
\usage{
browse(x) <- value

name(x)

name(x) <- value
}
\description{
The package defines a set of generic functions. This doc file
only provides an overview. Please consult the documentation
of the classes to learn which methods can be applied to a class
of a certain type.
}

