% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/ranges.R
\docType{class}
\name{ranges-class}
\alias{ranges-class}
\alias{as.data.table.ranges}
\title{Ranges of query matches.}
\usage{
\method{as.data.table}{ranges}(x, ...)
}
\arguments{
\item{x}{A \code{ranges} class object.}

\item{...}{Additional arguments (unused).}
}
\description{
S4 class to manage ranges of corpus positions for query matches. The class
inherits from the classes \code{regions} and \code{corpus}.
}
\section{Slots}{

\describe{
\item{\code{query}}{A length-one \code{character} string, query used for query
matches.}
}}

\seealso{
Other classes to manage corpora: 
\code{\link{corpus-class}},
\code{\link{phrases-class}},
\code{\link{regions}},
\code{\link{subcorpus}}
}
\concept{classes to manage corpora}
