% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_method.R
\docType{methods}
\name{read,count-method}
\alias{read,count-method}
\title{rdname read-method}
\usage{
\S4method{read}{count}(.Object, col, partitionBundle, cqp = FALSE,
  highlight = list(), cpos = FALSE, ...)
}
\description{
rdname read-method
}

