% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_method.R
\docType{methods}
\name{read,data.table,ANY-method}
\alias{read,data.table,ANY-method}
\title{rdname read-method}
\usage{
\S4method{read}{data.table,ANY}(.Object, highlight = list(), cqp = FALSE,
  partitionBundle, col, ...)
}
\description{
rdname read-method
}

