% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_method.R
\docType{methods}
\name{read,data.table-method}
\alias{read,data.table-method}
\title{rdname read-method}
\usage{
\S4method{read}{data.table}(.Object, partitionBundle, col, ...)
}
\description{
rdname read-method
}

