% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regions_class.R
\docType{class}
\name{Regions-class}
\alias{Regions-class}
\title{Regions of a CWB corpus.}
\description{
Regions of a CWB corpus.
}
\section{Slots}{

\describe{
\item{\code{cpos}}{a data.table that will include a "cpos_left" and "cpos_right" column}

\item{\code{corpus}}{the CWB corpus (character vector length 1)}

\item{\code{encoding}}{the encoding of the CWB corpus (character vector length 1)}
}}

