% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{corpus_get_name}
\alias{corpus_get_name}
\alias{corpus_get_id}
\alias{corpus_get_home}
\alias{corpus_get_info}
\alias{corpus_get_encoding}
\alias{corpus_get_p_attributes}
\alias{corpus_get_s_attributes}
\alias{corpus_get_properties}
\title{Evaluate registry file.}
\usage{
corpus_get_name(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))

corpus_get_id(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))

corpus_get_home(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))

corpus_get_info(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))

corpus_get_encoding(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))

corpus_get_p_attributes(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))

corpus_get_s_attributes(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))

corpus_get_properties(corpus, registry = Sys.getenv("CORPUS_REGISTRY"))
}
\arguments{
\item{corpus}{name of the CWB corpus}

\item{registry}{directory of the registry (defaults to CORPUS_Registry environment variable)}
}
\description{
Functions to extract information from a registry file describing a corpus.
Several operations could be accomplished with the 'cwb-regedit' tool,
the functions defined here ensure that manipulating the registry is 
possible without a full installation of the CWB.
}
\details{
An appendix to the 'Corpus Encoding Tutorial' (http://cwb.sourceforge.net/files/CWB_Encoding_Tutorial.pdf)
includes an explanation of the registry file format.
}
