% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed.R
\name{renamed}
\alias{renamed}
\alias{sAttributes}
\alias{pAttributes}
\alias{getTokenStream}
\alias{getTerms}
\alias{getEncoding}
\title{Renamed Functions}
\usage{
sAttributes(...)

pAttributes(...)

getTokenStream(...)

getTerms(...)

getEncoding(...)
}
\arguments{
\item{...}{argument that are passed to the renamed function}
}
\description{
These functions have been renamed in order to have a consistent coding style
that follows the snake_case convention. The "old" function still work to
maintain backwards compatiblity.
}
