% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplot_method.R
\docType{methods}
\name{scatterplot}
\alias{scatterplot}
\alias{scatterplot,data.frame-method}
\alias{scatterplot-method}
\title{word scatterplot}
\usage{
\S4method{scatterplot}{data.frame}(object, xmax = c(), ymax = c(),
  fontSize = 0.7, rotation = 45, type = 0)
}
\arguments{
\item{object}{a data frame with the ranks of cooccurrences}

\item{xmax}{maximum on x axis for plot}

\item{ymax}{maximum on y axis for plot}

\item{fontSize}{the expansion factor for the words}

\item{rotation}{rotation of the text}

\item{type}{defaults to 0 for output of terms, can be set to symbol integer}
}
\value{
the plot
}
\description{
plot a word scatterplot
}
\author{
Andreas Blaette
}

