\name{scatterplotCollocates}
\alias{scatterplotCollocates}
\title{word scatterplot}
\usage{
scatterplotCollocates(comp, xmax = c(), ymax = c(), fontSize = 0.7,
  rotation = 45, type = 0)
}
\arguments{
  \item{comp}{a data frame with the ranks of collocates}

  \item{xmax}{maximum on x axis for plot}

  \item{ymax}{maximum on y axis for plot}

  \item{fontSize}{the expansion factor for the words}

  \item{rotation}{rotation of the text}

  \item{type}{defaults to 0 for output of terms, can be set
  to symbol integer}
}
\value{
the plot
}
\description{
plot a word scatterplot
}
\author{
Andreas Blaette
}

