% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{session-class}
\alias{createProject}
\alias{project,missing-method}
\alias{project,session-method}
\alias{project<-,session,character-method}
\alias{session}
\alias{session-class}
\title{session class}
\usage{
\S4method{project}{session,character}(x) <- value

\S4method{project}{session}(object)

\S4method{project}{missing}()

createProject(project)
}
\arguments{
\item{x}{bla}

\item{value}{to assign}

\item{object}{bla}

\item{project}{bla}
}
\description{
default settings for current session
}
\section{Slots}{

\describe{
\item{\code{project}}{currently active project}

\item{\code{defaultCorpus}}{corpus to use if none is provided}

\item{\code{pAttribute}}{default pAttribute}

\item{\code{leftContext}}{default left context}

\item{\code{rightContext}}{default right context}

\item{\code{minSignificance}}{default minimum significance}

\item{\code{minFrequency}}{default minimum frequency}

\item{\code{posFilter}}{default filter for POS}

\item{\code{filterType}}{default filter type}

\item{\code{kwicMetadata}}{default metadata displayed for kwic}

\item{\code{kwicNo}}{default number of kwic lines}

\item{\code{metadata}}{default metadata}

\item{\code{multicore}}{default multicore}

\item{\code{consoleEncoding}}{default console encoding}

\item{\code{smtpServer}}{default server to use for sending mail}

\item{\code{smtpPort}}{default port}

\item{\code{email}}{default email address}

\item{\code{partitionDir}}{default partition directory}

\item{\code{defaultKwicCollocate}}{default collocate for kwic display}

\item{\code{defaultKwicNode}}{default node for kwic analysis}
}}

