% Generated by roxygen2 (4.0.1): do not edit by hand
\name{shinyDispersion}
\alias{shinyDispersion}
\title{Shiny app for exploring the distribution}
\usage{
shinyDispersion(type)
}
\arguments{
\item{type}{either "simple", "multi" or "crosstab"}
}
\value{
not applicable
}
\description{
Fires shiny apps for analysing the distribution of a query
across subcorpora. This is a wrapper!
}
\details{
Shiny will open a window in your browser.
}
\author{
Andreas Blaette
}

