\name{shinyKeyness}
\alias{shinyKeyness}
\title{Shiny app for keyness statistics}
\usage{
shinyKeyness(partition1, partition2)
}
\arguments{
  \item{partition1}{a partition object (corpus of
  interest)}

  \item{partition2}{a partition object (reference corpus)}
}
\value{
not applicable
}
\description{
Fires up a shiny app. The default values of
\code{drillingControls} are used.
}
\details{
Shiny will open a window in your browser. To other
partitions, the function has to be called again.
}
\author{
Andreas Blaette
}

