\name{shinyVolatility}
\alias{shinyVolatility}
\title{Shiny app for volatility analysis}
\usage{
shinyVolatility()
}
\description{
Fires up a shiny app. The default values of
\code{drillingControls} are used.
}
\details{
Shiny will open a window in your browser. To other
partitions, the function has to be called again.
}
\author{
Andreas Blaette
}

