% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_method.R
\docType{methods}
\name{size}
\alias{size}
\alias{size,DocumentTermMatrix-method}
\alias{size,character-method}
\alias{size,partition-method}
\title{get corpus size}
\usage{
size(x)

\S4method{size}{character}(x)

\S4method{size}{partition}(x)

\S4method{size}{DocumentTermMatrix}(x)
}
\arguments{
\item{x}{object}
}
\description{
get corpus size
}

