% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{tag}
\alias{tag}
\alias{tag,character-method}
\title{run treetagger over a character vector}
\usage{
tag(object, ...)

\S4method{tag}{character}(object)
}
\arguments{
\item{object}{a character vector}

\item{...}{further parameters}
}
\description{
treetagger is used as a backend
}

