% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/tempcorpus.R
\docType{class}
\name{tempcorpus-class}
\alias{tempcorpus-class}
\alias{tempcorpus}
\title{S4 class to capture core information on a temporary CWB corpus}
\usage{
tempcorpus(.Object, ...)
}
\arguments{
\item{.Object}{a partition object}

\item{...}{further parameters}
}
\description{
S4 class to capture core information on a temporary CWB corpus

Based on the corpus positions defining a partition, a temporary CWB corpus is 
generated that is stored in a temporary directory.
}
\section{Slots}{

\describe{
\item{\code{cpos}}{matrix with start/end corpus positions}

\item{\code{dir}}{directory where the tempcorpus is stored}

\item{\code{registry}}{directory of the registry dir (subdirectory of dir)}

\item{\code{indexed}}{directory of the dir with the indexed files}
}}

