% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R
\docType{class}
\name{tempcorpus_class}
\alias{tempcorpus_class}
\alias{tempcorpus-class}
\title{S4 class to capture core information on a temporary CWB corpus}
\description{
S4 class to capture core information on a temporary CWB corpus
}
\section{Slots}{

\describe{
\item{\code{cpos}}{matrix with start/end corpus positions}

\item{\code{dir}}{directory where the tempcorpus is stored}

\item{\code{registry}}{directory of the registry dir (subdirectory of dir)}

\item{\code{indexed}}{directory of the dir with the indexed files}
}}

