% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates_method.R
\docType{methods}
\name{getTemplate}
\alias{getTemplate}
\alias{getTemplate,character-method}
\alias{getTemplate,missing-method}
\alias{getTemplate,partition-method}
\alias{setTemplate}
\alias{setTemplate,character-method}
\title{Get and set templates.}
\usage{
getTemplate(.Object, ...)

\S4method{getTemplate}{character}(.Object)

\S4method{getTemplate}{partition}(.Object)

\S4method{getTemplate}{missing}(.Object)

setTemplate(.Object, ...)

\S4method{setTemplate}{character}(.Object, template)
}
\arguments{
\item{...}{further parameters}

\item{template}{template to assign (a list)}

\item{x}{either a "character" vector, if missing, the names of available templates will be returned}
}
\description{
Templates are used to format the markdown/html output of partitions. Upon loading 
the polmineR package, templates for corpora are loaded into the option 'polmineR.templates'.
}

