% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TermDocumentMatrix_methods.R, R/noise_method.R, R/size_method.R
\docType{methods}
\name{TermDocumentMatrix}
\alias{TermDocumentMatrix}
\alias{noise,DocumentTermMatrix-method}
\alias{size,DocumentTermMatrix-method}
\alias{size,TermDocumentMatrix-method}
\title{Methods for TermDocumentMatrix / DocumentTermMatrix}
\usage{
\S4method{noise}{DocumentTermMatrix}(.Object, minTotal = 2,
  minTfIdfMean = 0.005, sparse = 0.995, stopwordsLanguage = "german",
  minNchar = 2, specialChars = "^[a-zA-ZéäöüÄÖÜ-ß|-]+$",
  numbers = "^[0-9\\\\.,]+$", verbose = TRUE)

\S4method{size}{DocumentTermMatrix}(x)

\S4method{size}{TermDocumentMatrix}(x)
}
\arguments{
\item{.Object}{an .Object of class \code{"DocumentTermMatrix"}}

\item{minTfIdfMean}{minimum mean value for tf-idf to qualify a term as non-noise}

\item{sparse}{will be passed into \code{"removeSparseTerms"} from \code{"tm"}-package}

\item{stopwordsLanguage}{e.g. "german", to get stopwords defined in the tm package}

\item{minNchar}{min char length ti qualify a term as non-noise}

\item{x}{object}

\item{minSum}{minimum colsum (for DocumentTermMatrix) to qualify a term as non-noise}
}
\value{
a list
}
\description{
Methods for TermDocumentMatrix / DocumentTermMatrix

detect noise
}

