% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textstat-class.R, R/textstat-methods.R, R/view-method.R
\docType{class}
\name{textstat-class}
\alias{as.DataTables,textstat-method}
\alias{as.data.frame,textstat-method}
\alias{colnames,textstat-method}
\alias{dim,textstat-method}
\alias{head,textstat-method}
\alias{names,textstat-method}
\alias{rownames,textstat-method}
\alias{show,textstat-method}
\alias{tail,textstat-method}
\alias{textstat-class}
\alias{trim,textstat-method}
\alias{view,textstat-method}
\title{S4 textstat class}
\usage{
\S4method{head}{textstat}(x, ...)

\S4method{tail}{textstat}(x, ...)

\S4method{view}{textstat}(.Object)
}
\arguments{
\item{x}{an object}

\item{...}{further parameters}

\item{.Object}{an object}
}
\description{
superclass for keyness and conctext class
}
\section{Slots}{
 \describe{ \item{\code{pAttribute}:}{Object of class
  \code{"character"} p-attribute of the query } \item{\code{corpus}:}{Object
  of class \code{"character"} the CWB corpus used }
  \item{\code{stat}:}{Object of class \code{"data.frame"} statistics of the
  analysis } \item{\code{encoding}:}{Object of class \code{"character"}
  encoding of the corpus } \item{\code{cutoff}:}{Object of class
  \code{"list"} cutoff levels that have been applied } }
}

