% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_method.R, R/pmi_method.R
\name{textstatistics}
\alias{textstatistics}
\title{Log-likelihood-test}
\usage{
ll(.Object, ...)

pmi(.Object)
}
\arguments{
\item{ids}{the numeric ids (integer) of the tokens}

\item{windowFreq}{term frequency of the tokens in corpus A}

\item{corpusFreq}{term frequency of the tokens in the corpus B}

\item{windows.total}{total size of the window}

\item{corpus.total}{total size of the corpus}

\item{windowIds}{ids of cooccurring words, used for matching}
}
\value{
a Matrix with six columns
}
\description{
Log-likelihood-test

There may be a (small) problem: the counts used are not cooccurrences ...
}

