% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tf-method.R
\docType{methods}
\name{tf}
\alias{tf}
\alias{tf,character-method}
\alias{tf,partition-method}
\alias{tf,partitionCluster-method}
\alias{tf-method}
\title{get term frequencies}
\usage{
tf(object, ...)

\S4method{tf}{partition}(object, token, method = "in", pAttribute = NULL)

\S4method{tf}{partitionCluster}(object, token, method = "in",
  pAttribute = NULL, rel = FALSE)

\S4method{tf}{character}(object, token, pAttribute, method = "in")
}
\arguments{
\item{object}{either a partition or a partitionCluster object}

\item{...}{further parameters}

\item{token}{a character vector (one or multiple terms to be looked up)}

\item{method}{either "in", "grep" or "cqp" (defaults to "in")}

\item{pAttribute}{if NULL, the pAttributes available in the partition object will be reported}

\item{rel}{logical, defaults to FALSE}
}
\description{
Method to obtain term frequencies for one or multiple terms or queries.
}
\examples{
# generate a partition for testing
test <- partition("PLPRBTTXT", list(text_date=".*"), tf=c("word", "lemma"), method="grep")
tf(test, "Wir") # get frequencies for one token
tf(test, c("Wir", "lassen", "uns")) # get frequencies for multiple tokens
tf(test, c("Zuwander.*", "Integration.*"), method="grep") # get frequencies using "grep"-method
tf("PLPRTXT", c("machen", "Integration"), "word")
}
\seealso{
tf
}

