% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TokenStream_class.R
\docType{class}
\name{TokenStream-class}
\alias{TokenStream-class}
\title{Class for token stream operations.}
\description{
Class for token stream operations.
}
\section{Slots}{

\describe{
\item{\code{tokenStream}}{a data.table that will include a column "token"}

\item{\code{corpus}}{the CWB corpus (character vector length 1)}

\item{\code{encoding}}{the encoding of the CWB corpus (character vector length 1)}
}}

