% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_method.R
\docType{methods}
\name{trim,DocumentTermMatrix-method}
\alias{trim,DocumentTermMatrix-method}
\title{shrink a matrix for topicmodelling}
\usage{
\S4method{trim}{DocumentTermMatrix}(object, minTotal = 2,
  minTfIdfMean = 0.005, sparse = 0.995, stopwordsLanguage = "german",
  removeTrash = TRUE, verbose = TRUE)
}
\arguments{
\item{object}{an object of class \code{"DocumentTermMatrix"}}

\item{minTfIdfMean}{minimum mean value for tf-idf (suggested: 0.005)}

\item{sparse}{will be passed into \code{"removeSparseTerms"} from \code{"tm"}-package}

\item{stopwordsLanguage}{e.g. "german", to get stopwords defined in the tm package}

\item{removeTrash}{if TRUE, terms containing special chars, numbers or length = 1 will be removed}

\item{minCount}{minimum overall occurrence of a term in the matrix}
}
\description{
shrink a matrix for topicmodelling
}

