% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{trim,partitionCluster-method}
\alias{trim,partitionCluster-method}
\title{trim partitionCluster Object}
\usage{
\S4method{trim}{partitionCluster}(object, pAttribute = NULL,
  minFrequency = 0, posFilter = NULL, tokenFilter = NULL, drop = NULL,
  minSize = 0, keep = NULL, mc = NULL, ...)
}
\arguments{
\item{object}{a partitionCluster object}

\item{pAttribute}{character vector: "word", "lemma", or both}

\item{minFrequency}{minimum frequency of tokens}

\item{posFilter}{pos to keep}

\item{tokenFilter}{a character vector with tokens to keep in the tf list specified by pAttribute}

\item{drop}{partitionObjects you want to drop, specified either by number or by label}

\item{minSize}{a minimum size for the partitions to be kept}

\item{keep}{specify labels of partitions to keep, everything else is dropped}

\item{mc}{if not NULL logical - whether to use multicore parallelization}

\item{...}{further arguments (unused)}
}
\value{
partitionCluster
}
\description{
Apply the trim method to all partitions in a partitionCluster. In addition,
partitions can be dropped.
}

