% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/use-method.R
\docType{methods}
\name{use}
\alias{use}
\alias{use,character-method}
\title{use corpus}
\usage{
use(pkg)

\S4method{use}{character}(pkg)
}
\arguments{
\item{pkg}{the object to be used}
}
\description{
Use a cwb corpus shipped in a package.
}

