% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weigh-method.R
\docType{methods}
\name{weigh}
\alias{weigh}
\alias{weigh,DocumentTermMatrix-method}
\alias{weigh,TermDocumentMatrix-method}
\title{weigh a matrix}
\usage{
weigh(.Object, ...)

\S4method{weigh}{TermDocumentMatrix}(.Object, method = "tfidf")

\S4method{weigh}{DocumentTermMatrix}(.Object, method = "tfidf")
}
\arguments{
\item{.Object}{the matrix to be weighed}

\item{method}{the kind of weight to apply}
}
\description{
weigh a matrix
}

