% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/weigh-method.R
\docType{methods}
\name{weigh,matrix-method}
\alias{weigh}
\alias{weigh,TermDocumentMatrix-method}
\alias{weigh,matrix-method}
\title{weigh a matrix}
\usage{
\S4method{weigh}{matrix}(tdm, method = "tfidf", corpusSizes)
}
\arguments{
\item{tdm}{a matrix (which is a term-document-matrix, terms in the rows, docs in the columns)}

\item{method}{which weighting to apply (currently only tfidf is implemented)}

\item{corpusSizes}{the total number of tokens in the documents}
}
\description{
weigh a matrix
}

