% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addPos-method.R
\docType{methods}
\name{addPos}
\alias{addPos}
\alias{addPos,keyness-method}
\alias{addPos,partition-method}
\alias{addPos-method}
\title{add pos information}
\usage{
addPos(object, ...)

\S4method{addPos}{partition}(object, pAttribute)

\S4method{addPos}{keyness}(object, Partition = NULL)
}
\arguments{
\item{object}{either a partition, a partitionBundle or a keyness object}

\item{...}{further arguments}

\item{pAttribute}{character vector (typically c("word") or c("lemma") or c("word", "lemma"))}

\item{Partition}{a partition object (the corpus of interest)}
}
\value{
the original, enhanced object

an augmented partition or partitionBundle object (includes pos now)

an enhanced keyness object
}
\description{
Add information on part-of-speech tags of tokens to tokens. The method is
available for objects of the classes 'partition', 'partitionBundle' and
'keyness' respectively.
\code{method?addPos("partitionBundle")}, \code{method?addPos("keyness")}).

The 'pos'-slot of the partition (or partitionBundle) object is filled with tables
providing a statistic on the frequency of a pos-tag of a token
in the partition.

A keyness object will be supplemented with pos information. The working of the
method is potentially slow. It is recommended to trim the object first, before
adding pos information.
}
\author{
Andreas Blaette
}

