% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/adjustEncoding-method.R
\docType{methods}
\name{adjustEncoding}
\alias{adjustEncoding}
\alias{adjustEncoding,character-method}
\title{adjust encoding}
\usage{
adjustEncoding(.Object, ...)

\S4method{adjustEncoding}{character}(.Object, corpusEncoding)
}
\arguments{
\item{.Object}{the object (usually a character vector, nothing else is implemented so far)}

\item{...}{further parameters}

\item{corpusEncoding}{the encoding of the corpus (e.g. "latin1")}
}
\description{
Utility method to help making sure that the encoding of a character string or the relevant parts of an object
conform to the encoding to the corpus.
}

