% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences.R, R/as.sparseMatrix.R
\docType{class}
\name{Cooccurrences-class}
\alias{Cooccurrences-class}
\alias{as.simple_triplet_matrix,Cooccurrences-method}
\alias{ll,Cooccurrences-method}
\alias{as_igraph}
\alias{as.sparseMatrix,Cooccurrences-method}
\title{Cooccurrences class for corpus/partition.}
\usage{
\S4method{as.simple_triplet_matrix}{Cooccurrences}(x)

\S4method{ll}{Cooccurrences}(.Object, verbose = TRUE)

as_igraph(x, edge_attributes = c("ll", "ab_count", "rank_ll"),
  vertex_attributes = "count", as.undirected = TRUE)

\S4method{as.sparseMatrix}{Cooccurrences}(x, col)
}
\arguments{
\item{x}{A \code{Cooccurrences} class object.}

\item{.Object}{A \code{Cooccurrences}-class object.}

\item{verbose}{Logical.}

\item{edge_attributes}{Attributes from stat \code{data.table} in x to add to edges.}

\item{vertex_attributes}{Vertex attributes to add to nodes.}

\item{as.undirected}{Logical, whether to return directed or undirected graph.}

\item{col}{A column to extract.}
}
\description{
The class inherits from the \code{textstat}-class.

Transform cooccurrence igraph object.
}
\details{
Return a sparse matrix with the term
  cooccurrence matrix.

Returns a simple triplet
  matrix based on the counts of term cooccurrences. If counts are not yet
  present, that is done first.
}
\section{Slots}{

\describe{
\item{\code{left}}{Single \code{integer} value, number of tokens to the left.}

\item{\code{right}}{Single \code{integer} value, number of tokens to the right.}

\item{\code{p_attribute}}{A \code{character} vector, the p-attribute the evaluation of the corpus is based on.}

\item{\code{corpus}}{Length-one \code{character} vector, the CWB corpus used.}

\item{\code{stat}}{A \code{data.table} with the statistical analysis of cooccurrences.}

\item{\code{encoding}}{Length-one \code{character} vector, the encoding of the corpus.}

\item{\code{method}}{A \code{character} vector stating the statistical test(s) used.}

\item{\code{keep}}{A \code{list} of named character vectors, names are p-attributes.}

\item{\code{drop}}{A \code{list} of named character vectors, names are p-attributes.}

\item{\code{partition}}{A \code{partition}.}

\item{\code{window_sizes}}{A \code{data.table} linking the number of tokens in the
context of a token identified by id.}

\item{\code{minimized}}{Logical, whether the object has been minimized.}
}}

