% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences.R
\docType{methods}
\name{Cooccurrences}
\alias{Cooccurrences}
\alias{Cooccurrences,character-method}
\alias{Cooccurrences,partition-method}
\title{Get all cooccurrences.}
\usage{
Cooccurrences(.Object, ...)

\S4method{Cooccurrences}{character}(.Object, ...)

\S4method{Cooccurrences}{partition}(.Object, p_attribute, left, right,
  stoplist = NULL, mc = getOption("polmineR.mc"), verbose = FALSE,
  progress = FALSE, ...)
}
\arguments{
\item{.Object}{A length-one character vector indicating a corpus, or a
\code{partition} object.}

\item{...}{Further arguments.}

\item{p_attribute}{The positional attribute.}

\item{left}{Size of left context.}

\item{right}{Size of right context.}

\item{stoplist}{Tokens to drop.}

\item{mc}{Logical.}

\item{verbose}{Logical.}

\item{progress}{Logical.}
}
\description{
Get all cooccurrences.
}
\examples{
\dontrun{
stopwords <- unname(unlist(noise(terms("REUTERS", p_attribute = "word"), stopwordsLanguage = "en")))
r <- Cooccurrences(.Object = "REUTERS", p_attribute = "word", left = 5L, right = 5L, stoplist = stopwords)
r <- ll(r)
r <- subset(r, ll > 11.83 & ab_count >= 5)
data.table::setorderv(r@stat, cols = "ll", order = -1L)
head(r, 25)

if (requireNamespace("igraph", quietly = TRUE)){
  g <- as_igraph(r, as.undirected = TRUE)
  plot(g)
}

a <- cooccurrences(r, query = "oil")
a <- data.table::as.data.table(a)

b <- Cooccurrences("REUTERS", query = "oil")
b <- data.table::as.data.table(b)
b <- b[!word \%in\% stopwords]

# now let's check whether results are identical
all(b[["word"]][1:5] == a[["word"]][1:5])
}
}
