% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/as.mallet-method.R
\name{as.mallet}
\alias{as.mallet}
\title{as.mallet}
\usage{
as.mallet(.Object, ...)
}
\arguments{
\item{.Object}{partition- or partitionCluster-object}

\item{...}{further parameters}
}
\description{
Save polmineR-object as mallet object
}

