% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.markdown-method.R, R/html-method.R
\docType{methods}
\name{as.markdown}
\alias{as.markdown}
\alias{html,character-method}
\title{turn partition into markdown}
\usage{
as.markdown(object, ...)

\S4method{html}{character}(object)
}
\arguments{
\item{object}{object to be converted}

\item{...}{further arguments}
}
\description{
for further processing to html
}

