% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.markdown_method.R
\docType{methods}
\name{as.markdown}
\alias{as.markdown}
\alias{as.markdown,numeric-method}
\alias{as.markdown,plprPartition-method}
\title{Turn partition into markdown}
\usage{
as.markdown(.Object, ...)

\S4method{as.markdown}{numeric}(.Object, corpus, meta, template)

\S4method{as.markdown}{plprPartition}(.Object, meta, cpos = FALSE,
  interjections = TRUE)
}
\arguments{
\item{.Object}{object to be converted}

\item{...}{further arguments}

\item{meta}{metainformation i.e. s-attributes) to be displayed}

\item{cpos}{logical, whether to add cpos as ids in span elements}

\item{interjections}{logical, whether to format interjections}
}
\description{
The method is the worker behind the html-method.
}

