% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.markdown_method.R
\docType{methods}
\name{as.markdown}
\alias{as.markdown}
\alias{as.markdown,numeric-method}
\alias{as.markdown,plprPartition-method}
\title{Generate markdown from a partition.}
\usage{
as.markdown(.Object, ...)

\S4method{as.markdown}{numeric}(.Object, corpus, meta = NULL, cpos = FALSE,
  cutoff = NULL, ...)

\S4method{as.markdown}{plprPartition}(.Object, meta = NULL,
  template = getTemplate(.Object), cpos = FALSE, interjections = TRUE,
  cutoff = NULL, ...)
}
\arguments{
\item{.Object}{object to be converted}

\item{...}{further arguments}

\item{corpus}{name of CWB corpus}

\item{meta}{metainformation i.e. s-attributes) to be displayed}

\item{cpos}{logical, whether to add cpos as ids in span elements}

\item{cutoff}{maximum number of tokens to reconstruct}

\item{template}{a template for formatting output}

\item{interjections}{logical, whether to format interjections}
}
\description{
The method is the worker behind the html-method.
}
\examples{
use("polmineR")
P <- partition("REUTERS", places = "argentina")
as.markdown(P)
as.markdown(P, meta = c("id", "places"))
if (interactive()) read(P, meta = c("id", "places"))
}
