% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.markdown_method.R
\docType{methods}
\name{as.markdown}
\alias{as.markdown}
\alias{as.markdown,plprPartition-method}
\alias{as.markdown,pressPartition-method}
\title{turn partition into markdown}
\usage{
as.markdown(object, ...)

\S4method{as.markdown}{plprPartition}(object, meta, cpos = FALSE,
  interjections = TRUE)

\S4method{as.markdown}{pressPartition}(object, meta = c("text_newspaper",
  "text_date"), cpos = FALSE)
}
\arguments{
\item{object}{object to be converted}

\item{...}{further arguments}

\item{meta}{metainformation i.e. s-attributes) to be displayed}

\item{cpos}{logical, whether to add cpos as ids in span elements}

\item{interjections}{logical, whether to format interjections}
}
\description{
for further processing to html
}

