% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.speeches.R
\name{as.speeches}
\alias{as.speeches}
\title{Split Corpus or Partition Into Speeches}
\usage{
as.speeches(.Object, s_attribute_date = grep("date", s_attributes(.Object),
  value = TRUE), s_attribute_name = grep("name", s_attributes(.Object), value
  = TRUE), gap = 500, mc = FALSE, verbose = TRUE, progress = TRUE)
}
\arguments{
\item{.Object}{a partition .Object}

\item{s_attribute_date}{the s-attribute that provides the dates of sessions}

\item{s_attribute_name}{the s-attribute that provides the names of speakers}

\item{gap}{number of tokens between strucs assumed to make the difference
whether a speech has been interrupted (by an interjection or question), or
whether to assume seperate speeches}

\item{mc}{whether to use multicore, defaults to FALSE}

\item{verbose}{logical, defaults to TRUE}

\item{progress}{logical}
}
\value{
A \code{partition_bundle}, the names of the objects in the bundle are
  the speaker name, the date of the speech and an index for the number of the
  speech on a given day, concatenated by underscores.
}
\description{
Split entire corpus or a partition into speeches. The heuristic is to split
the corpus/partition into partitions on day-to-day basis first, using the
s-attribute provided by \code{s_attribute_date}. These subcorpora are then
splitted into speeches by speaker name, using s-attribute
\code{s_attribute_name}. If there is a gap larger than the number of tokens
supplied by argument \code{gap}, contributions of a speaker are assumed to be
two seperate speeches.
}
\examples{
use("polmineR")
speeches <- as.speeches(
  "GERMAPARLMINI",
  s_attribute_date = "date", s_attribute_name = "speaker"
)
speeches_count <- count(speeches, p_attribute = "word")
tdm <- as.TermDocumentMatrix(speeches_count, col = "count")

bt <- partition("GERMAPARLMINI", date = "2009-10-27")
speeches <- as.speeches(bt, s_attribute_name = "speaker")
summary(speeches)
}
