% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.speeches_method.R
\docType{methods}
\name{as.speeches}
\alias{as.speeches}
\alias{as.speeches,partition-method}
\alias{speeches}
\alias{speeches,partition-method}
\alias{speeches-method}
\title{Split partition into speeches}
\usage{
as.speeches(.Object, ...)

\S4method{as.speeches}{partition}(.Object, sAttributeDates, sAttributeNames,
  gap = 500, mc = FALSE, verbose = FALSE)
}
\arguments{
\item{.Object}{a partition .Object}

\item{gap}{number of tokens between strucs to identify speeches}

\item{mc}{whether to use multicore, defaults to FALSE}

\item{verbose}{logical, defaults to TRUE}

\item{sAttribute}{the s-attribute that provides the speaker names}

\item{addMeta}{whether to add metainformation to partition name, if wanted, provide s-attribute with date as character vector}
}
\value{
a partitionCluster object
}
\description{
A method designed for corpora from the
PolMine corpora of plenary protocols. A partition is split into speeches.
}

