% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.VCorpus.R
\docType{methods}
\name{as.VCorpus,partition_bundle-method}
\alias{as.VCorpus,partition_bundle-method}
\title{Coerce partition_bundle to VCorpus.}
\usage{
\S4method{as.VCorpus}{partition_bundle}(x)
}
\arguments{
\item{x}{a \code{partition_bundle} object}
}
\description{
Coerce partition_bundle to VCorpus.
}
\examples{
use("polmineR")
P <- partition("GERMAPARLMINI", date = "2009-11-10")
PB <- partition_bundle(P, s_attribute = "speaker")
VC <- as.VCorpus(PB)
}
