% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{as.data.frame}
\alias{as.data.frame}
\alias{as.data.frame,context-method}
\alias{as.data.frame,keyness-method}
\title{Get statistics table from an object}
\usage{
\S4method{as.data.frame}{keyness}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{object with a statistics table}

\item{...}{any further arguments}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link{make.names}}) is optional.}
}
\description{
Get statistics table from an object
}

