% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/browse-method.R, R/generics.R, R/kwic-methods.R
\docType{methods}
\name{browse,textstat-method}
\alias{browse}
\alias{browse,context-method}
\alias{browse,html-method}
\alias{browse,kwic-method}
\alias{browse,partition-method}
\alias{browse,textstat-method}
\title{display in browser}
\usage{
\S4method{browse}{textstat}(object)

\S4method{browse}{context}(object)

\S4method{browse}{partition}(object, meta = NULL)

\S4method{browse}{html}(object)

browse(object, ...)

\S4method{browse}{kwic}(object, colnames = NULL)
}
\arguments{
\item{object}{what is to be displayed}

\item{meta}{metainformation to be displayed}

\item{...}{further parameters}

\item{colnames}{colnames to be used for data.frame}
}
\description{
bla
}

