% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\docType{methods}
\name{browse}
\alias{browse}
\alias{browse,textstat-method}
\alias{browse,cooccurrences-method}
\alias{browse,partition-method}
\alias{browse,html-method}
\alias{browse,kwic-method}
\alias{browse,press_partition-method}
\title{Display in browser}
\usage{
browse(object, ...)

\S4method{browse}{textstat}(object)

\S4method{browse}{cooccurrences}(object)

\S4method{browse}{partition}(object, meta = NULL)

\S4method{browse}{html}(object)

\S4method{browse}{kwic}(object, colnames = NULL)

\S4method{browse}{press_partition}(object, meta = c("text_newspaper",
  "text_date"))
}
\arguments{
\item{object}{what is to be displayed}

\item{...}{further parameters}

\item{meta}{metainformation to be displayed}

\item{colnames}{colnames to be used for data.frame}
}
\description{
Display in browser
}
