% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bubblegraph}
\alias{bubblegraph}
\title{bubblegraph visualisation}
\usage{
bubblegraph(ctab, rex = 1, leftMargin = 1.2, bottomMargin = 1.6,
  cex = 1, font = 1)
}
\arguments{
\item{ctab}{a crosstabulation (a slot of a crosstab object, see example)}

\item{rex}{a radius expansion factor to control the bubble size}

\item{leftMargin}{adjust left margin of the plot}

\item{bottomMargin}{adjust bottom margin of the plot}

\item{cex}{character expansion factor}

\item{font}{set font parameter from par}
}
\value{
a nice plot, hopefully
}
\description{
visualise the distribution of counts/frequencies in a crosstab object with a bubblegraph
}
\examples{
\dontrun{
bt <- partition(sAttributes=list(text_date="*"), method="grep", corpus="PLPRBTTXT", encoding="latin1")
dist <- dispersion(bt, "Politik", c("text_date", "text_party"))
bubblegraph(dist@total$rel)
}
}

