% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/as.sparseMatrix_method.R, R/bundle_class.R, R/corpus_method.R
\docType{methods}
\name{as.sparseMatrix,bundle-method}
\alias{+,bundle,bundle-method}
\alias{+,bundle,textstat-method}
\alias{[[,bundle-method}
\alias{as.bundle,list-method}
\alias{as.bundle,textstat-method}
\alias{as.matrix,bundle-method}
\alias{as.sparseMatrix,bundle-method}
\alias{bundle-class}
\alias{corpus,bundle-method}
\alias{length,bundle-method}
\alias{names,bundle-method}
\alias{names<-,bundle,character-method}
\alias{unique,bundle-method}
\title{bundle class}
\usage{
\S4method{as.sparseMatrix}{bundle}(x, col)

\S4method{length}{bundle}(x)

\S4method{names}{bundle}(x)

\S4method{names}{bundle,character}(x) <- value

\S4method{unique}{bundle}(x)

\S4method{+}{bundle,bundle}(e1, e2)

\S4method{+}{bundle,textstat}(e1, e2)

\S4method{[[}{bundle}(x, i)

\S4method{as.matrix}{bundle}(x, col)

\S4method{as.bundle}{list}(object, ...)

\S4method{as.bundle}{textstat}(object)

\S4method{corpus}{bundle}(object)
}
\arguments{
\item{x}{a bundle object}

\item{col}{columsn of the table to be used}

\item{value}{...}

\item{e1}{object 1}

\item{e2}{object 2}

\item{i}{index to address an object}

\item{object}{a bundle object}
}
\description{
A class to bundle several objects (partition, context, comp, cooccurrences objects)
in one S4 object.
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"}}

\item{\code{pAttribute}}{Object of class \code{"character"}}
}}
\author{
Andreas Blaette
}

