% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bundle_class.R
\docType{class}
\name{bundle-class}
\alias{[[,bundle,ANY-method}
\alias{bapply}
\alias{bapply,bundle-method}
\alias{bundle-class}
\alias{length,bundle-method}
\alias{names,bundle-method}
\alias{unique,bundle-method}
\title{bundle class}
\usage{
\S4method{[[}{bundle,ANY}(x, i)

\S4method{length}{bundle}(x)

\S4method{names}{bundle}(x)

\S4method{unique}{bundle}(x)

bapply(x, ...)

\S4method{bapply}{bundle}(x, f, ...)
}
\arguments{
\item{x}{a bundle object}

\item{i}{index to address an object}

\item{...}{further parameters}

\item{f}{a function that can be applied to each object contained in the bundle}
}
\description{
A class to bundle several objects (partition, context, keyness, collocations objects)
in one S4 object.
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"}}
}}
\author{
Andreas Blaette
}

