% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisquare_method.R
\docType{methods}
\name{chisquare}
\alias{chisquare}
\alias{chisquare,context-method}
\alias{chisquare,textstat-method}
\title{perform chisquare-text}
\usage{
chisquare(.Object, ...)

\S4method{chisquare}{textstat}(.Object)

\S4method{chisquare}{context}(.Object)
}
\arguments{
\item{.Object}{object}

\item{...}{further parameters}
}
\value{
a table
}
\description{
Perform Chisquare-Test based on a table with counts
}
\details{
This function deliberately uses a self-made chi-square test for performance
reason
}
\author{
Andreas Blaette
}

