% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coerce_methods.R
\docType{methods}
\name{as.TermDocumentMatrix}
\alias{as.TermDocumentMatrix}
\alias{as.TermDocumentMatrix,bundle-method}
\alias{as.TermDocumentMatrix,character-method}
\title{coerce object to matrix}
\usage{
as.TermDocumentMatrix(x, ...)

\S4method{as.TermDocumentMatrix}{character}(x, pAttribute, sAttribute,
  from = NULL, to = NULL, strucs = NULL, rmBlank = TRUE,
  verbose = TRUE, robust = FALSE, mc = FALSE)

\S4method{as.TermDocumentMatrix}{bundle}(x, col, pAttribute = NULL,
  verbose = TRUE)
}
\arguments{
\item{x}{some object}

\item{...}{to make the check happy}

\item{pAttribute}{the p-attribute}

\item{sAttribute}{the s-attribute}

\item{from}{bla}

\item{to}{bla}

\item{strucs}{bla}

\item{rmBlank}{bla}

\item{verbose}{bla}

\item{robust}{bla}

\item{mc}{logical}
}
\value{
a TermDocumentMatrix
}
\description{
Coerce an object, usually a bundle (or an object inheriting from bundle) to
a TermDocumentMatrix, DocumentTermMatrix, sparseMatrix etc.
}
\examples{
\dontrun{
foo <- as.TermDocumentMatrix(
  x="ZEIT", pAttribute="word", sAttribute="text_id",
  from="1946_01_auf-einen-von-bomben-zerschlagenen-engel.html",
  to="1951_08_gespraeche-zum-interzonenhandel.html", robust="LESUNG", mc=TRUE, verbose=TRUE, rmBlank=TRUE
)
}
}
\author{
Andreas Blaette

me
}

