% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/collocations-method.R
\docType{methods}
\name{collocations}
\alias{collocations}
\alias{collocations,partition-method}
\alias{collocations,partitionCluster-method}
\alias{collocations-method}
\title{calculate all collocations in a partition}
\usage{
\S4method{collocations}{partition}(object, pAttribute = "word", window = 5,
  method = "ll", filter = TRUE, posFilter = c("ADJA", "NN"),
  progress = TRUE, mc = FALSE)
}
\arguments{
\item{object}{a partition object}

\item{pAttribute}{p-attribute, typically "word" or "token"}

\item{window}{no of tokens to the left and to the right}

\item{method}{the statistical test to use}

\item{filter}{defaults to TRUE}

\item{posFilter}{what POS to keep}

\item{progress}{logical, whether to show progress bar}

\item{mc}{whether to use multicore}
}
\value{
a data frame
}
\description{
the result is meant to serve as a result for an analysis of collocation graphs
}
\examples{
\dontrun{
bt17merkel <- partition("PLPRTXT", list(text_lp="17", text_type="speech", text_speaker="Angela Merkel"))
bt17merkelColl <- collocations(bt17merkel)
}
}
\author{
Andreas Blaette
}

