% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/collocationsCluster-class.R
\docType{class}
\name{collocationsCluster-class}
\alias{as.TermDocumentMatrix,collocationsCluster-method}
\alias{collocationsCluster}
\alias{collocationsCluster-class}
\title{S4 collocationsCluster class}
\description{
class to organize information of multiple context analyses
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"} a list of context objects}
}}

