% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{collocationsReshaped}
\alias{collocationsReshaped}
\alias{merge,collocationsReshaped-method}
\title{Methods for manipulating collocationsReshaped-class-objects}
\usage{
\S4method{merge}{collocationsReshaped}(x, y)
}
\arguments{
\item{x}{collocations for a corpus of interest}

\item{y}{collocations for a reference corpus}
}
\description{
Methods for manipulating collocationsReshaped-class-objects
}

