% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_class.R, R/view_method.R
\docType{class}
\name{comp-class}
\alias{comp-class}
\alias{compBundle-class}
\alias{compCooccurrences-class}
\alias{compNgrams-class}
\alias{show,comp-method}
\alias{summary,comp-method}
\alias{summary,compBundle-method}
\alias{view,comp-method}
\title{S4 class for comparing corpora}
\usage{
\S4method{summary}{comp}(object)

\S4method{show}{comp}(object)

\S4method{summary}{compBundle}(object)

\S4method{view}{comp}(.Object)
}
\arguments{
\item{object}{an object}

\item{.Object}{an object}
}
\description{
object resulting from compare-method
}
\section{Slots}{

\describe{
\item{\code{corpus}}{Object of class \code{"character"}}

\item{\code{pAttribute}}{Object of class \code{"character"}}

\item{\code{encoding}}{Object of class \code{"character"}}

\item{\code{corpus}}{Object of class \code{"character"}}

\item{\code{stat}}{Object of class \code{"data.frame"}}

\item{\code{sizeCoi}}{Object of class \code{"numeric"}}

\item{\code{sizeRef}}{Object of class \code{"numeric"}}

\item{\code{included}}{Object of class \code{"logical"} whether corpus of interest is included in reference corpus}

\item{\code{method}}{Object of class \code{"character"} statisticalTest used}

\item{\code{call}}{Object of class \code{"character"} the call that generated the object
@section Methods:
 \describe{
  \item{summary}{\code{signature(object = "comp")}: Display essential information }
  \item{addPos}{\code{signature(object = "comp")}: add POS attribute to statistics table }
  \item{as.data.frame}{\code{signature(object = "comp")}: get the statistics table}
  }}

\item{\code{objects}}{an object of class \code{list}}
}}
\author{
Andreas Blaette
}

