% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coerce_methods.R, R/context_class.R, R/context_methods.R, R/dispersion_method.R, R/tf_method.R
\docType{methods}
\name{as.partitionBundle,context-method}
\alias{[,context,ANY,ANY,ANY-method}
\alias{[,context-method}
\alias{[[,context,ANY-method}
\alias{[[,context-method}
\alias{as.DataTables,context-method}
\alias{as.partitionBundle,context-method}
\alias{context-class}
\alias{dispersion,context-method}
\alias{head,context-method}
\alias{ll,context-method}
\alias{show,context-method}
\alias{summary,context-method}
\alias{tail,context-method}
\alias{tf,context-method}
\title{S4 context class}
\usage{
\S4method{as.partitionBundle}{context}(object, mc = FALSE)

\S4method{tail}{context}(x, n = 10)

\S4method{[[}{context,ANY}(x, i)

\S4method{dispersion}{context}(object, sAttribute)

\S4method{tf}{context}(object)
}
\arguments{
\item{object}{object}

\item{mc}{logical, whether to use multicore}

\item{x}{object}

\item{n}{for head and tail}

\item{i}{for indexing}

\item{sAttribute}{the s-attribute to use}
}
\description{
class to organize information of context analysis

Based on a context object, get distribution of hits
}
\section{Slots}{

\describe{
\item{\code{query}}{Object of class \code{"character"} node examined}

\item{\code{frequency}}{Object of class \code{"numeric"} number of hits}

\item{\code{partition}}{Object of class \code{"character"} the partition the analysis is based on}

\item{\code{partitionSize}}{Object of class \code{"numeric"} the size of the partition}

\item{\code{leftContext}}{Object of class \code{"numeric"} number of tokens to the right}

\item{\code{rightContext}}{Object of class \code{"numeric"} number of tokens to the left}

\item{\code{size}}{Object of class \code{"numeric"} number of tokens in the right and left context}

\item{\code{pAttribute}}{Object of class \code{"character"} p-attribute of the query}

\item{\code{sAttribute}}{Object of ...}

\item{\code{corpus}}{Object of class \code{"character"} the CWB corpus used}

\item{\code{stat}}{Object of class \code{"data.frame"} statistics of the analysis}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{posFilter}}{Object of class \code{"character"} part-of-speech tags filtered}

\item{\code{cpos}}{Object of class \code{"list"} corpus positions of the hits}

\item{\code{statisticalTest}}{Object of class \code{"character"} statistical test used}

\item{\code{cutoff}}{Object of class \code{"list"} cutoff levels that have been applied}

\item{\code{call}}{Object of class \code{"character"} call that generated the object}
}}
\section{Methods}{

  \describe{
    \item{show}{output of core information}
    \item{summary}{core statistical information}
    \item{[}{index the object}
    \item{[[}{specific collocates}
    \item{trim}{trim the object}
    \item{as.data.frame}{get the statistics table}
   }
}

