% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/context-method.R
\docType{methods}
\name{context}
\alias{as.TermContextMatrix,contextCluster-method}
\alias{as.matrix,contextCluster-method}
\alias{context}
\alias{context,collocations-method}
\alias{context,contextCluster-method}
\alias{context,missing-method}
\alias{context,partition-method}
\alias{context,partitionCluster-method}
\alias{ll}
\alias{ll-method}
\title{Analyze context of a node word}
\usage{
\S4method{context}{partition}(object, query, pAttribute = NULL,
  sAttribute = "text_id", leftContext = NULL, rightContext = NULL,
  minSignificance = 0, posFilter = NULL, filterType = "exclude",
  stoplist = NULL, positivelist = NULL, statisticalTest = "ll",
  mc = NULL, verbose = TRUE)

\S4method{context}{partitionCluster}(object, query, pAttribute = NULL,
  sAttribute = "text_id", leftContext = NULL, rightContext = NULL,
  minSignificance = NULL, posFilter = NULL, filterType = "exclude",
  stoplist = c(), positivelist = NULL, statisticalTest = "ll",
  mc = FALSE, verbose = TRUE)

\S4method{context}{collocations}(object, query, complete = FALSE)

\S4method{context}{missing}()
}
\arguments{
\item{object}{a partition or a partitionCluster object}

\item{query}{query, which may by a character vector or a cqpQuery object}

\item{pAttribute}{p-attribute of the query}

\item{sAttribute}{if provided, it will be checked that cpos do not extend beyond
the region defined by the s-attribute}

\item{leftContext}{no of tokens and to the left of the node word}

\item{rightContext}{no of tokens to the right of the node word}

\item{minSignificance}{minimum log-likelihood value}

\item{posFilter}{character vector with the POS tags to be included - may not
be empty!!}

\item{filterType}{either "include" or "exclude"}

\item{stoplist}{exclude a query hit from analysis if stopword(s) is/are in
context}

\item{positivelist}{character vector or numeric vector: include a query hit
only if token in positivelist is present. If positivelist is a character
vector, it is assumed to provide regex expressions (incredibly long if the
list is long)}

\item{statisticalTest}{either "LL" (default) or "pmi", if NULL, calculating
the statistics will be skipped}

\item{mc}{whether to use multicore; if NULL (default), the function will get
the value from the session settings}

\item{verbose}{report progress, defaults to TRUE}
}
\value{
depending on whether a partition or a partitionCluster serves as
  input, the return will be a context object, or a contextCluster object
}
\description{
Retrieve the concordances of a token and calculate the log-likelihood test
for collocates For formulating the query, CPQ syntax may be used (see
examples).
}
\examples{
\dontrun{
p <- partition("PLPRBTTXT", list(text_type="speech"))
a <- context(p, "Integration", "word")
}
}
\author{
Andreas Blaette
}

