% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences_class.R, R/cooccurrences_methods.R, R/trim_method.R, R/view_method.R
\docType{class}
\name{cooccurrences-class}
\alias{[,cooccurrences,ANY,ANY,ANY-method}
\alias{[,cooccurrences-method}
\alias{as.igraph,cooccurrences-method}
\alias{as.sparseMatrix,cooccurrences-method}
\alias{cooccurrences-class}
\alias{cooccurrencesReshaped-class}
\alias{show,cooccurrences-method}
\alias{summary,cooccurrences-method}
\alias{trim,cooccurrences-method}
\alias{view,cooccurrencesReshaped-method}
\title{cooccurrences}
\usage{
\S4method{show}{cooccurrences}(object)

\S4method{summary}{cooccurrences}(object)

\S4method{as.sparseMatrix}{cooccurrences}(x, col, mc = FALSE)

\S4method{trim}{cooccurrences}(object, by = NULL)

\S4method{view}{cooccurrencesReshaped}(.Object)
}
\arguments{
\item{object}{object}

\item{x}{object}

\item{col}{column}

\item{mc}{whether to use multicore}

\item{by}{by}

\item{.Object}{object}
}
\description{
class to organize information of context analysis
}
\section{Slots}{

\describe{
\item{\code{call}}{Object of class \code{"character"} the call that generated the object}

\item{\code{partition}}{Object of class \code{"character"} the partition the analysis is based on}

\item{\code{partitionSize}}{Object of class \code{"numeric"} the size of the partition}

\item{\code{left}}{Object of class \code{"numeric"} number of tokens to the right}

\item{\code{right}}{Object of class \code{"numeric"} number of tokens to the left}

\item{\code{pAttribute}}{Object of class \code{"character"} p-attribute of the query}

\item{\code{corpus}}{Object of class \code{"character"} the CWB corpus used}

\item{\code{stat}}{Object of class \code{"data.frame"} statistics of the analysis}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{pos}}{Object of class \code{"character"} part-of-speech tags filtered}

\item{\code{method}}{Object of class \code{"character"} statistical test(s) used}

\item{\code{cutoff}}{Object of class \code{"list"} cutoff levels that have been applied}

\item{\code{svg}}{Object of class \code{"character"} - valid XML with svg representation}
}}

