% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooccurrences_class.R, R/cooccurrencesReshaped_methods.R
\docType{methods}
\name{cooccurrencesReshaped}
\alias{cooccurrencesReshaped}
\alias{merge,cooccurrencesReshaped-method}
\title{Methods for manipulating cooccurrencesReshaped-class-objects}
\usage{
\S4method{merge}{cooccurrencesReshaped}(x, y)
}
\arguments{
\item{x}{cooccurrences for a corpus of interest}

\item{y}{cooccurrences for a reference corpus}
}
\description{
Methods for manipulating cooccurrencesReshaped-class-objects
}

