% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corpus_class.R
\docType{data}
\name{Corpus}
\alias{Corpus}
\title{Corpus class.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Corpus
}
\arguments{
\item{id2str}{logical, whether to perform id2str when doing count}
}
\description{
Corpus class.
}
\section{Fields}{

\describe{
\item{\code{corpus}}{character vector, CWB corpus}

\item{\code{encoding}}{encoding of the corpus}

\item{\code{count}}{data.table with counts}
}}
\section{Methods}{

\describe{
  \item{one}{bla}
}
}
\keyword{datasets}

