% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Corpus_class.R
\docType{data}
\name{Corpus}
\alias{Corpus}
\title{Corpus class.}
\format{\preformatted{Class 'R6ClassGenerator' <Corpus> object generator
  Public:
    corpus: NULL
    encoding: NULL
    pAttribute: NULL
    size: NULL
    stat: data.table, data.frame
    initialize: function (corpus, pAttribute = NULL) 
    count: function (pAttribute = getOption("polmineR.pAttribute")) 
    as.partition: function () 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:polmineR>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "Corpus_generator"
}}
\usage{
Corpus
}
\description{
Corpus class.
}
\section{Fields}{

\describe{
\item{\code{corpus}}{character vector, CWB corpus}

\item{\code{encoding}}{encoding of the corpus}

\item{\code{count}}{data.table with counts}
}}
\section{Methods}{

\describe{
  \item{one}{bla}
}
}
\keyword{datasets}

