% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpus_method.R
\docType{methods}
\name{corpus}
\alias{corpus}
\alias{corpus,partition-method}
\alias{corpus,bundle-method}
\alias{corpus,missing-method}
\title{Get corpus.}
\usage{
corpus(object)

\S4method{corpus}{partition}(object)

\S4method{corpus}{bundle}(object)

\S4method{corpus}{missing}()
}
\arguments{
\item{object}{the object}
}
\description{
Calling \code{corpus()} will return the corpora available. If the 
param 'packges' (logical) is TRUE, packages that include a corpus
are returned.
}
\details{
If object is a partition or partitionBundle-object, the corpus the
respective object is derived from is returned.
}
