% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/corpus.R, R/split.R
\docType{class}
\name{corpus-class}
\alias{corpus-class}
\alias{zoom}
\alias{name,corpus-method}
\alias{show,corpus-method}
\alias{$,corpus-method}
\alias{split,corpus-method}
\alias{split,subcorpus_bundle-method}
\title{S4 class to wrap information on CWB corpora.}
\usage{
\S4method{name}{corpus}(x)

\S4method{show}{corpus}(object)

\S4method{$}{corpus}(x, name)

\S4method{split}{corpus}(x, s_attribute, values = NULL, prefix = "",
  mc = getOption("polmineR.mc"), verbose = TRUE, progress = FALSE,
  type = get_type(x), ...)

\S4method{split}{subcorpus_bundle}(x, s_attribute, prefix = "",
  progress = TRUE, mc = getOption("polmineR.mc"))
}
\arguments{
\item{x}{An object of class \code{corpus}, or inheriting from it.}

\item{object}{An object.}

\item{name}{A (single) s-attribute.}

\item{s_attribute}{The s-attribute to vary.}

\item{values}{Values the s-attribute provided shall assume.}

\item{prefix}{A character vector that will be attached as a prefix to partition names.}

\item{mc}{Logical, whether to use multicore parallelization.}

\item{verbose}{Logical, whether to provide progress information.}

\item{progress}{Logical, whether to show progress bar.}

\item{type}{The type of \code{partition} to generate.}

\item{...}{Further arguments.}
}
\description{
S4 class to wrap information on CWB corpora.
}
\details{
A \code{corpus} object can have a name, which can be retrieved using
the \code{name}-method.

Applying the `$`-method on a corpus will return the values for the
  s-attribute stated with argument \code{name}.

Applying the \code{split}-method to a \code{subcorpus_bundle}-object
  will iterate through the subcorpus, and apply \code{split} on each
  \code{subcorpus} object in the bundle, splitting it up by the s-attribute
  provided by the argument \code{s_attribute}. The return value is a
  \code{subcorpus_bundle}, the names of which will be the names of the
  incoming \code{partition_bundle} concatenated with the s-attribute values
  used for splitting. The argument \code{prefix} can be used to achieve a
  more descriptive name.
}
\section{Slots}{

\describe{
\item{\code{corpus}}{A length-one \code{character} vector, a CWB corpus.}

\item{\code{data_dir}}{The directory where the files for the indexed corpus are.}

\item{\code{type}}{The type of the corpus (e.g. "plpr" for a corpus of plenary protocols).}

\item{\code{name}}{An additional name for the object that may be more telling than the corpus ID.}

\item{\code{encoding}}{The encoding of the corpus, given as a length-one \code{character} vector.}
}}

\examples{
g <- corpus("GERMAPARLMINI")
g$date
corpus("GERMAPARLMINI")$date

sc <- subset("GERMAPARLMINI", date == "2009-10-27")
sc$date
gparl <- corpus("GERMAPARLMINI")
b <- split(gparl, s_attribute = "date")
# split up objects in partition_bundle by using partition_bundle-method
use("polmineR")
y <- corpus("GERMAPARLMINI") \%>\%
  split(s_attribute = "date") \%>\%
  split(s_attribute = "speaker")

summary(y)
}
\seealso{
Other classes to manage corpora: \code{\link{regions}},
  \code{\link{subcorpus-class}}
}
\concept{classes to manage corpora}
