% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R
\docType{class}
\name{corpus-class}
\alias{corpus-class}
\alias{zoom}
\title{S4 class to wrap information on CWB corpora.}
\arguments{
\item{x}{A \code{corpus} object.}

\item{...}{Further arguments.}

\item{name}{An s-attribute that will be assigned as key to a
\code{partition}.}

\item{e1}{A first expression, a \code{partition} object in this case.}

\item{e2}{A second expression, the value of an s-attribute.}

\item{table}{Values a s-attribute shall assume.}
}
\description{
S4 class to wrap information on CWB corpora.
}
\section{Slots}{

\describe{
\item{\code{corpus}}{A length-one \code{character} vector, a CWB corpus.}

\item{\code{data_dir}}{The directory where the files for the indexed corpus are.}

\item{\code{type}}{The type of the corpus (e.g. "plpr" for a corpus of plenary protocols).}

\item{\code{encoding}}{The encoding of the corpus, given as a length-one \code{character} vector.}

\item{\code{key}}{A length-one \code{character} vector stating a s-attribute that
serves as a key.}
}}

\seealso{
Other classes to manage corpora: \code{\link{regions}},
  \code{\link{subcorpus-class}}
}
\concept{classes to manage corpora}
