% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/corpus.R
\docType{class}
\name{corpus-class}
\alias{corpus-class}
\alias{zoom}
\alias{name,corpus-method}
\alias{show,corpus-method}
\alias{$,corpus-method}
\title{S4 class to wrap information on CWB corpora.}
\usage{
\S4method{name}{corpus}(x)

\S4method{show}{corpus}(object)

\S4method{$}{corpus}(x, name)
}
\arguments{
\item{x}{An object of class \code{corpus}, or inheriting from it.}

\item{object}{An object.}

\item{name}{A (single) s-attribute.}

\item{...}{Further arguments.}
}
\description{
S4 class to wrap information on CWB corpora.
}
\details{
A \code{corpus} object can have a name, which can be retrieved using
the \code{name}-method.

Applying the `$`-method on a corpus will return the values for the
  s-attribute stated with argument \code{name}.
}
\section{Slots}{

\describe{
\item{\code{corpus}}{A length-one \code{character} vector, a CWB corpus.}

\item{\code{data_dir}}{The directory where the files for the indexed corpus are.}

\item{\code{type}}{The type of the corpus (e.g. "plpr" for a corpus of plenary protocols).}

\item{\code{name}}{An additional name for the object that may be more telling than the corpus ID.}

\item{\code{encoding}}{The encoding of the corpus, given as a length-one \code{character} vector.}
}}

\examples{
g <- corpus("GERMAPARLMINI")
g$date
corpus("GERMAPARLMINI")$date

sc <- subset("GERMAPARLMINI", date == "2009-10-27")
sc$date
}
\seealso{
Other classes to manage corpora: \code{\link{regions}},
  \code{\link{subcorpus}}
}
\concept{classes to manage corpora}
