% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_method.R
\docType{methods}
\name{count}
\alias{count}
\alias{count,character-method}
\alias{count,partition-method}
\alias{count,partitionBundle-method}
\alias{count-method}
\title{get counts}
\usage{
count(.Object, ...)

\S4method{count}{partition}(.Object, query, pAttribute = NULL, mc = F,
  verbose = T)

\S4method{count}{partitionBundle}(.Object, query, pAttribute = NULL, mc = F,
  verbose = T)

\S4method{count}{character}(.Object, query, pAttribute = NULL,
  verbose = TRUE)
}
\arguments{
\item{.Object}{either a partition or a partitionBundle object}

\item{...}{further parameters}

\item{query}{a character vector (one or multiple terms to be looked up)}

\item{mc}{logical, whether to use multicore (defaults to FALSE)}

\item{verbose}{logical, whether to be verbose}
}
\description{
Count number of occurrences of a query (CQP syntax can be used).
}
\examples{
# generate a partition for testing 
test <- partition("PLPRBTTXT")
count(test, "Wir") # get frequencies for one token
count(test, c("Wir", "lassen", "uns")) # get frequencies for multiple tokens
count("PLPRTXT", c("machen", "Integration"), "word")
}
\seealso{
count
}

