% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_method.R
\docType{methods}
\name{count}
\alias{count}
\alias{count,character-method}
\alias{count,partition-method}
\alias{count,partitionBundle-method}
\alias{count-method}
\title{get counts}
\usage{
count(.Object, ...)

\S4method{count}{partition}(.Object, query, pAttribute = NULL, mc = F,
  verbose = T)

\S4method{count}{partitionBundle}(.Object, query, pAttribute = NULL,
  freq = FALSE, mc = F, verbose = T)

\S4method{count}{character}(.Object, query, pAttribute = NULL,
  verbose = TRUE)
}
\arguments{
\item{.Object}{a \code{"partition"} or \code{"partitionBundle"} object, or a character vector (length 1) providing the name of a corpus}

\item{...}{further parameters}

\item{query}{a character vector (one or multiple terms to be looked up)}

\item{pAttribute}{the p-attribute(s) to use}

\item{mc}{logical, whether to use multicore (defaults to FALSE)}

\item{verbose}{logical, whether to be verbose}

\item{freq}{logical, if FALSE, counts will be reported, if TRUE, frequencies}
}
\value{
a \code{"data.table"}
}
\description{
Count number of occurrences of a query. The CQP syntax can be used to formulate 
the query.
}
\examples{
use("polmineR.sampleCorpus")
debates <- partition("PLPRBTTXT", list(text_id=".*"), regex=TRUE)
count(debates, "Arbeit") # get frequencies for one token
count(debates, c("Arbeit", "Freizeit", "Zukunft")) # get frequencies for multiple tokens
count("PLPRBTTXT", c("Migration", "Integration"), "word")

debates <- partitionBundle(
  object="PLPRBTTXT",
  def=list(text_id=".*"),
  var=list(text_date=sAttributes("PLPRBTTXT", "text_date")),
  regex=TRUE, mc=FALSE, verbose=FALSE
)
count(debates, c("Arbeit", "Integration", "Umwelt"))
}
\seealso{
count
}

