% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_method.R
\docType{methods}
\name{count}
\alias{count}
\alias{count,character-method}
\alias{count,partition-method}
\alias{count,partitionBundle-method}
\alias{count-method}
\title{get counts}
\usage{
count(.Object, ...)

\S4method{count}{partition}(.Object, query, cqp = is.cqp,
  pAttribute = getOption("polmineR.pAttribute"),
  mc = getOption("polmineR.cores"), verbose = T, progress = F)

\S4method{count}{partitionBundle}(.Object, query, pAttribute = NULL,
  freq = FALSE, total = T, mc = F, progress = T, verbose = FALSE)

\S4method{count}{character}(.Object, query,
  pAttribute = getOption("polmineR.pAttribute"), verbose = TRUE)
}
\arguments{
\item{.Object}{a \code{"partition"} or \code{"partitionBundle"} object, or a
character vector (length 1) providing the name of a corpus}

\item{...}{further parameters}

\item{query}{a character vector (one or multiple terms to be looked up), CQP
syntax can be used.}

\item{cqp}{either logical (TRUE if query is a CQP query), or a
function to check whether query is a CQP query or not (defaults to is.query
auxiliary function)}

\item{pAttribute}{the p-attribute(s) to use}

\item{mc}{logical, whether to use multicore (defaults to FALSE)}

\item{verbose}{logical, whether to be verbose}

\item{progress}{logical, whether to show progress}

\item{freq}{logical, if FALSE, counts will be reported, if TRUE, frequencies}

\item{total}{defaults to FALSE, if TRUE, the added value of counts (column:
TOTAL) will be amended to the data.table that is returned}
}
\value{
a \code{"data.table"}
}
\description{
Count number of occurrences of a query (CQP syntax may be used) in a
partition, or a partitionBundle.
}
\examples{
\dontrun{
if (require(polmineR.sampleCorpus) && require(rcqp)){
  use("polmineR.sampleCorpus")
  debates <- partition("PLPRBTTXT", list(text_id=".*"), regex=TRUE)
  x <- count(debates, "Arbeit") # get frequencies for one token
  x <- count(debates, c("Arbeit", "Freizeit", "Zukunft")) # get frequencies for multiple tokens
  x <- count("PLPRBTTXT", c("Migration", "Integration"), "word")

  debates <- partitionBundle(
    .Object="PLPRBTTXT",
    def=list(text_date=sAttributes("PLPRBTTXT", "text_date")),
    regex=TRUE, mc=FALSE, verbose=FALSE
  )
  aiu <- count(debates, c("Arbeit", "Integration", "Umwelt"))
}
}
}
\seealso{
For a metadata-based breakdown of counts
(i.e. a differentiation by s-attributes), see \code{"dispersion"}.

count
}

