% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpos_method.R
\docType{methods}
\name{cpos}
\alias{cpos}
\alias{cpos,character-method}
\alias{cpos,partition-method}
\title{corpus positions for a query}
\usage{
cpos(.Object, ...)

\S4method{cpos}{character}(.Object, query, pAttribute = NULL,
  encoding = NULL, verbose = TRUE)

\S4method{cpos}{partition}(.Object, query, pAttribute = NULL,
  verbose = TRUE)
}
\arguments{
\item{.Object}{a character vector (length 1) giving the name of a corpus, or a partition object}

\item{...}{further arguments}

\item{query}{a query, either a single term, or a CQP query}

\item{pAttribute}{the p-attribute to search}

\item{encoding}{the encoding of the corpus (if NULL, the
encoding provided in the registry file of the corpus will be used)}

\item{verbose}{logical, whether to be talkative}
}
\value{
You get a matrix with two columns, the first column giving the start cpos of the hits obtained,
the second column giving the end cpos of the respective hit. The number of rows is the number of hits.
If there are no hits, a NULL object will be returned.
}
\description{
Get the corpus positions (cpos) for a query. The query may also be a 
CQP query.
}

