% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cpos-method.R
\docType{methods}
\name{cpos}
\alias{cpos}
\alias{cpos,character-method}
\title{get the corpus positions for a query}
\usage{
cpos(.Object, ...)

\S4method{cpos}{character}(.Object, query, pAttribute = NULL,
  encoding = NULL, verbose = TRUE)
}
\arguments{
\item{.Object}{an Object}

\item{...}{further arguments}

\item{query}{a query, may be CQP}

\item{pAttribute}{which attribute to search}
}
\description{
get the corpus positions for a query
}

