% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.cqp}
\alias{as.cqp}
\alias{is.cqp}
\title{CQP queries}
\usage{
is.cqp(query)

as.cqp(queries, normalise_case = FALSE, collapse = FALSE)
}
\arguments{
\item{query}{query to be checked}

\item{queries}{a character vector}

\item{normalise_case}{logical}

\item{collapse}{whether collapse the queries into one}
}
\value{
a character vector
}
\description{
Basic test whether query is a CQP query.

Takes a simple string as an imput and converts it to a valid CQP query
}

