\docType{class}
\name{cqpQuery-class}
\alias{[,cqpQuery-method}
\alias{[[,cqpQuery-method}
\alias{cqpQuery-class}
\alias{show,cqpQuery-method}
\title{CQP query object}
\description{
Object to store and check a CQP query, and to supply a few
methods.
}
\section{Slots}{
  \describe{ \item{\code{query}:}{Object of class
  \code{"character"} the query } }
}

\section{Methods}{
  \describe{ \item{[}{get frequency of a query} }
}
\author{
Andreas Blaette
}

