% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_method.R
\docType{methods}
\name{decode,Corpus-method}
\alias{decode,Corpus-method}
\title{Decode corpus.}
\usage{
\S4method{decode}{Corpus}(.Object, verbose = TRUE)
}
\arguments{
\item{.Object}{the corpus to decode (character vector)}

\item{verbose}{logical}
}
\description{
An encoded CWB corpus is turned into a tibble.
}
\examples{
\dontrun{
PLPRBTTXT <- Corpus$new("PLPRBTTXT")
}
}

