% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode_method.R
\docType{methods}
\name{decode}
\alias{decode}
\alias{decode,Corpus-method}
\alias{decode,character-method}
\title{Decode corpus.}
\usage{
decode(.Object, ...)

\S4method{decode}{Corpus}(.Object, verbose = TRUE)

\S4method{decode}{character}(.Object, sAttribute, verbose = TRUE)
}
\arguments{
\item{.Object}{the corpus to decode (character vector)}

\item{...}{further parameters}

\item{verbose}{logical}

\item{sAttribute}{the s-attribute to decode}
}
\description{
An encoded CWB corpus is turned into a tibble.
}
\examples{
\dontrun{
PLPRBTTXT <- Corpus$new("PLPRBTTXT")
}
}

