% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dispersion-class.R
\docType{class}
\name{dispersion-class}
\alias{dispersion-class}
\title{dispersion class}
\description{
class to organize results from dispersion analysis
}
\section{Slots}{

\describe{
\item{\code{abs}}{Object of class \code{"character"} the call that generated the object}

\item{\code{rel}}{Object of class \code{"character"} the partition the analysis is based on}

\item{\code{subcorpusSize}}{Object of class \code{"numeric"} the size of the partition}
}}

