% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_method.R
\docType{methods}
\name{divide}
\alias{divide}
\alias{divide,matrix-method}
\alias{divide,vector-method}
\title{divide an object into equally sized parts}
\usage{
divide(.Object, n)

\S4method{divide}{matrix}(.Object, n)

\S4method{divide}{vector}(.Object, n)
}
\arguments{
\item{.Object}{object to divide}

\item{n}{number of objects}
}
\description{
divide an object into equally sized parts
}
