% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_method.R
\docType{methods}
\name{dotplot}
\alias{dotplot}
\alias{dotplot,comp-method}
\alias{dotplot,partition-method}
\alias{dotplot,textstat-method}
\title{dotplot}
\usage{
dotplot(.Object, ...)

\S4method{dotplot}{textstat}(.Object, col = NULL, n = 20, ...)

\S4method{dotplot}{comp}(.Object, col = NULL, n = 20, ...)

\S4method{dotplot}{partition}(.Object, col = "count", n = 20, ...)
}
\arguments{
\item{.Object}{object}

\item{...}{further arguments that will be passed into the dotchart function}

\item{col}{column}

\item{n}{number}
}
\description{
dotplot
}

