% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding.R
\docType{methods}
\name{encoding}
\alias{encoding}
\alias{encoding<-}
\alias{encoding,textstat-method}
\alias{encoding,bundle-method}
\title{Get and set encoding.}
\usage{
encoding(object)

encoding(object) <- value

\S4method{encoding}{textstat}(object)

\S4method{encoding}{bundle}(object)
}
\arguments{
\item{object}{the object with an 'encoding'-slot}

\item{value}{value to be assigned}
}
\description{
Method for \code{textstat} objects and classes inheriting from \code{textstat}.
}
