\docType{methods}
\name{enrich,keyness-method}
\alias{enrich,keyness-method}
\title{Enrich keyness object}
\usage{
\S4method{enrich}{keyness}(object, addPos = NULL, verbose = TRUE)
}
\arguments{
  \item{object}{a keyness object}

  \item{addPos}{no idea ....}

  \item{verbose}{whether R should be talkative}
}
\description{
Wrapper for adding a statistic on pos attributes to keyness
object
}

