% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrich-method.R
\docType{methods}
\name{enrich}
\alias{enrich}
\alias{enrich-method}
\title{enrich an object}
\usage{
enrich(object, ...)
}
\arguments{
\item{object}{a partition, partitionCluster or keyness object}

\item{...}{further parameters}
}
\description{
Method to fill slots of a partition, partitionCluster or keyness object that
have not been set up previously. See the respective documentation:
\describe{
 \item{partition:}{\code{method?enrich("partition")}}
 \item{partitionCluster:}{\code{method?enrich("partitionCluster")}}
 \item{keyness:}{\code{method?enrich("keyness")}}
}
}

