% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrich_method.R
\docType{methods}
\name{enrich}
\alias{enrich}
\alias{enrich,compBundle-method}
\alias{enrich-method}
\title{enrich an object}
\usage{
enrich(object, ...)

\S4method{enrich}{compBundle}(object, addPos = NULL, verbose = TRUE,
  mc = getOption("polmineR.mc"))
}
\arguments{
\item{object}{a partition, partitionBundle or comp object}

\item{...}{further parameters}

\item{addPos}{depracated}

\item{verbose}{logical}

\item{mc}{logical or, if numeric, providing the number of cores}
}
\description{
Method to fill slots of a partition, partitionBundle or comp object that
have not been set up previously. See the respective documentation:
}

