% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_method.R
\docType{methods}
\name{enrich}
\alias{enrich}
\alias{enrich-method}
\title{enrich an object}
\usage{
enrich(object, ...)
}
\arguments{
\item{object}{a partition, partitionBundle or comp object}

\item{...}{further parameters}
}
\description{
Method to fill slots of a partition, partitionBundle or comp object that 
have not been set up previously. See the respective documentation:
}

