% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich.R
\docType{methods}
\name{enrich}
\alias{enrich}
\alias{enrich-method}
\title{Enrich an object.}
\usage{
enrich(.Object, ...)
}
\arguments{
\item{.Object}{a \code{partition}, \code{partition_bundle} or comp object}

\item{...}{further parameters}
}
\description{
Methods to enrich objects with additional (statistical) information. The methods are documented
with the classes to which they adhere. See the references in the \code{seealso}-section.
}
\seealso{
The enrich method is defined for the following classes:
"partition", (see \code{\link{partition_class}}),
"partition_bundle" (see \code{\link{partition_bundle-class}}),
"kwic" (see \code{\link{kwic-class}}), and
"context" (see \code{\link{context-class}}). See the linked documentation
to learn how the enrich method can be applied to respective objects.
}
