% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/enrich_method.R
\docType{methods}
\name{enrich,partition-method}
\alias{enrich,partition-method}
\title{enrich partition object (S4 method)}
\usage{
\S4method{enrich}{partition}(object, size = FALSE, pAttribute = NULL,
  id2str = TRUE, meta = NULL, verbose = TRUE, mc = FALSE)
}
\arguments{
\item{object}{a partition object}

\item{size}{logical, defaults to FALSE}

\item{pAttribute}{character vector providing the p-attributes for which frequency
tables shall be generated}

\item{meta}{character vector providing s-attributes for which metainformation shall be supplied}

\item{verbose}{logical, defaults to TRUE}

\item{addPos}{character vector providing p-attributes}
}
\description{
Fill slots of a partition object: Add object size, metainformation, pos.
}

