% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flatten}
\alias{flatten}
\title{flatten a nested list}
\usage{
flatten(object)
}
\arguments{
\item{object}{a list (with partitionCluster objects)}
}
\value{
a partitionCluster object
}
\description{
If you have a list of partitionClusters, this function will flatten the data
structure and return a partition Cluster object.
}

