% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten}
\alias{flatten}
\title{flatten a nested list}
\usage{
flatten(object)
}
\arguments{
\item{object}{a list (with \code{partition_bundle} objects)}
}
\value{
a \code{partition_bundle} object
}
\description{
If you have a list of \code{partition_bundle} objects, this function will flatten the data
structure and return a \code{partition_bundle} object.
}
