% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten}
\alias{flatten}
\title{flatten a nested list}
\usage{
flatten(object)
}
\arguments{
\item{object}{a list (with partitionBundle objects)}
}
\value{
a partitionBundle object
}
\description{
If you have a list of partitionBundles, this function will flatten the data
structure and return a partition Bundle object.
}
