% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequencies_method.R
\docType{methods}
\name{frequencies}
\alias{frequencies}
\alias{frequencies-method}
\alias{frequencies,partition-method}
\alias{frequencies,partitionBundle-method}
\title{Frequency breakdown of the variation of query results}
\usage{
\S4method{frequencies}{partition}(object, query,
  pAttribute = getOption("polmineR.pAttribute"))
}
\arguments{
\item{object}{a partition object}

\item{query}{a query, CQP syntax may be used}

\item{pAttribute}{pAttribute}
}
\value{
a matrix
}
\description{
If one query is supplied, the function returns a frequency breakdown of the results of the query.
If several queries are supplied, frequencies for the individual queries are retrieved.
}
