% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEncoding_method.R
\docType{methods}
\name{getEncoding}
\alias{getEncoding}
\alias{getEncoding,character-method}
\title{get the encoding of a corpus}
\usage{
getEncoding(.Object)

\S4method{getEncoding}{character}(.Object)
}
\arguments{
\item{.Object}{the corpus name}
}
\description{
get the encoding of a corpus
}

