% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getTerms-method.R
\name{getTerms}
\alias{getTerms}
\title{get terms available in a corpus or partition}
\usage{
getTerms(.Object, ...)
}
\arguments{
\item{.Object}{the object}

\item{pAttribute}{the pAttribute}

\item{mc}{logical, whether to use multicore}

\item{robust}{logical, whether to be robus}
}
\description{
get terms available in a corpus or partition
}

