% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_method.R
\docType{methods}
\name{highlight}
\alias{highlight}
\alias{highlight,partition-method}
\alias{highlight,character-method}
\alias{highlight,html-method}
\title{Highlight tokens in fulltext.}
\usage{
highlight(.Object, ...)

\S4method{highlight}{partition}(.Object, html, highlight = list(),
  cqp = is.cqp, tooltips = NULL)

\S4method{highlight}{character}(.Object, highlight = list(),
  tooltips = NULL)

\S4method{highlight}{html}(.Object, highlight = list(), tooltips = NULL)
}
\arguments{
\item{.Object}{either a \code{"partition"} or \code{"character"} object}

\item{...}{further parameters}

\item{html}{character vector with a website}

\item{highlight}{a \code{"list"} of character vectors, the names need to provide the colors}

\item{cqp}{logical, whether to use CQP syntax}

\item{tooltips}{a \code{"list"} of character vectors, all names need to be included in lists of the highlight-list}
}
\description{
Highlight tokens by term or cpos in a html document generated by
read/html.
}
