% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hits_class.R
\docType{class}
\name{hits-class}
\alias{hits}
\alias{hits,character-method}
\alias{hits,partition-method}
\alias{hits,partitionBundle-method}
\alias{hits-class}
\alias{sample,hits-method}
\title{hits class}
\usage{
hits(.Object, ...)

\S4method{hits}{character}(.Object, query, sAttribute = NULL,
  pAttribute = "word", size = TRUE, freq = FALSE, mc = FALSE,
  verbose = TRUE, progress = TRUE)

\S4method{hits}{partition}(.Object, query, sAttribute = NULL,
  pAttribute = "word", size = FALSE, freq = FALSE, mc = FALSE,
  progress = FALSE, verbose = TRUE)

\S4method{hits}{partitionBundle}(.Object, query, pAttribute = "word",
  size = TRUE, freq = FALSE, mc = FALSE, progress = FALSE,
  verbose = TRUE)

\S4method{sample}{hits}(x, size)
}
\arguments{
\item{.Object}{a character, partition or partitionBundle object}

\item{...}{further parameters}

\item{query}{character vector}

\item{sAttribute}{s-attributes}

\item{pAttribute}{p-attribute (will be passed into cpos)}

\item{size}{logical}

\item{freq}{locial}

\item{mc}{logical}

\item{verbose}{logical}

\item{progress}{logical}

\item{x}{a hits object}
}
\description{
A potentially useful worker.
}
\section{Slots}{

\describe{
\item{\code{dt}}{a \code{"data.table"}}

\item{\code{corpus}}{a \code{"character"} vector}

\item{\code{query}}{Object of class \code{"character"}}
}}

