% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/hits.R
\docType{class}
\name{hits_class}
\alias{hits_class}
\alias{hits-class}
\alias{sample,hits-method}
\title{Hits class.}
\usage{
\S4method{sample}{hits}(x, size)
}
\arguments{
\item{x}{A \code{hits} object.}

\item{size}{A non-negative integer giving the number of items to choose.}
}
\description{
Hits class.
}
\section{Slots}{

\describe{
\item{\code{stat}}{a \code{"data.table"}}

\item{\code{corpus}}{a \code{"character"} vector}

\item{\code{query}}{Object of class \code{"character"}}

\item{\code{p_attribute}}{p-attribute that has been queried}

\item{\code{encoding}}{encoding of the corpus}

\item{\code{name}}{name of the object}
}}

