% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/html-method.R
\docType{methods}
\name{html,partition-method}
\alias{html}
\alias{html,partition-method}
\alias{html,partitionCluster-method}
\alias{html-method}
\alias{show,html-method}
\title{Convert partition/partition Cluster into html}
\usage{
\S4method{html}{partition}(object, meta = NULL, type = "debate")
}
\arguments{
\item{object}{a partition object}

\item{meta}{metadata for output}

\item{type}{the type of html to be generated}
}
\description{
A partition is converted into a html file and displayed in a browser. If a partitionCluster
is provided, the browser will open several windows.
}

