\docType{class}
\name{keyness-class}
\alias{keyness,summary-method}
\alias{keyness-class}
\title{S4 class for comparing corpora}
\description{
to keep results from a keyness analysis
}
\section{Objects from the class}{
  keyness objects are returned by the function call
  \code{keyness}
}

\section{Slots}{
  \describe{ \item{\code{corpus}:}{Object of class
  \code{"character"} ~~ } \item{\code{pattribute}:}{Object
  of class \code{"character"} ~~ }
  \item{\code{stat}:}{Object of class \code{"data.frame"}
  ~~ } }
}
\author{
Andreas Blaette
}

