% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keyness-method.R
\docType{methods}
\name{keyness,partition-method}
\alias{keyness,collocations-method}
\alias{keyness,missing-method}
\alias{keyness,partition-method}
\alias{keyness,partitionCluster-method}
\title{compute chi-square values for tokens in a corpus using a reference corpus}
\usage{
\S4method{keyness}{partition}(x, y, pAttribute = NULL, minFrequency = 0,
  included = FALSE, method = "chiSquare", digits = 2, verbose = TRUE)

\S4method{keyness}{partitionCluster}(x, y, pAttribute = NULL,
  minFrequency = 1, included = FALSE, method = "chiSquare",
  verbose = TRUE, mc = TRUE, progress = FALSE)

\S4method{keyness}{collocations}(x, y, minFrequency = 0, included = FALSE,
  method = "ll", digits = 2, mc = TRUE, verbose = TRUE)

\S4method{keyness}{missing}()
}
\arguments{
\item{x}{a partition or partitionCluster object}

\item{y}{a partition object, it is assumed that the coi is a subcorpus of
ref}

\item{pAttribute}{The P-Attribute that will be counted (usually either
'word' or 'lemma')}

\item{minFrequency}{the minimum frequency of collocates}

\item{included}{TRUE if coi is part of ref, defaults to FALSE}

\item{method}{the statistical test to apply (chisquare or log likelihood)}

\item{digits}{numeric}

\item{verbose}{defaults to TRUE}

\item{progress}{logical}
}
\value{
The function returns a data frame with the following structure:
- absolute frequencies in the first row
- ...
}
\description{
Pearson's chi-squared test is calculated to measure the keyness of a token
in a corpus of interest (COI). A reference corpus is required for the
computation
}
\details{
If pos.filter is supplied, the most propable pos-Tags will be added to the
statistical table and the table will be filtered. This may slow down the
procedure considerably.
}
\author{
Andreas Blaette
}
\references{
Manning / Schuetze ...
}

