\docType{class}
\name{kwic-class}
\alias{kwic-class}
\title{class for kwic output}
\description{
information for kwic output
}
\section{Slots}{
  \describe{ \item{\code{cpos}:}{Object of class
  \code{"list"} corpus positions }
  \item{\code{word}:}{Object of class \code{"list"} to be
  explained } }
}

