% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kwic_class.R, R/kwic_methods.R, R/view_method.R
\docType{class}
\name{kwic-class}
\alias{[,kwic,ANY,ANY,ANY-method}
\alias{[,kwic-method}
\alias{as.data.frame,kwic-method}
\alias{kwic-class}
\alias{view,kwic-method}
\title{kwic (S4 class)}
\usage{
\S4method{as.data.frame}{kwic}(x)

\S4method{view}{kwic}(.Object)
}
\arguments{
\item{x}{a kwic-class object}

\item{.Object}{a kwic object}
}
\description{
S4 class for organizing information for concordance output
}
\section{Slots}{

\describe{
\item{\code{metadata}}{Object of class \code{"character"} keeping the sAttributes of the metadata that are to be displayed}

\item{\code{leftContext}}{words to the left}

\item{\code{rightContext}}{words to the right}

\item{\code{table}}{Object of class \code{"data.frame"} a table with the relevant information for kwic output}

\item{\code{collocate}}{Object of class \code{"character"} collocate, if applicable}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}
}}
\section{Methods}{

  \describe{
   \item{[}{indexing for seeing only some concordances}
   \item{show}{get kwic output}
  }
}

