% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kwic_method.R
\docType{methods}
\name{kwic}
\alias{kwic}
\alias{kwic,context-method}
\alias{kwic,partition-method}
\alias{show,kwic-method}
\title{KWIC output}
\usage{
kwic(object, ...)

\S4method{kwic}{context}(object, meta = NULL, neighbor = c())

\S4method{kwic}{partition}(object, query, leftContext = NULL,
  rightContext = NULL, meta = NULL, pAttribute = "word", neighbor = c(),
  verbose = TRUE)
}
\arguments{
\item{object}{a partition object}

\item{...}{further parameters to be passed}

\item{meta}{metainformation to display}

\item{neighbor}{only show kwic if a certain word is present}

\item{query}{what to look up}

\item{leftContext}{to the left}

\item{rightContext}{to the right}

\item{pAttribute}{typically 'word' or 'lemma'}

\item{verbose}{whether to be talkative}
}
\description{
Prepare and show 'keyword in context' (KWIC). The same result can be achieved by
applying the kwich method on either a partition or a context object.
}
\examples{
bt <- partition("PLPRTXT", def=list(text_date=".*"), regex=TRUE)
foo <- kwic(bt, "Integration")
foo <- kwic(bt, "Integration", leftContext=20, rightContext=20, meta=c("text_date", "text_name", "text_party"))
}

