% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kwic_method.R
\docType{methods}
\name{kwic}
\alias{kwic}
\alias{kwic,context-method}
\alias{kwic,missing-method}
\alias{kwic,partition-method}
\title{KWIC output}
\usage{
kwic(.Object, ...)

\S4method{kwic}{context}(.Object, meta = NULL, neighbor = NULL)

\S4method{kwic}{partition}(.Object, query, left = NULL, right = NULL,
  meta = NULL, pAttribute = "word", neighbor = NULL, verbose = TRUE)

\S4method{kwic}{missing}(.Object, ...)
}
\arguments{
\item{.Object}{a partition object}

\item{...}{further parameters to be passed}

\item{meta}{metainformation to display}

\item{neighbor}{only show kwic if a certain word is present}

\item{query}{what to look up}

\item{left}{to the left}

\item{right}{to the right}

\item{pAttribute}{typically 'word' or 'lemma'}

\item{verbose}{whether to be talkative}
}
\description{
Prepare and show 'keyword in context' (KWIC). The same result can be achieved by 
applying the kwich method on either a partition or a context object.
}
\examples{
bt <- partition("PLPRBTTXT", def=list(text_date=".*"), regex=TRUE)
foo <- kwic(bt, "Integration")
foo <- kwic(bt, "Integration", left=20, right=20, meta=c("text_date", "text_name", "text_party")) 
}

