% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_method.R
\name{label}
\alias{label}
\alias{label<-}
\title{Assign and get labels.}
\usage{
label(x)

label(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{length (character vector, length 1)}
}
\description{
Assign and get labels.
}
