% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label.R
\docType{methods}
\name{label}
\alias{label}
\alias{label<-}
\alias{label,kwic-method}
\title{Assign and get labels.}
\usage{
label(x, ...)

label(x) <- value

\S4method{label}{kwic}(x, n = NULL)
}
\arguments{
\item{x}{object}

\item{...}{further parameters}

\item{value}{length (character vector, length 1)}

\item{n}{label index}
}
\description{
Assign and get labels.
}
