% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels.R
\docType{class}
\name{Labels-class}
\alias{Labels-class}
\alias{Labels}
\title{Labels class.}
\arguments{
\item{n}{length of character vector in field labels}

\item{choices}{choices to be assigned to field choices}

\item{expandable}{whether choices are expandable}
}
\description{
Labels class.
}
\section{Fields}{

\describe{
\item{\code{labels}}{character vector with labels; if logical or numeric labels
are intended, assign them as character vector anyway}

\item{\code{choices}}{character vector, a list of choices for labels}

\item{\code{expandable}}{whether choices may be expanded (logical)}
}}


