% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mail_method.R
\docType{methods}
\name{mail}
\alias{mail}
\alias{mail,comp-method}
\alias{mail,context-method}
\alias{mail,data.frame-method}
\alias{mail,dispersion-method}
\alias{mail,kwic-method}
\alias{mail,partition-method}
\alias{mail-method}
\title{Mail result.}
\usage{
mail(object, ...)

\S4method{mail}{partition}(object, to = NULL,
  filename = "drillerExport.html", what = "html")

\S4method{mail}{context}(object, to = NULL, nrow = NULL,
  fileFormat = c("csv", "xlsx"))

\S4method{mail}{comp}(object, to = NULL, nrow = NULL,
  fileFormat = c("csv", "xlsx"))

\S4method{mail}{kwic}(object, to = NULL, nrow = NULL,
  fileFormat = c("csv", "xlsx"))

\S4method{mail}{dispersion}(object, to = NULL, nrow = NULL,
  fileFormat = c("csv", "xlsx"))

\S4method{mail}{data.frame}(object, to = NULL, nrow = NULL,
  fileFormat = c("csv", "xlsx"))
}
\arguments{
\item{object}{object to deliver}

\item{...}{further parameters}

\item{to}{the receiver of the mail message}

\item{filename}{filename}

\item{what}{what to send (defaults to "html")}

\item{nrow}{the number of rows of the table (if NULL, the whole table will be sent)}

\item{fileFormat}{csv or xlsx, or both}
}
\description{
Mail a result (to yourself).
}

