\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,partition-method}
\alias{meta-method}
\title{get values of metadata}
\usage{
\S4method{meta}{partition}(object, sAttribute)
}
\arguments{
  \item{object}{a partition object}

  \item{sAttribute}{the s-attribute to be looked up}
}
\value{
a character vector
}
\description{
Returns the values that an sAttribute takes within a
partition.
}

