% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams_method.R
\docType{class}
\name{ngrams-class}
\alias{ngrams}
\alias{ngrams,partition-method}
\alias{ngrams-class}
\title{get ngrams}
\usage{
ngrams(.Object, ...)

\S4method{ngrams}{partition}(.Object, n = 2, pAttribute = "word")
}
\arguments{
\item{.Object}{object of class \code{partition}}

\item{...}{further parameters}

\item{n}{number of tokens}

\item{pAttribute}{p-attribute}
}
\description{
get ngrams
}

