% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams_method.R
\name{ngrams}
\alias{ngrams}
\title{get ngrams}
\usage{
ngrams(.Object, ...)
}
\arguments{
\item{.Object}{object of class \code{partition}}

\item{n}{number of tokens}

\item{pAttribute}{p-attribute}
}
\description{
get ngrams
}

