% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition_method.R
\docType{methods}
\name{partition,character-method}
\alias{partition}
\alias{partition,character-method}
\alias{partition,list-method}
\alias{partition,missing-method}
\alias{partition,partition-method}
\title{Initialize a partition}
\usage{
\S4method{partition}{character}(object, def = NULL, name = c(""),
  encoding = NULL, pAttribute = NULL, meta = NULL, regex = FALSE,
  xml = "flat", id2str = TRUE, type = NULL, mc = FALSE,
  verbose = TRUE)

\S4method{partition}{list}(object, name = c(""), encoding = NULL,
  pAttribute = NULL, meta = NULL, regex = FALSE, xml = "flat",
  id2str = TRUE, type = NULL, mc = FALSE, verbose = TRUE)

\S4method{partition}{missing}()

\S4method{partition}{partition}(object, def, name = c(""), regex = FALSE,
  pAttribute = NULL, id2str = TRUE, type = NULL, verbose = TRUE,
  mc = FALSE)
}
\arguments{
\item{object}{character-vector - the CWB-corpus to be used}

\item{def}{list consisting of a set of character vectors (see
details)}

\item{name}{name of the new partition, defaults to "noName"}

\item{encoding}{encoding of the corpus (typically "LATIN1 or "(UTF-8)), if NULL, the encoding provided in the registry file of the corpus (charset="...") will be used b}

\item{pAttribute}{the pAttribute(s) for which term frequencies shall be retrieved}

\item{meta}{a character vector}

\item{regex}{logical, whether strucs will be filtered by applying a regex (via grep)}

\item{xml}{either 'flat' (default) or 'nested'}

\item{id2str}{whether to turn token ids to strings (set FALSE to minimize object.size / memory consumption)}

\item{type}{character vector (length 1) specifying the type of corpus / partition (e.g. "plpr")}

\item{mc}{whether to use multicore (for counting terms)}

\item{verbose}{logical, defaults to TRUE}

\item{value}{a character string that will be the name of the partition}

\item{from}{from}

\item{to}{to}

\item{x}{CHECK}
}
\value{
An object of the S4 class 'partition'
}
\description{
Set up an object of the partition class. Usually, this will include frequency lists.
}
\details{
The function sets up a partition based on a list of s-attributes with respective values.
The s-attributes defining the partition are a list, e.g. list(text_type="speech", text_year="2013").
The values of the list may contain regular expressions, but then the length always needs to be 1,
and the method needs to be set to "grep". Alternatively, the method can be set to "in", then
the length of the list may be > 1.
For the s-attributes, list elements may be character vectors with a length > 2.
However, the initialization of a partition object is considerably slower, if
you supply a >2 vector. The alternative being creating a set of partitions with
\code{partition.bundle} and then merging the partitions, this may still be ok.
For s-attributes, regular expressions can be used. Please note that for R grep,
double backlashes have to be used.
For some purposes (c.g. computation of crosstabulations), term frequencies will
not be needed in the setup of the partition object. In that case pAttribute=FALSE and
metadata=FALSE will speed up the initialization of the object a lot.
A date range with a specific start and end date can be specified by providing dateRange.
Note that the sequence of the s-Attributes will matter. Things will speed up if you start 
with the conditions narrowing down the corpus the most.
}
\examples{
use(polmineR.sampleCorpus)
spd <- partition("PLPRBTTXT", def=list(text_party="SPD", text_type="speech"))
kauder <- partition("PLPRBTTXT", def=list(text_name="Volker Kauder"), pAttribute=c("word"))
merkel <- partition("PLPRBTTXT", list(text_name=".*Merkel"), pAttribute="word", regex=TRUE)
sAttributes(merkel, "text_date")
sAttributes(merkel, "text_name")
merkel <- partition("PLPRBTTXT", list(text_name="Angela Dorothea Merkel", text_date="2009-11-10", text_type="speech"), pAttribute="word")
merkel <- subset(merkel, !word \%in\% punctuation)
merkel <- subset(merkel, !word \%in\% tm::stopwords("de"))
}
\author{
Andreas Blaette
}

