% Generated by roxygen2 (4.0.1): do not edit by hand
\name{partition}
\alias{partition}
\title{Initialize a partition}
\usage{
partition(corpus, def, label = c(""), encoding = NULL, tf = c("word",
  "lemma"), meta = NULL, method = "grep", xml = "flat", mc = FALSE,
  verbose = TRUE)
}
\arguments{
\item{corpus}{the CWB-corpus to be used}

\item{def}{list consisting of a set of character vectors (see
details)}

\item{label}{label of the new partition, defaults to "noLabel"}

\item{encoding}{encoding of the corpus (typically "LATIN1 or "(UTF-8)), if NULL, the encoding provided in the registry file of the corpus (charset="...") will be used b}

\item{tf}{the pAttributes for which term frequencies shall be retrieved}

\item{meta}{a character vector}

\item{method}{either 'grep' or 'in' to specify the filtering method to get relevant strucs}

\item{xml}{either 'flat' (default) or 'nested'}

\item{mc}{whether to use multicore (for tf lists)}

\item{verbose}{logical, defaults to TRUE}
}
\value{
An object of the S4 class 'partition'
}
\description{
Set up an object of the partition class. Usually, this will include frequency lists.
}
\details{
The function sets up a partition based on a list of s-attributes with respective values.
The s-attributes defining the partition are a list, e.g. list(text_type="speech", text_year="2013").
The values of the list may contain regular expressions, but then the length always needs to be 1,
and the method needs to be set to "grep". Alternatively, the method can be set to "in", then
the length of the list may be > 1.
For the s-attributes, list elements may be character vectors with a length > 2.
However, the initialization of a partition object is considerably slower, if
you supply a >2 vector. The alternative being creating a set of partitions with
\code{partition.cluster} and then merging the partitions, this may still be ok.
For s-attributes, regular expressions can be used. Please note that for R grep,
double backlashes have to be used.
For some purposes (c.g. computation of crosstabulations), term frequencies will
not be needed in the setup of the partition object. In that case tf=FALSE and
metadata=FALSE will speed up the initialization of the object a lot.
A date range with a specific start and end date can be specified by providing dateRange.
Note that the sequence of the s-Attributes will matter. Things will speed up if you start
with the conditions narrowing down the corpus the most.
}
\examples{
\dontrun{
spd <- partition("PLPRBTTXT", def=list(text_party="SPD", text_type="speech"))
}
}
\author{
Andreas Blaette
}

