% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/html_method.R, R/partitionBundle_class.R, R/partitionBundle_methods.R, R/sAttributes_method.R
\docType{methods}
\name{html,partitionBundle-method}
\alias{+,partitionBundle,ANY-method}
\alias{+,partitionBundle,partition-method}
\alias{+,partitionBundle,partitionBundle-method}
\alias{+,partitionBundle-method}
\alias{[,partitionBundle,ANY,ANY,ANY-method}
\alias{[,partitionBundle-method}
\alias{[[,partitionBundle,ANY-method}
\alias{[[,partitionBundle-method}
\alias{as.DocumentTermMatrix}
\alias{as.matrix,partitionBundle-method}
\alias{as.partitionBundle,list-method}
\alias{as.sparseMatrix}
\alias{as.sparseMatrix,TermDocumentMatrix-method}
\alias{as.sparseMatrix,partitionBundle-Method}
\alias{as.sparseMatrix,partitionBundle-method}
\alias{barplot,partitionBundle-method}
\alias{html,partitionBundle-method}
\alias{length,partitionBundle-method}
\alias{merge,partitionBundle-method}
\alias{names,partitionBundle-method}
\alias{names<-,partitionBundle,character-method}
\alias{partitionBundle-class}
\alias{sAttributes,partitionBundle-method}
\alias{show,partitionBundle-method}
\alias{summary,partitionBundle-method}
\alias{unique,partitionBundle-method}
\title{partitionBundle class}
\usage{
\S4method{html}{partitionBundle}(object, filename = c(), type = "debate")

\S4method{show}{partitionBundle}(object)

\S4method{summary}{partitionBundle}(object)

\S4method{[[}{partitionBundle,ANY}(x, i)

\S4method{[}{partitionBundle,ANY,ANY,ANY}(x, i)

\S4method{names}{partitionBundle}(x)

\S4method{names}{partitionBundle,character}(x) <- value

\S4method{unique}{partitionBundle}(x)

\S4method{length}{partitionBundle}(x)

\S4method{sAttributes}{partitionBundle}(object, sAttribute)
}
\arguments{
\item{object}{a partitionBundle object}

\item{filename}{filename}

\item{type}{type}

\item{x}{a partitionBundle object}

\item{i}{integer index}

\item{value}{the value}

\item{sAttribute}{the s-attribute to use}

\item{pAttribute}{the p-attribute}
}
\description{
A bundle of partition objects.
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"} the partitions making up the bundle}

\item{\code{corpus}}{Object of class \code{"character"} the CWB corpus the partition is based on}

\item{\code{sAttributesFixed}}{Object of class \code{"list"} fixed sAttributes}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{explanation}}{Object of class \code{"character"} an explanation of the partition}

\item{\code{xml}}{Object of class \code{"character"} whether the xml is flat or nested}

\item{\code{call}}{Object of class \code{"character"} the call that generated the partitionBundle}
}}
\author{
Andreas Blaette
}

