% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionBundle_class.R, R/enrich_method.R,
%   R/partitionBundle_method.R, R/sAttributes_method.R
\docType{class}
\name{partitionBundle-class}
\alias{partitionBundle-class}
\alias{[,partitionBundle-method}
\alias{[[,partitionBundle-method}
\alias{as.matrix,partitionBundle-method}
\alias{merge,partitionBundle-method}
\alias{+,partitionBundle-method}
\alias{names,partitionBundle-method}
\alias{summary,partitionBundle-method}
\alias{+,partitionBundle,ANY-method}
\alias{[,partitionBundle,ANY,ANY,ANY-method}
\alias{+,partitionBundle,partition-method}
\alias{+,partitionBundle,partitionBundle-method}
\alias{as.partitionBundle,list-method}
\alias{barplot,partitionBundle-method}
\alias{show,partitionBundle-method}
\alias{summary,partitionBundle-method}
\alias{[,partitionBundle,ANY,ANY,ANY-method}
\alias{barplot,partitionBundle-method}
\alias{enrich,partitionBundle-method}
\alias{as.partitionBundle,list-method}
\alias{partitionBundle,environment-method}
\alias{sAttributes,partitionBundle-method}
\title{Bundle of partitions (partitionBundle class).}
\usage{
\S4method{show}{partitionBundle}(object)

\S4method{summary}{partitionBundle}(object)

\S4method{[}{partitionBundle,ANY,ANY,ANY}(x, i)

\S4method{barplot}{partitionBundle}(height, ...)

\S4method{enrich}{partitionBundle}(.Object, mc = FALSE, progress = TRUE,
  verbose = FALSE, ...)

\S4method{as.partitionBundle}{list}(.Object, ...)

\S4method{partitionBundle}{environment}(.Object)

\S4method{sAttributes}{partitionBundle}(.Object, sAttribute)
}
\arguments{
\item{object}{a partitionBundle object}

\item{x}{a partitionBundle object}

\item{i}{integer index}

\item{height}{height}

\item{...}{further parameters}

\item{.Object}{a partitionBundle object}

\item{mc}{logical or, if numeric, providing the number of cores}

\item{progress}{logical}

\item{verbose}{logical}

\item{sAttribute}{the s-attribute to use}
}
\description{
Class and methods to manage bundles of partitions.
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"} the partitions making up the bundle}

\item{\code{corpus}}{Object of class \code{"character"} the CWB corpus the partition is based on}

\item{\code{sAttributesFixed}}{Object of class \code{"list"} fixed sAttributes}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{explanation}}{Object of class \code{"character"} an explanation of the partition}

\item{\code{xml}}{Object of class \code{"character"} whether the xml is flat or nested}

\item{\code{call}}{Object of class \code{"character"} the call that generated the partitionBundle}
}}

\author{
Andreas Blaette
}
