% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitionBundle_method.R
\docType{methods}
\name{partitionBundle}
\alias{partitionBundle}
\alias{partitionBundle,character-method}
\alias{partitionBundle,list-method}
\title{Generate a list of partitions}
\usage{
\S4method{partitionBundle}{character}(object, def, var, prefix = c(""),
  encoding = NULL, pAttribute = NULL, meta = NULL, regex = FALSE,
  xml = "flat", id2str = TRUE, type = NULL, mc = NULL, verbose = TRUE)

\S4method{partitionBundle}{list}(object, var, prefix = c(""),
  encoding = NULL, pAttribute = NULL, meta = NULL, regex = FALSE,
  xml = "flat", id2str = TRUE, mc = FALSE, verbose = TRUE)
}
\arguments{
\item{object}{character string, the CWB corpus to be used}

\item{def}{a list with the definition of a partition that shall be prepared}

\item{var}{list indicating the s-attribute to be variabel}

\item{prefix}{a character vector that will serve as a prefix for partition names}

\item{encoding}{encoding of the corpus, if not provided, encoding provided in the registry file will be used}

\item{pAttribute}{the pAttributes for which term frequencies shall be retrieved}

\item{meta}{a character vector}

\item{regex}{logical}

\item{xml}{either 'flat' (default) or 'nested'}

\item{id2str}{whether to turn token ids to strings (set FALSE to minimize object.size / memory consumption)}

\item{type}{the type of the partition objects}

\item{mc}{logical, whether to use multicore parallelization}

\item{verbose}{logical, whether to provide progress information}
}
\value{
a S4 class 'partitionBundle', which is a list with partition objects
}
\description{
A list of partition objects with fixed s-attributes and one variable
s-attribute is generated
}
\author{
Andreas Blaette
}

