% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/partitionBundle_method.R
\docType{methods}
\name{partitionBundle}
\alias{partitionBundle}
\alias{partitionBundle,character-method}
\alias{partitionBundle,list-method}
\title{Generate a list of partitions}
\usage{
partitionBundle(object, ...)

\S4method{partitionBundle}{character}(object, def, var, prefix = c(""), ...,
  mc = NULL, verbose = TRUE)

\S4method{partitionBundle}{list}(object, var, prefix = c(""), ...)
}
\arguments{
\item{object}{character string, the CWB corpus to be used}

\item{...}{further parameters that will be pased into the partition-method called to generate the partitions in the bundle}

\item{def}{a list with the definition of a partition that shall be prepared}

\item{var}{a list that indicates the s-attribute to be variable and that
provides a character string of values (e.g. var=list(text_year=c("2005", "2006"))}

\item{prefix}{a character vector that will serve as a prefix for partition names}

\item{mc}{logical, whether to use multicore parallelization}

\item{verbose}{logical, whether to provide progress information}
}
\value{
a S4 class 'partitionBundle', which is a list with partition objects
}
\description{
A list of partition objects with fixed s-attributes and one variable
s-attribute is generated
}
\author{
Andreas Blaette
}
\seealso{
\code{\link{partition}} and \code{\link{bundle-class}}
}

