% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/partitionBundle-method.R
\docType{methods}
\name{partitionBundle}
\alias{partitionBundle}
\alias{partitionBundle,character-method}
\alias{partitionBundle,list-method}
\alias{partitionBundle,missing-method}
\title{Generate a list of partitions}
\usage{
\S4method{partitionBundle}{character}(object, def, var, prefix = c(""),
  encoding = NULL, tf = c("word", "lemma"), meta = NULL,
  method = "grep", xml = "flat", id2str = TRUE, type = NULL,
  mc = NULL, verbose = TRUE)

\S4method{partitionBundle}{list}(object, var, prefix = c(""),
  encoding = NULL, tf = c("word", "lemma"), meta = NULL,
  method = "grep", xml = "flat", id2str = TRUE, mc = FALSE,
  verbose = TRUE)

\S4method{partitionBundle}{missing}()
}
\arguments{
\item{object}{character string, the CWB corpus to be used}

\item{def}{a list with the definition of a partition that shall be prepared}

\item{var}{list indicating the s-attribute to be variabel}

\item{prefix}{a character vector that will serve as a prefix for partition labels}

\item{encoding}{encoding of the corpus, if not provided, encoding provided in the registry file will be used}

\item{tf}{the pAttributes for which term frequencies shall be retrieved}

\item{meta}{a character vector}

\item{method}{either 'grep' or 'in'}

\item{xml}{either 'flat' (default) or 'nested'}

\item{id2str}{whether to turn token ids to strings (set FALSE to minimize object.size / memory consumption)}

\item{type}{the type of the partition objects}

\item{mc}{logical, whether to use multicore parallelization}

\item{verbose}{logical, whether to provide progress information}
}
\value{
a S4 class 'partitionBundle', which is a list with partition objects
}
\description{
A list of partition objects with fixed s-attributes and one variable
s-attribute is generated
}
\details{
If var is list(text_date=NULL) for instance, all values for the provided s-attribute
in the partition
defined by sAttributesFixed will be retrived and used for defining the
partitions.
While generally S4 methods are used in the driller package, the return is a S3 method.
The reasons is that the number of partitions kept in the bundle is not known before the initialization.
Setting multicore to TRUE will speed up things. Error handling is less benevolent, risk of overheating, no verbose output.
}
\author{
Andreas Blaette
}

