% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/addPos-method.R, R/html-method.R, R/partitionCluster-class.R, R/partitionCluster-methods.R, R/sAttributes-method.R
\docType{methods}
\name{addPos,partitionCluster-method}
\alias{+,partitionCluster,ANY-method}
\alias{+,partitionCluster,partition-method}
\alias{+,partitionCluster,partitionCluster-method}
\alias{+,partitionCluster-method}
\alias{[,partitionCluster,ANY,ANY,ANY-method}
\alias{[,partitionCluster-method}
\alias{[[,partitionCluster,ANY-method}
\alias{[[,partitionCluster-method}
\alias{addPos,partitionCluster-method}
\alias{as.DocumentTermMatrix}
\alias{as.matrix,partitionCluster-method}
\alias{as.partitionCluster,list-method}
\alias{as.sparseMatrix}
\alias{as.sparseMatrix,TermDocumentMatrix-method}
\alias{as.sparseMatrix,partitionCluster-Method}
\alias{as.sparseMatrix,partitionCluster-method}
\alias{barplot,partitionCluster-method}
\alias{html}
\alias{html,partition-method}
\alias{html,partitionCluster-method}
\alias{html-method}
\alias{label,partitionCluster-method}
\alias{label<-,partitionCluster,character-method}
\alias{length,partitionCluster-method}
\alias{merge,partitionCluster-method}
\alias{names,partitionCluster-method}
\alias{partitionCluster-class}
\alias{sAttributes,partitionCluster-method}
\alias{show,html-method}
\alias{show,partitionCluster-method}
\alias{summary,partitionCluster-method}
\alias{unique,partitionCluster-method}
\title{#}
\usage{
\S4method{addPos}{partitionCluster}(object, pAttribute)

\S4method{html}{partition}(object, meta = NULL)

\S4method{html}{partitionCluster}(object, filename = c(), type = "debate")

\S4method{show}{partitionCluster}(object)

\S4method{[[}{partitionCluster,ANY}(x, i)

\S4method{[}{partitionCluster,ANY,ANY,ANY}(x, i)

\S4method{+}{partitionCluster,partitionCluster}(e1, e2)

\S4method{+}{partitionCluster,partition}(e1, e2)

\S4method{label}{partitionCluster}(object)

\S4method{label}{partitionCluster,character}(object) <- value

\S4method{unique}{partitionCluster}(x)

\S4method{length}{partitionCluster}(x)

\S4method{sAttributes}{partitionCluster}(object, sAttribute)
}
\arguments{
\item{object}{a partitionCluster object}

\item{pAttribute}{the p-attribute}

\item{meta}{metadata for output}

\item{filename}{filename}

\item{type}{type}

\item{x}{a partitionCluster object}

\item{i}{integer index}

\item{e1}{partition 1}

\item{e2}{partition 2}

\item{value}{the value}

\item{sAttribute}{the s-attribute to use}
}
\description{
#

A cluster of partition objects.
}
\section{Slots}{

\describe{
\item{\code{objects}}{Object of class \code{"list"} the partitions making up the cluster}

\item{\code{corpus}}{Object of class \code{"character"} the CWB corpus the partition is based on}

\item{\code{sAttributesFixed}}{Object of class \code{"list"} fixed sAttributes}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{explanation}}{Object of class \code{"character"} an explanation of the partition}

\item{\code{xml}}{Object of class \code{"character"} whether the xml is flat or nested}

\item{\code{call}}{Object of class \code{"character"} the call that generated the partitionCluster}
}}
\author{
Andreas Blaette
}

