% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed.R
\docType{data}
\name{polmineR-defunct}
\alias{polmineR-defunct}
\alias{browse}
\alias{Corpus}
\title{Defunct methods and functions.}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
browse(...)

Corpus
}
\arguments{
\item{...}{Any arguments that may be passed into the defunct function/method.}
}
\description{
Methods and functions not in use any more or that have been
  superseded by renamed functions.
}
\details{
The RefClass `Corpus` has been superseded by the S4 `corpus`-class
  and adherent methods.
}
\keyword{datasets}
