% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polmineR-defunct.R, R/renamed.R
\name{polmineR-defunct}
\alias{polmineR-defunct}
\alias{store}
\alias{mail}
\alias{browse}
\title{Defunct functionality}
\usage{
store(...)

mail(...)

browse(...)
}
\arguments{
\item{...}{Any arguments that may be passed into the defunct function/method.}
}
\description{
Methods and functions not in use any more or that have been
superseded by renamed functions.
}
