% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polmineR.R, R/info.R
\name{polmineR-generics}
\alias{polmineR-generics}
\alias{get_info}
\alias{show_info}
\title{Generic methods defined in the polmineR package}
\usage{
get_info(x)

show_info(x)
}
\arguments{
\item{x}{An S4 class object.}
}
\description{
This documentation object gives an overview over the generic
  methods defined in the polmineR package that have no individual man page
  but are documented directly with the classes they are defined for.
}
