% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polmineR.R, R/browse.R
\name{polmineR-generics}
\alias{polmineR-generics}
\alias{name}
\alias{name<-}
\alias{browse<-}
\title{generic methods defined in the polmineR-package}
\usage{
name(x)

name(x) <- value

browse(x) <- value
}
\arguments{
\item{x}{object}

\item{value}{value}
}
\description{
The package defines a set of generic functions. This doc file
only provides an overview. Please consult the documentation
of the classes to learn which methods can be applied to a class
of a certain type.
}
