% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_method.R
\docType{methods}
\name{read}
\alias{read}
\alias{read,partition,list-method}
\alias{read,partitionBundle,list-method}
\title{Return to the original text and read}
\usage{
read(.Object, highlight, ...)

\S4method{read}{partition,list}(.Object, highlight = list(), cqp = FALSE,
  meta = NULL, verbose = TRUE, ...)

\S4method{read}{partitionBundle,list}(.Object, highlight = list(),
  cqp = FALSE, ...)
}
\arguments{
\item{.Object}{an object to be read (\code{"partition" or "partitionBundle"})}

\item{...}{further parameters passed into read}

\item{cqp}{a list of character vectors with regular expressions to
highlight relevant terms or expressions; the names of the list provide the
colors}

\item{meta}{a character vector supplying s-attributes for the metainformation
to be printed, if not stated explicitly, session settings will be used}

\item{regex}{a list of character vectors with regular expressions to
highlight relevant terms or expressions; the names of the list provide the
colors (see examples)}
}
\description{
Generate text (i.e. html) and read it in the viewer pane of RStudio. If called on
a \code{"partitionBundle"}-object, skip through the partitions contained in the
bundle.
}
\examples{
use("polmineR.sampleCorpus")
merkel <- partition(
  "PLPRBTTXT",
  list(text_date="2009-11-10", text_name="Angela Dorothea Merkel"),
  type="plpr"
)
read(merkel, meta=c("text_name", "text_date"))
read(
  merkel,
  highlight=list(yellow=c("Deutschland", "Bundesrepublik"), lightgreen="Regierung")
  )
\dontrun{
all <- partition("PLPRBTTXT", list(text_id=".*"), regex=TRUE, type="plpr")
speeches <- as.speeches(all, sAttributeDates="text_date", sAttributeNames="text_name", gap=500)
read(speeches)

migVocab <- count(speeches, query=c("Migration", "Integration", "Zuwanderung"))
read(migVocab, speeches, col="Integration")
}
}
\seealso{
For concordances / a keword-in-context display, see \code{\link{kwic}}.
}

