% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_method.R
\docType{methods}
\name{read}
\alias{read}
\alias{read,plprPartition-method}
\alias{read,pressPartition-method}
\title{read an object}
\usage{
read(.Object, ...)

\S4method{read}{plprPartition}(.Object, meta = c("text_name", "text_party",
  "text_date"))

\S4method{read}{pressPartition}(.Object, meta = c("text_newspaper",
  "text_date"))
}
\arguments{
\item{.Object}{an object to be read}

\item{...}{further parameters}
}
\description{
A html page will be generated and displayed in the viewer pane of RStudio.
}

