% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{registry}
\alias{registry}
\title{Get session registry directory.}
\usage{
registry()
}
\description{
The \code{polmineR} package uses a subdirectory of the per-session temporary directory
as a (temporary) registry. The \code{registry} function will return the path 
to this directory.
}
\examples{
registry()
}
