% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renamed.R
\docType{methods}
\name{renamed}
\alias{renamed}
\alias{sAttributes}
\alias{pAttributes}
\alias{getTokenStream}
\alias{getTerms}
\alias{getEncoding}
\alias{partitionBundle}
\alias{as.partitionBundle}
\alias{setTemplate}
\alias{getTemplate}
\alias{corpus,textstat-method}
\alias{corpus,bundle-method}
\alias{corpus,kwic-method}
\title{Renamed Functions}
\usage{
sAttributes(...)

pAttributes(...)

getTokenStream(...)

getTerms(...)

getEncoding(...)

partitionBundle(...)

as.partitionBundle(...)

setTemplate(...)

getTemplate(...)

\S4method{corpus}{textstat}(.Object)

\S4method{corpus}{bundle}(.Object)

\S4method{corpus}{kwic}(.Object)
}
\arguments{
\item{...}{argument that are passed to the renamed function}

\item{.Object}{A \code{kwic} object.}
}
\description{
These functions have been renamed in order to have a consistent coding style
that follows the snake_case convention. The "old" function still work to
maintain backwards compatiblity.
}
