% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sAttributes_method.R
\docType{methods}
\name{sAttributes,character-method}
\alias{sAttributes}
\alias{sAttributes,character-method}
\alias{sAttributes,partition-method}
\title{Print S-Attributes in a partition or corpus}
\usage{
\S4method{sAttributes}{character}(object, sAttribute = NULL, unique = TRUE,
  regex = NULL)

\S4method{sAttributes}{partition}(object, sAttribute = NULL)
}
\arguments{
\item{object}{either a partition or a character vector specifying a CWB corpus}

\item{sAttribute}{bla}

\item{unique}{logical, whether to return unique values only}

\item{regex}{apply a regex}
}
\value{
the S-Attributes are immediately printed
}
\description{
Convencience function - just to access the s-sttributes in a partition
quickly.
}

