% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/score_method.R
\docType{methods}
\name{score}
\alias{score}
\alias{score,partitionBundle-method}
\alias{score-method}
\title{score partitions in a bundle}
\usage{
\S4method{score}{partitionBundle}(object, token, method, pAttribute,
  rel = FALSE, tfMethod = FALSE)
}
\arguments{
\item{object}{a partitionBundle object}

\item{token}{a character vector supplying tokens to be looked up}

\item{method}{either "additive" or "vsm"}

\item{pAttribute}{the p-attribute the analysis is based on}

\item{rel}{logical, whether to use relative frequencies}

\item{tfMethod}{what method to use for term frequencies}
}
\value{
a named numeric vector (names of the partitions in the bundle)
}
\description{
Assign scores to the partitions in a partitionBundle object.
}

