% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{session-class}
\alias{corpus}
\alias{corpus,session-method}
\alias{corpus<-}
\alias{corpus<-,session,character-method}
\alias{createProject}
\alias{email}
\alias{email,session-method}
\alias{email<-}
\alias{email<-,session,character-method}
\alias{filterType}
\alias{filterType,session-method}
\alias{filterType<-}
\alias{filterType<-,session,character-method}
\alias{kwicMetadata}
\alias{kwicMetadata,session-method}
\alias{kwicMetadata<-}
\alias{kwicMetadata<-,session,character-method}
\alias{kwicNo}
\alias{kwicNo,session-method}
\alias{kwicNo<-}
\alias{kwicNo<-,session,numeric-method}
\alias{leftContext}
\alias{leftContext,session-method}
\alias{leftContext<-}
\alias{leftContext<-,session,numeric-method}
\alias{metadata}
\alias{metadata,session-method}
\alias{metadata<-}
\alias{metadata<-,session,character-method}
\alias{minFrequency}
\alias{minFrequency,session-method}
\alias{minFrequency<-}
\alias{minFrequency<-,session,numeric-method}
\alias{minSignificance}
\alias{minSignificance,session-method}
\alias{minSignificance<-}
\alias{minSignificance<-,session,numeric-method}
\alias{multicore}
\alias{multicore,session-method}
\alias{multicore<-}
\alias{multicore<-,session,logical-method}
\alias{pAttribute}
\alias{pAttribute,session-method}
\alias{pAttribute<-}
\alias{pAttribute<-,session,character-method}
\alias{posFilter}
\alias{posFilter,session-method}
\alias{posFilter<-}
\alias{posFilter<-,session,character-method}
\alias{project}
\alias{project,missing-method}
\alias{project,session-method}
\alias{project<-}
\alias{project<-,session,character-method}
\alias{rightContext}
\alias{rightContext,session-method}
\alias{rightContext<-}
\alias{rightContext<-,session,numeric-method}
\alias{session}
\alias{session-class}
\alias{show,session-method}
\alias{smtpPort}
\alias{smtpPort,session-method}
\alias{smtpPort<-}
\alias{smtpPort<-,session,character-method}
\alias{smtpServer}
\alias{smtpServer,session-method}
\alias{smtpServer<-}
\alias{smtpServer<-,session,character-method}
\title{session class}
\usage{
createProject(project)

\S4method{corpus}{session,character}(object) <- value

\S4method{corpus}{session}(object)

\S4method{project}{session,character}(object) <- value

\S4method{project}{session}(object)

\S4method{project}{missing}()

\S4method{pAttribute}{session}(object)

\S4method{pAttribute}{session,character}(object) <- value

\S4method{leftContext}{session}(x)

\S4method{leftContext}{session,numeric}(x) <- value

\S4method{rightContext}{session}(x)

\S4method{rightContext}{session,numeric}(x) <- value

\S4method{minSignificance}{session}(x)

\S4method{minSignificance}{session,numeric}(x) <- value

\S4method{minFrequency}{session}(x)

\S4method{minFrequency}{session,numeric}(x) <- value

\S4method{posFilter}{session}(x)

\S4method{posFilter}{session,character}(x) <- value

\S4method{filterType}{session}(x)

\S4method{filterType}{session,character}(x) <- value

\S4method{kwicMetadata}{session}(x)

\S4method{kwicMetadata}{session,character}(x) <- value

\S4method{kwicNo}{session}(x)

\S4method{kwicNo}{session,numeric}(x) <- value

\S4method{metadata}{session}(x)

\S4method{metadata}{session,character}(x) <- value

\S4method{multicore}{session}(x)

\S4method{multicore}{session,logical}(x) <- value

\S4method{smtpServer}{session}(x)

\S4method{smtpServer}{session,character}(x) <- value

\S4method{smtpPort}{session}(x)

\S4method{smtpPort}{session,character}(x) <- value

\S4method{email}{session}(x)

\S4method{email}{session,character}(x) <- value

\S4method{show}{session}(object)
}
\arguments{
\item{project}{the project}

\item{object}{a session class object}

\item{value}{character string to be assigned}

\item{x}{an object}
}
\description{
default settings for current session
}
\section{Slots}{

\describe{
\item{\code{project}}{currently active project}

\item{\code{corpus}}{corpus to use if none is provided}

\item{\code{pAttribute}}{default pAttribute}

\item{\code{leftContext}}{default left context}

\item{\code{rightContext}}{default right context}

\item{\code{minSignificance}}{default minimum significance}

\item{\code{minFrequency}}{default minimum frequency}

\item{\code{posFilter}}{default filter for POS}

\item{\code{filterType}}{default filter type}

\item{\code{kwicMetadata}}{default metadata displayed for kwic}

\item{\code{kwicNo}}{default number of kwic lines}

\item{\code{metadata}}{default metadata}

\item{\code{multicore}}{default multicore}

\item{\code{consoleEncoding}}{default console encoding}

\item{\code{smtpServer}}{default server to use for sending mail}

\item{\code{smtpPort}}{default port}

\item{\code{email}}{default email address}

\item{\code{partitionDir}}{default partition directory}

\item{\code{defaultKwicCollocate}}{default collocate for kwic display}

\item{\code{defaultKwicNode}}{default node for kwic analysis}
}}

