% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size_method.R
\docType{methods}
\name{size}
\alias{size}
\alias{size,DocumentTermMatrix-method}
\alias{size,character-method}
\alias{size,partition-method}
\title{Get number of tokens.}
\usage{
size(x, ...)

\S4method{size}{character}(x, sAttribute = NULL, verbose = TRUE)

\S4method{size}{partition}(x, sAttribute = NULL)

\S4method{size}{DocumentTermMatrix}(x)
}
\arguments{
\item{x}{object to get size(s) for}

\item{...}{further arguments}

\item{sAttribute}{character vector with s-attributes (one or more)}

\item{verbose}{logical, whether to print messages}
}
\description{
Get number of tokens.
}

