% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\docType{methods}
\name{split,partition-method}
\alias{split,partition-method}
\alias{split,partition}
\title{split partition into partition_bundle}
\usage{
\S4method{split}{partition}(x, gap, drop = FALSE, ...)
}
\arguments{
\item{x}{a partition object}

\item{gap}{an integer specifying the minimum gap for performing the split}

\item{drop}{not yet implemented}

\item{...}{further arguments}
}
\value{
a \code{partition_bundle}
}
\description{
Split a partition object into a \code{partition_bundle} if gap between strucs
exceeds a minimum number of tokens specified by 'gap'. Relevant to 
split up a plenary protocol into speeches. Note: To speed things up, the
returned partitions will not include frequency lists. The lists can be
prepared by applying \code{enrich} on the \code{partition_bundle} object that
is returned.
}
