% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/restore_method.R, R/store_method.R
\docType{methods}
\name{restore,session-method}
\alias{restore}
\alias{restore,character-method}
\alias{restore,session-method}
\alias{store}
\alias{store,session-method}
\alias{store,textstat-method}
\title{store objects}
\usage{
\S4method{restore}{session}(object)

\S4method{restore}{character}(object)

\S4method{store}{textstat}(object)

\S4method{store}{session}(object)
}
\arguments{
\item{object}{an object}
}
\description{
store and restore
}

