% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\docType{methods}
\name{store}
\alias{store}
\alias{store,textstat-method}
\alias{store,data.frame-method}
\alias{store,kwic-method}
\title{Store objects as Excel-file.}
\usage{
store(.Object, ...)

\S4method{store}{textstat}(.Object, filename = tempfile(fileext = ".xlsx"),
  rows = 1L:nrow(.Object))

\S4method{store}{data.frame}(.Object, filename = tempfile(fileext = ".xlsx"),
  rows = 1L:nrow(.Object))

\S4method{store}{kwic}(.Object, filename = tempfile(fileext = ".xlsx"),
  rows = 1L:nrow(.Object))
}
\arguments{
\item{.Object}{An object that can be processed.}

\item{...}{Further arguments.}

\item{filename}{Name of the file to write.}

\item{rows}{The rows of the table to export.}
}
\description{
Store objects as Excel-file.
}
