% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4classes.R, R/corpus.R
\docType{class}
\name{subcorpus-class}
\alias{subcorpus-class}
\alias{subcorpus}
\alias{subset,corpus-method}
\alias{subset,subcorpus-method}
\title{S4 subcorpus class.}
\usage{
\S4method{subset}{corpus}(x, ...)

\S4method{subset}{subcorpus}(x, ...)
}
\arguments{
\item{x}{A \code{corpus} or \code{subcorpus} object.}

\item{...}{An expression that will be used to create a subcorpus from
s-attributes.}
}
\description{
S4 subcorpus class.
}
\section{Slots}{

\describe{
\item{\code{s_attributes}}{A named \code{list} with the structural attributes
defining the subcorpus.}

\item{\code{cpos}}{A \code{matrix} with left and right corpus positions defining
regions (two columns).}

\item{\code{annotations}}{Object of class \code{list}.}

\item{\code{size}}{Total size of the subcorpus (length-one \code{integer} vector).}

\item{\code{metadata}}{Object of class \code{data.frame}, metadata information.}

\item{\code{strucs}}{Object of class \code{integer}, the strucs defining the
subcorpus.}

\item{\code{xml}}{Object of class \code{character}, whether the xml is "flat" or
"nested".}

\item{\code{s_attribute_strucs}}{Object of class \code{character}, the base node.}
}}

\examples{
use("polmineR")
a <- corpus("GERMAPARLMINI")
sc <- subset(a, grepl("Merkel", speaker))
sc <- subset(a, speaker == "Angela Dorothea Merkel")
sc <- subset(a, speaker == "Angela Dorothea Merkel" & date == "2009-10-28")
sc <- subset(a, grepl("Merkel", speaker) & date == "2009-10-28")
a <- corpus("GERMAPARLMINI")
b <- subset(a, date == "2009-11-10")
c <- subset(b, speaker == "Frank-Walter Steinmeier")
}
\seealso{
Other classes to manage corpora: \code{\link{corpus-class}},
  \code{\link{regions}}
}
\concept{classes to manage corpora}
