% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\docType{methods}
\name{t_test}
\alias{t_test}
\alias{t_test,context-method}
\title{perform t-test}
\usage{
t_test(.Object)

\S4method{t_test}{context}(.Object)
}
\arguments{
\item{.Object}{a context or features object}
}
\description{
S4 method for context object to perform t-test
}
