% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_method.R
\docType{methods}
\name{tag}
\alias{tag}
\alias{tag,character-method}
\title{run treetagger over a character vector}
\usage{
tag(object, ...)

\S4method{tag}{character}(object)
}
\arguments{
\item{object}{a character vector}

\item{...}{further parameters}
}
\description{
treetagger is used as a backend
}

