% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcorpus.R
\name{tempcorpus}
\alias{tempcorpus}
\title{create a tempcorpus}
\usage{
tempcorpus(.Object, ...)
}
\arguments{
\item{.Object}{a partition object}

\item{...}{further parameters}
}
\description{
Based on the corpus positions defining a partition, a temporary CWB corpus is 
generated that is stored in a temporary directory.
}
