% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates_method.R
\docType{methods}
\name{getTemplate}
\alias{getTemplate}
\alias{getTemplate,character-method}
\alias{getTemplate,missing-method}
\alias{setTemplate}
\title{Manage templates.}
\usage{
getTemplate(x)

\S4method{getTemplate}{character}(x)

\S4method{getTemplate}{missing}()

setTemplate(corpus, template)
}
\arguments{
\item{x}{either a "character" vector, if missing, the names of available templates will be returned}

\item{value}{list to assign}
}
\description{
Manage templates.
}

