% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates_method.R
\docType{methods}
\name{getTemplate}
\alias{getTemplate}
\alias{getTemplate,character-method}
\alias{getTemplate,missing-method}
\alias{setTemplate}
\alias{setTemplate,character-method}
\title{Manage templates.}
\usage{
getTemplate(.Object, ...)

\S4method{getTemplate}{character}(.Object)

\S4method{getTemplate}{missing}(.Object)

setTemplate(.Object, ...)

\S4method{setTemplate}{character}(.Object, template)
}
\arguments{
\item{...}{further parameters}

\item{template}{template to assign (a list)}

\item{x}{either a "character" vector, if missing, the names of available templates will be returned}
}
\description{
When the polmineR package is loaded, an environment .templates will be assigned to the global 
environment that will is used when preparing the html output for reading the text in a partition.
The
}

