% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{get_template}
\alias{get_template}
\alias{get_template,subcorpus-method}
\alias{get_template,character-method}
\alias{get_template,corpus-method}
\alias{get_template,partition-method}
\title{Get template for reconstructing full text.}
\usage{
get_template(.Object, ...)

\S4method{get_template}{character}(.Object, warn = TRUE)

\S4method{get_template}{corpus}(.Object, warn = TRUE)

\S4method{get_template}{partition}(.Object, warn = TRUE)

\S4method{get_template}{subcorpus}(.Object, warn = TRUE)
}
\arguments{
\item{.Object}{A \code{corpus}, \code{subcorpus} or \code{partition} object,
or a length-one \code{character} vector with a corpus ID.}

\item{...}{Further arguments to be defined.}

\item{warn}{A \code{logical} value, whether to issue a warning if template 
is not available.}
}
\description{
Templates are used to format the markdown/html output of partitions.
}
