% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\docType{methods}
\name{get_template}
\alias{get_template}
\alias{get_template,subcorpus-method}
\alias{get_template,character-method}
\alias{get_template,partition-method}
\alias{get_template,missing-method}
\alias{set_template}
\alias{set_template,character-method}
\alias{set_template,missing-method}
\title{Get and set templates.}
\usage{
get_template(.Object, ...)

\S4method{get_template}{character}(.Object)

\S4method{get_template}{partition}(.Object)

\S4method{get_template}{subcorpus}(.Object)

\S4method{get_template}{missing}(.Object)

set_template(.Object, ...)

\S4method{set_template}{character}(.Object)

\S4method{set_template}{missing}(.Object, verbose = FALSE)
}
\arguments{
\item{.Object}{object}

\item{...}{further parameters}

\item{verbose}{logical, whether to be verbose}
}
\description{
Templates are used to format the markdown/html output of partitions. Upon loading 
the polmineR package, templates for corpora are loaded into the option 'polmineR.templates'.
}
