% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TermDocumentMatrix_methods.R, R/size_method.R
\docType{methods}
\name{TermDocumentMatrix}
\alias{TermDocumentMatrix}
\alias{size,DocumentTermMatrix-method}
\alias{size,TermDocumentMatrix-method}
\title{Methods for TermDocumentMatrix / DocumentTermMatrix}
\usage{
\S4method{size}{DocumentTermMatrix}(x)

\S4method{size}{TermDocumentMatrix}(x)
}
\arguments{
\item{x}{object}
}
\description{
Methods for TermDocumentMatrix / DocumentTermMatrix
}

