% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TermDocumentMatrix.R, R/size.R
\docType{methods}
\name{TermDocumentMatrix}
\alias{TermDocumentMatrix}
\alias{size,TermDocumentMatrix-method}
\title{Methods for TermDocumentMatrix / DocumentTermMatrix}
\usage{
\S4method{size}{TermDocumentMatrix}(x)
}
\arguments{
\item{x}{object}
}
\description{
Methods for TermDocumentMatrix / DocumentTermMatrix
}
