% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms_method.R
\docType{methods}
\name{terms-partition-method}
\alias{terms,partition-method}
\alias{terms-partition-method}
\title{get terms available in a corpus}
\usage{
\S4method{terms}{partition}(x, pAttribute, regex = NULL)
}
\arguments{
\item{x}{a partition object}

\item{pAttribute}{the pAttribute to be queried}

\item{regex}{a regex to filter results}
}
\description{
get terms available in a corpus
}

