% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/textstat_class.R, R/textstat_methods.R, R/view_method.R
\docType{class}
\name{textstat-class}
\alias{as.DataTables,textstat-method}
\alias{as.data.frame,textstat-method}
\alias{colnames,textstat-method}
\alias{dim,textstat-method}
\alias{head,textstat-method}
\alias{names,textstat-method}
\alias{nrow,textstat-method}
\alias{round,textstat-method}
\alias{rownames,textstat-method}
\alias{show,textstat-method}
\alias{tail,textstat-method}
\alias{textstat-class}
\alias{view,textstat-method}
\title{S4 textstat class}
\usage{
\S4method{head}{textstat}(x, ...)

\S4method{tail}{textstat}(x, ...)

\S4method{nrow}{textstat}(x)

\S4method{round}{textstat}(x, digits = 2)

\S4method{view}{textstat}(.Object)
}
\arguments{
\item{x}{an object}

\item{...}{further parameters}

\item{digits}{}

\item{.Object}{an object}

\item{x}{textstat object}
}
\description{
superclass for keyness and context class
}
\section{Slots}{

\describe{
\item{\code{pAttribute}}{Object of class \code{"character"} p-attribute of the query}

\item{\code{corpus}}{Object of class \code{"character"}}

\item{\code{stat}}{Object of class \code{"data.frame"} statistics of the analysis}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}

\item{\code{cutoff}}{Object of class \code{"list"} cutoff levels that have been applied}
}}

