% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{textstat-class}
\alias{as.data.frame,textstat-method}
\alias{colnames,textstat-method}
\alias{dim,textstat-method}
\alias{nrow,textstat-method}
\alias{rownames,textstat-method}
\alias{show,textstat-method}
\alias{textstat-class}
\alias{trim,textstat-method}
\title{S4 textstat class}
\description{
superclass for keyness and conctext class
}
\section{Slots}{
 \describe{ \item{\code{pAttribute}:}{Object of class
  \code{"character"} p-attribute of the query } \item{\code{corpus}:}{Object
  of class \code{"character"} the CWB corpus used }
  \item{\code{stat}:}{Object of class \code{"data.frame"} statistics of the
  analysis } \item{\code{encoding}:}{Object of class \code{"character"}
  encoding of the corpus } \item{\code{cutoff}:}{Object of class
  \code{"list"} cutoff levels that have been applied } }
}

