% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textstat_class.R, R/view_method.R
\docType{class}
\name{textstat-class}
\alias{+,textstat,textstat-method}
\alias{[,textstat,ANY,ANY,ANY-method}
\alias{[[,textstat-method}
\alias{as.DataTables,textstat-method}
\alias{as.bundle}
\alias{as.data.frame,textstat-method}
\alias{as.data.table,textstat-method}
\alias{colnames,textstat-method}
\alias{dim,textstat-method}
\alias{head,textstat-method}
\alias{names,textstat-method}
\alias{nrow,textstat-method}
\alias{round,textstat-method}
\alias{rownames,textstat-method}
\alias{show,textstat-method}
\alias{sort,textstat-method}
\alias{subset,textstat-method}
\alias{tail,textstat-method}
\alias{textstat-class}
\alias{view,textstat-method}
\title{S4 textstat class}
\usage{
\S4method{head}{textstat}(x, ...)

\S4method{tail}{textstat}(x, ...)

\S4method{dim}{textstat}(x)

\S4method{nrow}{textstat}(x)

\S4method{round}{textstat}(x, digits = 2)

\S4method{colnames}{textstat}(x)

\S4method{sort}{textstat}(x, by, decreasing = TRUE)

as.bundle(object, ...)

\S4method{+}{textstat,textstat}(e1, e2)

\S4method{subset}{textstat}(x, ...)

\S4method{as.data.table}{textstat}(x)

\S4method{[[}{textstat}(x, i)

\S4method{[}{textstat,ANY,ANY,ANY}(x, i)

\S4method{view}{textstat}(.Object)
}
\arguments{
\item{x}{textstat object}

\item{...}{further parameters}

\item{digits}{no of digits}

\item{by}{by}

\item{decreasing}{logical}

\item{object}{an object}

\item{e1}{object 1}

\item{e2}{object 2}

\item{i}{index}

\item{.Object}{an object}
}
\description{
superclass for comp and context class
}
\section{Slots}{

\describe{
\item{\code{pAttribute}}{Object of class \code{"character"} p-attribute of the query}

\item{\code{corpus}}{Object of class \code{"character"}}

\item{\code{stat}}{Object of class \code{"data.frame"} statistics of the analysis}

\item{\code{name}}{name of the object}

\item{\code{encoding}}{Object of class \code{"character"} encoding of the corpus}
}}

