% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ll_method.R, R/pmi_method.R
\docType{methods}
\name{ll}
\alias{ll}
\alias{ll,comp-method}
\alias{ll,context-method}
\alias{ll,cooccurrences-method}
\alias{pmi}
\alias{pmi,context-method}
\title{text statistics}
\usage{
ll(.Object, ...)

\S4method{ll}{context}(.Object)

\S4method{ll}{cooccurrences}(.Object)

\S4method{ll}{comp}(.Object)

pmi(.Object)

\S4method{pmi}{context}(.Object)
}
\arguments{
\item{.Object}{an object}

\item{...}{further parameters}
}
\description{
text statistics
}

