% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tf_method.R
\docType{methods}
\name{tf}
\alias{tf}
\alias{tf,character-method}
\alias{tf,partition-method}
\alias{tf,partitionBundle-method}
\alias{tf-method}
\title{get term frequencies}
\usage{
tf(object, ...)

\S4method{tf}{partition}(object, query = NULL, pAttribute = NULL,
  method = "in", mc = F, verbose = T)

\S4method{tf}{partitionBundle}(object, query, pAttribute = NULL,
  method = "in", rel = FALSE)

\S4method{tf}{character}(object, query, pAttribute = NULL, method = "in",
  verbose = TRUE)
}
\arguments{
\item{object}{either a partition or a partitionBundle object}

\item{...}{further parameters}

\item{query}{a character vector (one or multiple terms to be looked up)}

\item{pAttribute}{if NULL, the pAttributes available in the partition object will be reported}

\item{method}{either "in", "grep" or "cqp" (defaults to "in")}

\item{mc}{logical, whether to use multicore (defaults to FALSE)}

\item{verbose}{logical, whether to be verbose}

\item{rel}{logical, defaults to FALSE}
}
\description{
Method to obtain term frequencies for one or multiple terms or queries. The
query may be formulated using regular expression syntax (method="grep"),
or may use the CQP syntax.
}
\examples{
# generate a partition for testing
test <- partition("PLPRBTTXT")
tf(test, "Wir") # get frequencies for one token
tf(test, c("Wir", "lassen", "uns")) # get frequencies for multiple tokens
tf(test, c("Zuwander.*", "Integration.*"), method="grep") # get frequencies using "grep"-method
tf("PLPRTXT", c("machen", "Integration"), "word")
}
\seealso{
tf
}

