% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/trim-method.R
\docType{methods}
\name{trim,keyness-method}
\alias{trim,keyness-method}
\title{trim keyness object}
\usage{
\S4method{trim}{keyness}(object, minSignificance = NULL, minFrequency = 0,
  maxRank = 0, rankBy = NULL, tokenFilter = NULL, posFilter = NULL,
  filterType = "include", digits = NULL, verbose = TRUE)
}
\arguments{
\item{object}{a keyness object}

\item{minSignificance}{minimum significance level}

\item{minFrequency}{the minimum frequency}

\item{maxRank}{maximum rank}

\item{rankBy}{the column of the statistics-table to use for ordering}

\item{tokenFilter}{tokens to exclude from table}

\item{posFilter}{pos to keep}

\item{filterType}{either "include" or "exclude"}

\item{digits}{a list}

\item{verbose}{whether to be talkative}
}
\value{
a keyness object
}
\description{
Trim a context object by applying different filters.
}
\details{
Maybe it would be more efficient to use the subset function.-
}
\author{
Andreas Blaette
}

