\docType{methods}
\name{trim}
\alias{trim}
\alias{trim-method}
\title{trim an object}
\usage{
trim(object, ...)
}
\arguments{
  \item{object}{the object to be trimmed}

  \item{...}{further parameters}
}
\description{
Use this method to trim 'context', 'keyness', 'context',
partition' and 'partitionCluster' objects. See respective
documentation available through
\code{method?trim("context")},
\code{method?trim("keyness")},
\code{method?trim("partition")},
\code{method?trim("partitionCluster")},
\code{method?trim("crosstab")}.
}
\author{
Andreas Blaette
}

