% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{trim}
\alias{trim}
\alias{trim-method}
\title{trim an object}
\usage{
trim(object, ...)
}
\arguments{
\item{object}{the object to be trimmed}

\item{...}{further arguments}
}
\description{
Method that includes varying to adjust objects from the driller package by
applying thresholds, minimum frequencies etc. It can be applied to 'context',
'keyness', 'context', 'partition' and 'partitionCluster' objects. See
the respective documentation:
\describe{
 \item{context:}{\code{method?trim("context")}}
 \item{keyness:}{\code{method?trim("keyness")}}
 \item{partition:}{\code{method?trim("partition")}}
 \item{partitionCluster:}{\code{method?trim("partitionCluster")}}
 \item{crosstab:}{\code{method?trim("crosstab")}}
}
}
\author{
Andreas Blaette
}

