% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_method.R, R/utils.R
\docType{methods}
\name{trim}
\alias{punctuation}
\alias{trim}
\alias{trim,TermDocumentMatrix-method}
\alias{trim,compBundle-method}
\alias{trim,partitionBundle-method}
\alias{trim-method}
\title{trim an object}
\format{An object of class \code{character} of length 13.}
\usage{
trim(object, ...)

\S4method{trim}{compBundle}(object, minSignificance = 0, minFrequency = 0,
  maxRank = 0, tokenFilter = NULL, posFilter = NULL,
  filterType = "include", mc = FALSE)

\S4method{trim}{partitionBundle}(object, pAttribute = NULL,
  minFrequency = 0, posFilter = NULL, tokenFilter = NULL, drop = NULL,
  minSize = 0, keep = NULL, mc = NULL, ...)

\S4method{trim}{TermDocumentMatrix}(object, minTotal = NULL,
  stopwords = NULL, keep = NULL, verbose = TRUE)

punctuation
}
\arguments{
\item{object}{the object to be trimmed}

\item{...}{further arguments}

\item{minSignificance}{minimum significance level}

\item{maxRank}{maximum rank}

\item{tokenFilter}{tokens to exclude from table}

\item{posFilter}{exclude words with a POS tag not in this list}

\item{filterType}{either "include" or "exclude"}

\item{mc}{if not NULL logical - whether to use multicore parallelization}

\item{pAttribute}{character vector, either lemma or word}

\item{drop}{partitionObjects you want to drop, specified either by number or name}

\item{minSize}{a minimum size for the partitions to be kept}

\item{keep}{specify names of partitions to keep, everything else is dropped}

\item{minTotal}{the minimum frequency}

\item{stopwords}{words/tokens to drop}

\item{verbose}{whether to be talkative}
}
\description{
Method to trim and adjust objects by 
applying thresholds, minimum frequencies etc. It can be applied to 'context',
'comp', 'context', 'partition' and 'partitionBundle' objects.
}
\author{
Andreas Blaette
}
\keyword{datasets}

