% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim.R, R/utils.R
\docType{methods}
\name{trim}
\alias{trim}
\alias{trim-method}
\alias{trim,TermDocumentMatrix-method}
\alias{trim,TermDocumentMatrix-method}
\alias{trim,DocumentTermMatrix-method}
\alias{punctuation}
\title{trim an object}
\format{An object of class \code{character} of length 13.}
\usage{
trim(object, ...)

\S4method{trim}{TermDocumentMatrix}(object, termsToKeep = NULL,
  termsToDrop = NULL, docsToKeep = NULL, docsToDrop = NULL,
  verbose = TRUE)

\S4method{trim}{DocumentTermMatrix}(object, ...)

punctuation
}
\arguments{
\item{object}{the object to be trimmed}

\item{...}{further arguments}

\item{termsToKeep}{...}

\item{termsToDrop}{...}

\item{docsToKeep}{...}

\item{docsToDrop}{...}

\item{verbose}{logical}
}
\description{
Method to trim and adjust objects by 
applying thresholds, minimum frequencies etc. It can be applied to \code{context},
\code{features}, \code{context}, \code{partition} and \code{partition_bundle} objects.
}
\author{
Andreas Blaette
}
\keyword{datasets}
