% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tTest.R
\docType{methods}
\name{tTest}
\alias{tTest}
\alias{tTest,context-method}
\title{perform t-test}
\usage{
tTest(object, ...)

\S4method{tTest}{context}(object, partitionObject)
}
\arguments{
\item{object}{a context or keyness object}

\item{...}{further parameters}

\item{partitionObject}{partition to be passed in}
}
\description{
S4 method for context object to perform t-test
}

