% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tTest.R
\docType{methods}
\name{tTest}
\alias{tTest}
\alias{tTest,context-method}
\title{perform t-test}
\usage{
tTest(.Object)

\S4method{tTest}{context}(.Object)
}
\arguments{
\item{.Object}{a context or features object}
}
\description{
S4 method for context object to perform t-test
}
