% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_method.R
\name{use}
\alias{use}
\title{use corpus}
\usage{
use(pkg = "default")
}
\arguments{
\item{pkg}{the package with a cwb corpus that shall be used, defaults to "default"
(will reset original registry)}
}
\description{
Use a cwb corpus shipped in a package or return to default registry.
}

