% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_method.R
\name{use}
\alias{use}
\title{Use packaged corpus.}
\usage{
use(pkg = NULL)
}
\arguments{
\item{pkg}{the package with a cwb corpus that shall be used (defaults to NULL, to return to default registry)
(will reset original registry)}
}
\value{
the function returns invisibly the registry that was previously set
}
\description{
Use a cwb corpus shipped in a package or return to default registry.
}
