% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weigh.R
\docType{methods}
\name{weigh}
\alias{weigh}
\alias{weigh,TermDocumentMatrix-method}
\alias{weigh,DocumentTermMatrix-method}
\title{Apply Weight to Matrix}
\usage{
weigh(.Object, ...)

\S4method{weigh}{TermDocumentMatrix}(.Object, method = "tfidf")

\S4method{weigh}{DocumentTermMatrix}(.Object, method = "tfidf")
}
\arguments{
\item{.Object}{the matrix to be weighed}

\item{...}{further parameters}

\item{method}{the kind of weight to apply}
}
\description{
Apply Weight to Matrix
}
