% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zoom-method.R
\docType{methods}
\name{zoom}
\alias{zoom}
\alias{zoom,partition-method}
\title{zoom into a partition}
\usage{
\S4method{zoom}{partition}(object, def, name = c(""), method = "in",
  tf = c("word", "lemma"), id2str = TRUE, type = NULL, verbose = TRUE)
}
\arguments{
\item{object}{a partition object}

\item{def}{a list supplying further sAttributes}

\item{name}{a name for the new partition}

\item{method}{either "in" or "grep"}

\item{tf}{character vector, pAttributes for which term frequencies shall be retrieved}

\item{id2str}{whether to transfer ids to strings}

\item{type}{the type of the resulting partition}

\item{verbose}{logical, show progress report or not (defaults to TRUE)}
}
\description{
add a further specification of a s-attribute to an existing partition
}

